==> 0_b_funciones.Rmd <==
---
title: "Funciones"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: html_document
---


```{r librerias, echo=FALSE, message=FALSE, include=FALSE, warning=FALSE}
source("0_a_librerias.R")

# creo una función ficticia para ver si se han cargado las funciones
se_ha_cargado_f <- "Sí"

```

# Descripción

***

En este .Rmd van a estar todas las funciones que se utilizarán para la creación del target trial, así como otras funciones auxiliares que se van a utilizar para el limpiado de la base y la creación de los resultados. El objetivo último, es acabar convirtiendo este .Rmd en un paquete de R. Como manera de estructurar el trabajo, cada función irá por separado en un chunk.

# Funciones auxiliares

***

En esta sección se encuentran las funciones auxiliares.

## fymd

Función que transforma los 10 primeros carácteres de una cadena en una fecha mediante la función de ymd de la librería lubridate.

```{r fymd}
# fecha en texto a ymd
fymd <- function(x) {
  x = as.Date(ymd(str_sub(x, 1, 10)))
  }

```

## fnullna

Función que transforma una cadena con texto "NULL" a NA.

```{r fnullna}

fnullna <- function(x) {
  x = if_else(x == "NULL", NA_character_, x)
  }

```

## fnull0

Función que transforma una cadena con texto "NULL" a "0".

```{r}

fnull0 <- function(x) {
  x = if_else(x == "NULL", 0, as.numeric(x))
  }

```


# Funciones del Target Trial

***

## excluir

Función que realiza un criterio de exclusión a partir de una variable de un tibble.

```{r}
# función para excluir siguiendo un criterio-----------------------------------
excluir <- function(pob, 
                    crit,
                    var)
{
  
  if (is.logical(pob %>% pull(var)) == TRUE) {
  print(glue("Variable lógica. {crit}"))
  pob <- pob %>% filter(is.na(get(var)) |
                     get(var) != TRUE)
    print(glue("n : {nrow(pob)}"))
  pob
  }
  else {
  print(glue("Variable no lógica. {crit}"))
  pob <- pob %>% filter(!is.na(get(var)))
    print(glue("n : {nrow(pob)}"))
  pob
  }
}

```

## hacer_flujo

Está función aplica los **criterios de exclusión que no dependen del tiempo**, uno detrás de otro, informando del flow.

```{r}
# flujo------------------------------------------------------------------------
hacer_flujo <- function(data,
                        exc = lista_exclusion
                   ) 
  {
  print(glue("n de partida: {nrow(data)}"))

  for( i in seq_along(exc$criterios)) {
  data <- excluir(pob = data,
                  crit = exc$criterios[[i]],
                  var = exc$variables[[i]])
}
  data
  }
  
```

## excluir_dep

Función que realiza un criterio de exclusión a partir de una variable de un tibble que es dependiente del tiempo_0 elegido.

```{r}
# función para excluir siguiendo un criterio-----------------------------------
excluir_dep <- function(pob, 
                        crit,
                        var,
                        t_var,
                        t0,
                        vb)
{
  if (t_var == "exc"){
    if (vb == TRUE) print(glue("Variable exclusión. {crit}"))
  pob <- pob %>% filter(is.na(get(var)) | get(var) >= t0)
    if (vb == TRUE) print(glue("n : {nrow(pob)}"))
  pob
  }
  else {
    if (vb == TRUE) print(glue("Variable inclusión. {crit}"))
  pob <- pob %>% filter(get(var) <= t0)
    if (vb == TRUE) print(glue("n : {nrow(pob)}"))
  pob
  }
}

```

## hacer_flujo_dep

Está función aplica los **criterios de exclusión que sí dependen del tiempo**, uno detrás de otro, informando del flow.

```{r}
# flujo------------------------------------------------------------------------
hacer_flujo_dep <- function(data,
                            exc = lista_exclusion_dep,
                            tiempo_0 = tiempo_0,
                            verbose = TRUE
                   ) 
  {
  tiempo_0 <- ymd(tiempo_0)
  print(glue("n de partida: {nrow(data)}"))
  for(i in seq_along(exc$criterios)) {
  data <- excluir_dep(
                       pob = data,
                       crit = exc$criterios[[i]],
                       var = exc$variables[[i]],
                       t_var = exc$tipo_var[[i]],
                       t0 = tiempo_0,
                       vb = verbose)
}
  data
  }

```

## f_matching_tt

Función que, para una fecha determinada (tiempo_0) realiza:

- Los criterios de exclusión dependientes del tiempo.
- Define la exposición en la población.
- Realiza matching (1:1) de esa población.

Además, permite definir cuál va a ser la variable de censura, cuál es el tiempo de seguimiento y si va a haber alguna variable de matching que sea dependiente del tiempo. 

Asimismo, permite introducir una base de test previos, para crear una variable de test negativos previos en el tiempo_0, que será introducida como una variable de matching.

Por último, se permite definir la cantidad de información que queremos que aparezca por pantalla (*verbose*) y el valor de la semilla (*seed*) para obtener siempre el mismo resultado en el matching.

```{r}
# función de matching----------------------------------------------------------
f_matching_tt <- function(data, 
                          tiempo_0, 
                          crit_exc,
                          exposure,
                          out,
                          vector_match,
                          match_with_replacement = TRUE,
                          base_test = NULL,
                          dias_a_excluir_test_previos = NULL,
                          temporal_var = NULL,
                          temporal_var_unit = "days",
                          birth_date = NULL,
                          calcular_edad = NULL,
                          calcular_test = NULL,
                          calcular_temp_var = NULL,
                          age_first = NULL,
                          age_last = NULL,
                          age_by = NULL,
                          censoring = "fdef",
                          delayed_exposure_censoring = NULL,
                          follow_up = 60,
                          verbose = TRUE,
                          seed = 1,
                          fecha_fin_estudio = ymd("2100-01-01")) {
  
  set.seed(seed)
  tiempo_0 <- ymd(tiempo_0)
  print(glue("tiempo_0 = {tiempo_0}"))
  if (verbose == TRUE) print(glue("Outcome = {out}"))
  if (verbose == TRUE) print(glue("n de partida: {nrow(data)}"))
  
  # exclusiones dependientes del tiempo
  for(i in seq_along(crit_exc$criterios)) {
  data <- excluir_dep(
                       pob = data,
                       crit = crit_exc$criterios[[i]],
                       var = crit_exc$variables[[i]],
                       t_var = crit_exc$tipo_var[[i]],
                       t0 = tiempo_0,
                       vb = verbose)
  }
  
  # exclusión por el outcome
  if (verbose == TRUE) print(glue("Variable exclusión. {out}"))
     data <- data %>% filter(is.na(get(out)) | get(out) > tiempo_0)
  if (verbose == TRUE) print(glue("n : {nrow(pob)}"))
  
  # obtención número de test negativos previos al primer positivo
  if (!is.null(base_test)) {
  data <- base_test %>% filter(id %in% data$id,
                               tpos == FALSE,
                               ftes < tiempo_0) %>% 
    group_by(id) %>% 
    summarise(n_0 = max(n)) %>% 
    right_join(data, by = "id") %>% 
    mutate(n_0 = case_when(is.na(n_0) ~ 0L, 
                 n_0 > 3 ~ 3L,
                 T ~ n_0)) 
  vector_match <- c(vector_match, "n_0")
  }
  
  # contar número de test negativos previos (si no está en el matching)
  if (!is.null(calcular_test)) {
  data <- calcular_test %>% filter(id %in% data$id,
                               tpos == FALSE,
                               ftes < tiempo_0) %>% 
    group_by(id) %>% 
    summarise(n_0 = max(n)) %>% 
    right_join(data, by = "id") %>% 
    mutate(n_0 = case_when(is.na(n_0) ~ 0L, 
                 n_0 > 3 ~ 3L,
                 T ~ n_0)) 
  }
     
  if (!is.null(dias_a_excluir_test_previos)) {
    ids_a_excluir <- base_test %>% 
      filter(id %in% data$id,
             ftes > tiempo_0 %m-% days(dias_a_excluir_test_previos),
             ftes <= tiempo_0 ) %>%
      distinct(id) %>% pull()
    data <- data %>% 
      filter(id %nin% ids_a_excluir)
  }   
        
  # creación de variable temporal respecto a tiempo_0
  if (!is.null(temporal_var)) {
  data <- data %>% 
  mutate(temp_var = as.numeric(tiempo_0 - 
                    floor_date(get(temporal_var), unit = temporal_var_unit)))
  vector_match <- c(vector_match, "temp_var")
  }
  
  # Calcular temp_var respecto a tiempo_0 (cuando no está en el matching)
  if (!is.null(calcular_temp_var)) {
  data <- data %>% 
  mutate(temp_var = as.numeric(tiempo_0 - 
                    floor_date(get(calcular_temp_var), unit = temporal_var_unit)))
  }   
     
  # introducir edad como variable de matching respecto a tiempo_0
  if (!is.null(birth_date)) {
    if(is.null(age_first) || is.null(age_last) || is.null(age_by)) {
      stop(strwrap(prefix = " ", initial = "", 
           "Have you introduced 'age_first', 'age_last',
           and, 'age_by' arguments properly?"))
    }
  data <- data %>% 
    mutate(p_edad_cat = cut(as.numeric(difftime(tiempo_0, get(birth_date), 
                                              unit = "days"))/365.25,
                          right = FALSE, 
                          breaks = c(0,
                                     seq(age_first, age_last, by = age_by),
                                     200))) %>% 
    mutate(p_edad_cat_label = p_edad_cat)
  
  vector_match <- c(vector_match, "p_edad_cat")
  }
     
  # calcular edad (cuando la edad no está en el matching)
  if (!is.null(calcular_edad)) {
    if(is.null(age_first) || is.null(age_last) || is.null(age_by)) {
      stop(strwrap(prefix = " ", initial = "", 
           "Have you introduced 'age_first', 'age_last',
           and, 'age_by' arguments properly?"))
    }
  data <- data %>% 
    mutate(p_edad_cat = cut(as.numeric(difftime(tiempo_0, get(calcular_edad), 
                                              unit = "days"))/365.25,
                          right = FALSE, 
                          breaks = c(0,
                                     seq(age_first, age_last, by = age_by),
                                     200))) %>% 
    mutate(p_edad_cat_label = p_edad_cat)
  }
     

  # exposicion-seguimiento
  data <- data %>% 
    mutate(exposicion = if_else(tiempo_0 == get(exposure) & 
                         !is.na(get(exposure)), 1, 0),
           date_start = tiempo_0,
           date_end = pmin(tiempo_0 %m+% days(follow_up), fecha_fin_estudio))
  
    if (is.null(delayed_exposure_censoring)){
    data <- data %>%
      mutate(date_end = if_else(exposicion == 0 & 
                                !is.na(get(exposure)) &
                                get(exposure) < date_end,  
                                get(exposure), date_end))
    } else {
    data <- data %>%
      mutate(date_end = if_else(exposicion == 0 & 
                                !is.na(get(exposure)) &
                                get(exposure) < date_end,  
      pmin(get(exposure) %m+% days(delayed_exposure_censoring), date_end), 
      date_end))
    }
  # meter variables del matching como numéricas  
  suppressWarnings(data <- data %>%
    mutate(across(.cols = all_of(vector_match), 
                  .fns = ~if_else(is.na(as.numeric(.x)) == TRUE, 
                                  as.numeric(as.factor(.x)),
                                  as.numeric(.x)))))      
  
  # hacer matching (investigar distintas opciones)
  try({
    greedymatch <- Matching::Match(Tr = data$exposicion, M = 1, 
                   X = data[vector_match], exact = TRUE, ties = FALSE,
                   replace = match_with_replacement)
    a <- data[unlist(greedymatch[c("index.treated", "index.control")]),]
    par <- rep(1:(nrow(a)/2), 2)
    bind_cols(par = par, a) %>%
      group_by(par) %>%
      mutate(
        date_end = min(date_end),
        # censurar si fallece (sin censurar al otro componente de la pareja)
        date_end = if_else(!is.na(get(censoring)) &
                   get(censoring) < date_end, get(censoring), date_end),     
        # creo la variable de resultado y finalizo el seguimiento cuando aparece)
         caso = if_else(get(out) <= date_end & !is.na(get(out)), 1, 0),
         date_end = if_else(caso == 1, get(out), date_end)) %>% 
      as_tibble() 
    }
  )
}

```

## f_matching_tt_mult 

Extensión de la función f_matching_tt, en la cuál se obtiene un ensayo clínico emulado para cada tiempo_0 y se unen en un único *tibble*.

```{r}
# función de matching para múltiples fechas------------------------------------
f_matching_tt_mult <- function(
                             data, 
                             f_ini,
                             f_fin,
                             crit_exc,
                             exposure,
                             out,
                             vector_match,
                             match_with_replacement = TRUE,
                             base_test = NULL,
                             dias_a_excluir_test_previos = NULL,
                             temporal_var = NULL,
                             temporal_var_unit = "days",
                             birth_date = NULL,
                             calcular_edad = NULL,
                             calcular_test = NULL,
                             calcular_temp_var = NULL,
                             age_first = NULL,
                             age_last = NULL,
                             age_by = NULL,
                             censoring = "fdef",
                             delayed_exposure_censoring = NULL,               
                             follow_up = 60,
                             verbose = TRUE,
                             seed = 1,
                             fecha_fin_estudio = ymd("2100-01-01")){
  
  fechas <- seq.Date(from = ymd(f_ini), to = ymd(f_fin), by = "days")
  
  tt_tibble =  map(fechas, 
      f_matching_tt,
      data = data,
      crit_exc = crit_exc,
      vector_match = vector_match,
      match_with_replacement = match_with_replacement,
      exposure = exposure,
      out = out, 
      base_test = base_test,
      dias_a_excluir_test_previos = dias_a_excluir_test_previos,
      temporal_var = temporal_var,
      temporal_var_unit = temporal_var_unit,
      birth_date = birth_date,
      calcular_edad = calcular_edad,
      calcular_test = calcular_test,
      calcular_temp_var = calcular_temp_var,
      age_first = age_first,
      age_last = age_last,
      age_by = age_by,
      censoring = censoring,
      delayed_exposure_censoring = delayed_exposure_censoring,
      follow_up = follow_up,
      verbose = verbose,
      seed = seed,
      fecha_fin_estudio = fecha_fin_estudio) %>% 
    # descartar los días en los que no se encuentra matching
    keep(~is_tibble(.x)) %>%  
    # unir todos los tibbles de la lista en uno único
    enframe() %>%
    rename(target_trial_number = name) %>% 
    unnest(value) %>% 
    # para evitar errores en el matcheo
    filter(!is.na(exposicion))

  if (!is.null(base_test)) {
  vector_match <- c(vector_match, "n_0")
  }  
  
  if (!is.null(temporal_var)) {
  vector_match  <- c(vector_match, 
                     glue("{temporal_var} (in {temporal_var_unit})"))
  }  
  
  if (!is.null(birth_date)) {
  vector_match  <- c(vector_match, 
                     glue("age (in {age_by}-years groups)"))
  }  
  
  return(lst(tt_tibble = tt_tibble,
       out = out,
       vector_match = vector_match))  
}


```

## shorten_tibble

Función para acortar el follow-up sobre los resultados con un follow-up mayor.

```{r}
# función de matching para acortar el follow-up--------------------------------
shorten_tibble <- function(result, shorter_follow_up, outcome_var){
  new_result <- result
  
  new_result$tt_tibble <- result$tt_tibble |>
  # shrinkage of the follow-up
  mutate(date_end2 = pmin(date_end, date_start %m+% days(shorter_follow_up))) |> 
  # reasess the case
  mutate(caso2 = if_else(caso == 1 & date_end2 >= {{outcome_var}}, 1, 0)) |> 
  # update caso and date_end
  select(-caso, -date_end) |> 
  rename(caso = caso2, date_end = date_end2)
  return(new_result)
}

```


## f_matching_tncc

En esta función se realiza un diseño de caso-control negativo desde la realización de un test (positivo = caso / negativo = control).

```{r}
# función de matching tncc
f_matching_tncc <- function(
                          data, 
                          tiempo_0,
                          caso_control_var,
                          caso_control_date,
                          vector_match,
                          crit_exc = NULL,
                          match_with_replacement = TRUE,
                          base_test = NULL,
                          dias_a_excluir_test_previos = NULL,
                          temporal_var = NULL,
                          temporal_var_unit = "days",
                          birth_date = NULL,
                          age_first = NULL,
                          age_last = NULL,
                          age_by = NULL,
                          verbose = TRUE,
                          seed = 1) {
  
  set.seed(seed)
  tiempo_0 <- ymd(tiempo_0)
  print(glue("tiempo_0 = {tiempo_0}"))
  # exclusiones dependientes del tiempo
  if (!missing(crit_exc)){
  for(i in seq_along(crit_exc$criterios)) {
  data <- excluir_dep(
                       pob = data,
                       crit = crit_exc$criterios[[i]],
                       var = crit_exc$variables[[i]],
                       t_var = crit_exc$tipo_var[[i]],
                       t0 = tiempo_0,
                       vb = verbose)
  } 
    }
  
    data <- data %>%
    # exclusión covid positivo en los 90 días previos al test
    filter(
      is.na(fcv1) |
      as.numeric(tiempo_0 - fcv1) >= 90 | 
      as.numeric(tiempo_0 - fcv1) <= 0,
      is.na(fcv2) |
      as.numeric(tiempo_0 - fcv2) >= 90 | 
      as.numeric(tiempo_0 - fcv2) <= 0) %>% 
    # obtención variable de covid previo
    mutate(fcv1_prev = if_else((as.numeric(tiempo_0 - fcv1) >= 90) == TRUE,
                               fcv1, NA_Date_))
  
    data <- base_test %>% filter(id %in% data$id,
                               get(caso_control_date) == tiempo_0) %>%
     inner_join(data, by = "id")
  
  # introducir edad como variable de matching respecto a tiempo_0
  if (!missing(birth_date)) {
    if(is.null(age_first) || is.null(age_last) || is.null(age_by)) {
      stop(strwrap(prefix = " ", initial = "", 
           "Have you introduced 'age_first', 'age_last',
           and, 'age_by' arguments properly?"))
    }
    data <- data %>% 
      mutate(p_edad_cat = cut(as.numeric(difftime(tiempo_0, get(birth_date), 
                                              unit = "days"))/365.25,
                          right = FALSE, 
                          breaks = c(0,
                                     seq(age_first, age_last,
                                         by = age_by),
                                     200))) %>% 
    mutate(p_edad_cat_label = p_edad_cat)
  # actualizar variables matching  
  vector_match <- c(vector_match, "p_edad_cat")
  }
  
  # exposicion-seguimiento
  data <- data %>%
    mutate(across(.cols = all_of(vector_match), 
                  .fns = ~if_else(is.na(as.numeric(.x)) == TRUE, 
                                  as.numeric(as.factor(.x)),
                                  as.numeric(.x))))  %>%     
    mutate(caso_control = if_else(get(caso_control_var) == TRUE, 1, 0),
           date_start = tiempo_0)
    
  # hacer matching (investigar distintas opciones)
  try({
    greedymatch <- Matching::Match(Tr = data$caso_control, M = 1, 
                   X = data[vector_match], exact = TRUE, ties = FALSE,
                   replace = match_with_replacement)
    a <- data[unlist(greedymatch[c("index.treated", "index.control")]),]
    par <- rep(1:(nrow(a)/2), 2)
    bind_cols(par = par, a) %>%
      group_by(par) %>% 
      as_tibble() 
    }
  )
}

```

## f_matching_tncc_mult 

Extensión de la función f_matching_tncc, en la cuál se obtiene un ensayo clínico emulado para cada tiempo_0 y se unen en un único *tibble*.

```{r}
# función de matching para múltiples fechas------------------------------------
f_matching_tncc_mult <- function(
                             data, 
                             f_ini,
                             f_fin,
                             caso_control_var,
                             caso_control_date,
                             vector_match,
                             crit_exc = NULL,
                             match_with_replacement = TRUE,
                             base_test = NULL,
                             dias_a_excluir_test_previos = NULL,
                             temporal_var = NULL,
                             temporal_var_unit = "days",
                             birth_date = NULL,
                             age_first = NULL,
                             age_last = NULL,
                             age_by = NULL,
                             verbose = TRUE,
                             seed = 1){
  
  fechas <- seq.Date(from = ymd(f_ini), to = ymd(f_fin), by = "days")
  
  tt_tibble =  map(fechas, 
      f_matching_tncc,
      data = data,
      crit_exc = crit_exc,
      caso_control_var = caso_control_var,
      caso_control_date = caso_control_date,
      vector_match = vector_match,
      match_with_replacement = match_with_replacement,
      base_test = base_test,
      birth_date = birth_date,
      age_first = age_first,
      age_last = age_last,
      age_by = age_by,
      verbose = verbose,
      seed = seed) %>% 
    # descartar los días en los que no se encuentra matching
    keep(~is_tibble(.x)) %>%  
    # unir todos los tibbles de la lista en uno único
    enframe() %>%
    rename(target_trial_number = name) %>% 
    unnest(value) 
    # para evitar errores en el matcheo
    # filter(!is.na(caso_control))    # filter(!is.na(caso_control))

  if (!missing(birth_date)) {
  vector_match  <- c(vector_match, 
                     glue("age (in {age_by}-years groups)"))
  }  
  
  return(lst(tt_tibble = tt_tibble,
       vector_match = vector_match))  
}


```

## f_analisis

En esta función se analizan los resultados obtenidos en la función *f_matching_tt_mult*. En la lista que se genera en la función anterior, se incluye el *outcome* y *las variables de matching*, las cuales se muestran por pantalla en la función de análisis. Además, se puede introducir el tiempo necesario para que la exposición surta efecto (*tiempo_efecto*). En el caso de que haya un *tiempo_efecto*, se muestra un aumento de la figura de Kaplan-Meier para ese intervalo de tiempo, en el que idealmente, las curvas se deberían superponer. Asimismo, se puede configurar el límite inferior del eje y (*lim_1* y *lim_zoom*) de la gráfica de Kaplan-Meier.

```{r}
# función que realiza el análisis de supervivencia-----------------------------
f_analisis <- function(
                lista,
                tiempo_efecto = 0,
                lim_1 = 0.8,
                lim_zoom = 0.95,
                tipo = NULL,
                breaks = 7,
                title = NULL
                       ) {
  
  base_analisis <- lista$tt_tibble %>% 
    mutate(survtime = date_end - date_start)
  
  max_survtime <- base_analisis %>% 
    summarise(max_survtime = as.numeric(max(survtime))) %>% pull()
  
  print(glue("Se está analizando el outcome: {lista$out} \\
             mediante la emulación de un ensayo clínico  \\
             con múltiples inicios de tiempo. Se ha utilizado matching por: \\
             {str_c(lista$vector_match, collapse = ', ')}.
             
             "))
  
  print(glue("Análisis de supervivencia días 0-{max_survtime}"))  
  print(base_analisis %>% 
             tabyl(exposicion, caso) %>%
             adorn_totals(where = c('row', 'col')) %>%
             adorn_title())

  fit <- survfit(Surv(survtime, caso) ~ exposicion, data = base_analisis) 
  
  if (missing(tipo)){
      print(ggsurvplot(fit, data = base_analisis, xlab = "days of follow-up",
           ylab = "Survival probability",
           main = "Product-Limit Survival Estimates", 
           ylim = c(lim_1, 1),
           risk.table = TRUE, conf.int = TRUE))}else{
      print(ggsurvplot(fit, data = base_analisis,
           fun = tipo,
           surv.plot.height	= 0.7,
           risk.table.height	= 0.3,
           palette = "lancet",
           xlab = "Days of Follow-Up",
           break.x.by = breaks,
           ylab = "Cumulative Incidence",
           ylim = c(0, 1 - lim_1),
           risk.table = TRUE,
           fontsize = 3,
           order.legend = TRUE,
           legend = "none",
           legend.labs = c("No booster", "Booster"),
           # legend.title ="Treatment Group",
           title = title,
           conf.int = FALSE))
        }

    
  
  cpp0 = coxph(formula = Surv(survtime, caso) ~ exposicion, 
              data = base_analisis)
  
  # imprimir resultado
  print(glue("
  
             # Estimated effectiveness in Navarra for {lista$out} is : \\
             {round((1 - exp(coef(cpp0))) * 100, 1)} % \\
             ({round((1 - exp(confint(cpp0))[2]) * 100, 1)} - \\
             {round((1 - exp(confint((cpp0)))[1]) * 100, 1)} %)"))
  
  # El tiempo_efecto se introduce como un argumento
  if (tiempo_efecto != 0) {
  base_analisis_2 <- base_analisis %>% 
  group_by(target_trial_number, par) %>% 
    filter(min(survtime) >= tiempo_efecto) %>%
  ungroup() 
    
  if (missing(tipo)){
      print(ggsurvplot(fit, data = base_analisis, 
           xlab = "days of follow-up",
           ylab = "Survival probability",
           main = "Product-Limit Survival Estimates: Zoom at tiempo-efecto", 
           xlim = c(0, tiempo_efecto),
           ylim = c(lim_zoom, 1),
           risk.table = TRUE, conf.int = TRUE))}
   
   print(glue("
              
               ####################################\\
              ####################################
              
              Análisis de supervivencia días \\
              {tiempo_efecto}-{max_survtime}"))  
   
   print(base_analisis_2 %>% 
             tabyl(exposicion, caso) %>%
             adorn_totals(where = c('row', 'col')) %>%
             adorn_title())  
    
  cpp = coxph(formula = Surv(survtime, caso) ~ exposicion, 
              data = base_analisis_2)
  
  # imprimir resultado
  print(glue("
  
             # Estimated effectiveness in Navarra for {lista$out} is : \\
             {round((1 - exp(coef(cpp))) * 100, 1)} % \\
             ({round((1 - exp(confint(cpp))[2]) * 100, 1)} - \\
             {round((1 - exp(confint((cpp)))[1]) * 100, 1)} %)"))
  }
}

```

## f_analisis lancet rr

Hacer el análisis para la carta: el rr.

```{r, fig.height = 4, fig.width = 6.5}
f_analisis_lancet_rr <- function(
                lista,
                tiempo_efecto = 0,
                title = "Put a fancy title") {
  
  base_analisis <- lista$tt_tibble %>%
    mutate(survtime = date_end - date_start)
  
  max_survtime <- base_analisis %>% 
    summarise(max_survtime = as.numeric(max(survtime))) %>% pull()
  
  print(glue("Se está analizando el outcome: {lista$out} \\
             mediante la emulación de un ensayo clínico  \\
             con múltiples inicios de tiempo. Se ha utilizado matching por: \\
             {str_c(lista$vector_match, collapse = ', ')}.
             
             "))
  
  print(glue("Análisis de supervivencia días 0-{max_survtime}"))  
  print(base_analisis %>% 
             tabyl(exposicion, caso) %>%
             adorn_totals(where = c('row', 'col')) %>%
             adorn_title())

  fit <- survfit(Surv(survtime, caso) ~ exposicion, data = base_analisis) 
  
# Cálculos de la RR
print("Cálculo del RR")

# arreglar tibble
  tt_tibble_eff <- base_analisis %>% 
  # filtrar parejas con seguimiento después del día 7
      group_by(target_trial_number, par) %>% 
      filter(min(survtime) >= tiempo_efecto) %>% 
      ungroup()

tibble_eff <- hacer_grupo_tabla_2(
         tt_tibble = tt_tibble_eff, 
         group_description = title,
         number_of_bootstrap = 500)

rr_eff <- tibble_eff %>% pull(rr_eff) %>% str_replace_all("\\.", "·")

lst(tibble_eff, rr_eff)
}

```

## f_tabla_1

A continuación, se presentan dos funciones auxiliares, que permiten hacer la Tabla 1.

```{r}
# hacer basal discretas--------------------------------------------------------
hacer_basal_discreta <- function(base, 
                                 variable,
                                 relabel = NULL) {
  
  tabla <- base %>% tabyl({{variable}}, exposicion) %>% 
    adorn_totals("col") %>%  
    adorn_percentages("col") %>%
    adorn_pct_formatting(digits = 1) %>%
    adorn_ns("front") %>% 
    rename(Variable = {{variable}}) %>% 
    mutate(Variable = as.character(Variable))
  
  if (!is.null(relabel)){
  tabla <- tabla %>% 
    mutate(Variable = relabel)
  }
  tabla
}

```

```{r}
# hacer basal continuas--------------------------------------------------------
hacer_basal_continua <- function(base, variable) {
base %>%
  group_by(exposicion) %>% 
  get_summary_stats({{variable}}, type = "mean_sd") %>% 
  mutate(mean = format(round(mean, 2), nsmall =2)) %>% 
  mutate(sd = format(round(sd, 2), nsmall =2)) %>% 
  mutate(mean_sd = glue("{mean} ({sd})")) %>% 
  select(trat_medico, mean_sd) %>% 
  pivot_wider(values_from = mean_sd, names_from = trat_medico) %>% 
  bind_cols( (base %>% 
              get_summary_stats({{variable}}, type = "mean_sd") %>% 
  mutate(mean = format(round(mean, 2), nsmall = 2)) %>% 
  mutate(sd = format(round(sd, 2), nsmall = 2)) %>% 
  mutate(mean_sd = glue("{mean} ({sd})")) %>% 
  select(mean_sd) %>% 
  rename(Total = mean_sd))) %>% 
  mutate(Variable = "mean (sd)", .before = 1)
}

```

```{r}
# Comorbidities----------------------------------------------------------------
 # función
hacer_tabla_basal_comor <- function(base, comor, name = "1") {
  base %>%
  tabyl({{comor}}, exposicion) %>% 
  adorn_totals("col") %>%  
  adorn_percentages("col") %>%
  adorn_pct_formatting(digits = 1) %>%
  adorn_ns("front") %>% 
  filter({{comor}} == 1) %>% 
  rename(Variable = {{comor}}) %>% 
  mutate(Variable = name)
  }

```


```{r}
hacer_tabla_1 <- function(tt_tibble,
                          adaptada = NULL,
                          title = "baseline characteristics",
                          desc = "tabla1",
                          path = getwd()) {

  tabla_basal_edad <- hacer_basal_discreta(base = tt_tibble,
                                         variable = p_edad_cat_label,
                                         relabel = c(
                                      "0-39",
                                     "40-44", 
                                     "45-49",
                                     "50-54",
                                     "55-59",
                                     "60-64",
                                     "65-69",
                                     "70-74",
                                     "75-79",
                                     "80-84",
                                     "85-89",
                                     "&ge; 90")
                                     ) %>% 
    filter(Variable != "0-39")

  tabla_basal_sexo <- hacer_basal_discreta(base = tt_tibble,
                                         variable = gen,
                                         relabel = c(
                                           "Men",
                                           "Women")
                                         )
                                         
  tabla_basal_n_test <- hacer_basal_discreta(base = tt_tibble,
                                         variable = n_0,
                                         relabel = c(
                                           "0",
                                           "1",
                                           "2",
                                           "&ge; 3")
                                         )

  tabla_basal_vaccine_type <- hacer_basal_discreta(base = tt_tibble,
                                         variable = tvc1_label,
                                         relabel = NULL
                                         )

  
  tabla_basal_booster_type <- hacer_basal_discreta(base = tt_tibble,
                                         variable = tvc3,
                                         relabel = NULL
                                         ) %>% 
    mutate(`0` = NA) %>% 
    filter(Variable != "<NA>")

  tabla_basal_t_pauta_comp <- hacer_basal_discreta(tt_tibble %>% 
    mutate(
      t_desde_pauta_completa = 90 + temp_var,
      t_desde_pauta_completa_cat = case_when(
        between(t_desde_pauta_completa, 90, 149) ~  "90-149 days",
        between(t_desde_pauta_completa, 150, 179) ~ "150-179 days",
        t_desde_pauta_completa >= 180 ~  "&ge; 180 days"),
      t_desde_pauta_completa_cat = factor(t_desde_pauta_completa_cat,
        levels = c("90-149 days", "150-179 days", "&ge; 180 days"))),
    variable = t_desde_pauta_completa_cat)
  
  if (missing(adaptada)){
  tabla_basal_ucn8 <- hacer_basal_discreta(base = tt_tibble,
                                         variable = ucn8,
                                           relabel = c(
                                           "1",
                                           "2",
                                           "3",
                                           "4",
                                           "5",
                                           "6",
                                           "7",
                                           "&ge; 8"))}else{
    tabla_basal_ucn8 <- hacer_basal_discreta(base = tt_tibble,
                                         variable = ucn8,
                                           relabel = c(
                                           "2",
                                           "3",
                                           "4",
                                           "5",
                                           "6",
                                           "7",
                                           "&ge; 8"))
                                             
                                           }
  
  tabla_basal <- bind_rows(
                           tabla_basal_edad,
                           tabla_basal_sexo,
                           tabla_basal_n_test,
                           tabla_basal_vaccine_type,
                           tabla_basal_t_pauta_comp,
                           tabla_basal_booster_type,
                      hacer_tabla_basal_comor(tt_tibble, vgp20,
                        name = "Flu vaccinated in 2020"),
                      hacer_tabla_basal_comor(tt_tibble, vgp21,
                        name = "Flu vaccinated in 2021"),
                        tabla_basal_ucn8,
                      hacer_tabla_basal_comor(tt_tibble, diab,
                        name = "Diabetes mellitus"),
                      hacer_tabla_basal_comor(tt_tibble, corp,
                        name = "Coronary artery disease"),
                      hacer_tabla_basal_comor(tt_tibble, irc,
                        name = "Chronic kidney failure"),
                      hacer_tabla_basal_comor(tt_tibble, epoc,
                        name = "Chronic obstructive pulmonary disease"),
                      hacer_tabla_basal_comor(tt_tibble, hlip,
                        name = "hyperdyslipidemia"),
                      hacer_tabla_basal_comor(tt_tibble, hta,
                        name = "Hypertension"),
                      hacer_tabla_basal_comor(tt_tibble, ictu,
                        name = "Stroke"),
                      hacer_tabla_basal_comor(tt_tibble, obes,
                        name = "Obesity")) %>% 
    select(-Total)
  
  n_unexposed <- tt_tibble %>% 
    filter(exposicion == 0) %>%
    nrow()
    
  n_exposed <- tt_tibble %>% 
    filter(exposicion == 1) %>%
    nrow()
  
  id_uniques <- tt_tibble %>% 
    filter(exposicion == 0) %>%
    distinct(id) %>%
    nrow()

saveRDS(tabla_basal, file.path(path, "Resultados", "RDS",
                               glue("{desc}.RDS")))  
    
if (missing(adaptada)){  
  kable(tabla_basal, align = c("c"), escape = FALSE,
      col.names = c("Variable",
                    glue("No booster group (n = {n_unexposed}*)"),
                    glue("Booster group (n = {n_exposed})")),
      table.attr = "style = \"color: black;\"",
      caption = title) %>%
  kable_styling(bootstrap_options = c("striped", 
                                      "hover", 
                                      "condensed"),
                full_width = FALSE, fixed_thead = T) %>%
  row_spec(0, bold = T, color = "white", background = "dodgerblue",
           extra_css = "vertical-align:middle") %>% 
  pack_rows("Age group (years), n (%)", 1, 11) %>%
  pack_rows("Sex, n (%)", 12, 13) %>%
  pack_rows("Number of previous SARS-CoV-2 tests, n (%)", 14, 17) %>%
  pack_rows("Type of previous vaccination, n (%)", 18, 21) %>%
  pack_rows("Time since primary vaccine schedule, n (%)", 22, 24) %>%
  pack_rows("Type of booster, n (%)", 25, 26) %>%
  pack_rows("Vaccinated against flu, n (%)", 27, 28) %>%
  pack_rows("Household members, n (%)", 29, 36) %>%
  pack_rows("Comorbidities, n (%)", 37, 44) %>%
  footnote(footnote_as_chunk = TRUE, 
           general = glue("NA = not applicable, *{id_uniques} unique individuals in the no booster group."))}else{
              kable(tabla_basal, align = c("c"), escape = FALSE,
      col.names = c("Variable",
                    glue("No booster group (n = {n_unexposed}*)"),
                    glue("Booster group (n = {n_exposed})")),
      table.attr = "style = \"color: black;\"",
      caption = title) %>%
  kable_styling(bootstrap_options = c("striped", 
                                      "hover", 
                                      "condensed"),
                full_width = FALSE, fixed_thead = T) %>%
  row_spec(0, bold = T, color = "white", background = "dodgerblue",
           extra_css = "vertical-align:middle") %>% 
  pack_rows("Age group (years), n (%)", 1, 11) %>%
  pack_rows("Sex, n (%)", 12, 13) %>%
  pack_rows("Number of previous SARS-CoV-2 tests, n (%)", 14, 17) %>%
  pack_rows("Type of previous vaccination, n (%)", 18, 21) %>%
  pack_rows("Time since primary vaccine schedule, n (%)", 22, 24) %>%
  pack_rows("Type of booster, n (%)", 25, 26) %>%
  pack_rows("Vaccinated against flu, n (%)", 27, 28) %>%
  pack_rows("Household members, n (%)", 29, 35) %>%
  pack_rows("Comorbidities, n (%)", 36, 43) %>%
  footnote(footnote_as_chunk = TRUE, 
           general = glue("NA = not applicable, *{id_uniques} unique individuals in the no booster group."))
           }

  }

```

## f_tabla_2

A continuación se realiza la función *hacer_grupo_tabla_2*, que permite obtener los resultados de la tabla 2 introduciendo: 

- El estrato de interés (tibble filtrado por condición del estrato con n parejas expuesto/no expuesto).
- La descripción del estrato.
- Número de réplicas bootstrap para obtener los intervalos de confianza.

```{r}
hacer_grupo_tabla_2 <- function(tt_tibble, 
         group_description,
         number_of_bootstrap = 500){
  
  print(group_description)
  print(tt_tibble %>% 
             tabyl(exposicion, caso) %>%
             adorn_totals(where = c('row', 'col')) %>%
             adorn_title())  
  
  max_survtime <- tt_tibble %>% 
    summarise(max_survtime = as.numeric(max(survtime))) %>% pull()

  fit <- survfit(Surv(survtime, caso) ~ exposicion, data = tt_tibble) 

  s_unexposed <- Surv(tt_tibble %>% 
                      filter(exposicion == 0) %>% 
                      pull(survtime), 
                    tt_tibble %>% 
                      filter(exposicion == 0) %>% 
                      pull(caso))
  boot_unexposed <- bootkm(s_unexposed, times = max_survtime, B = number_of_bootstrap) 

  s_exposed <- Surv(tt_tibble %>% 
                      filter(exposicion == 1) %>% 
                      pull(survtime), 
                    tt_tibble %>% 
                      filter(exposicion == 1) %>% 
                      pull(caso))
  boot_exposed <- bootkm(s_exposed, times = max_survtime, B = number_of_bootstrap) 

  unexposed_events <- tt_tibble %>% 
    filter(exposicion == 0, caso == 1) %>% nrow()

  unexposed_risk <- (1 - fit$surv[length(fit$surv)/2]) * 1e4

  exposed_events <- tt_tibble %>% 
    filter(exposicion == 1, caso == 1) %>% nrow()

  exposed_risk <- (1 - fit$surv[length(fit$surv)]) * 1e4

  point_estimate_rr <- 100 * (1 - (exposed_risk/unexposed_risk))
  qci <- quantile(100*(1-(1 - boot_exposed)/(1 - boot_unexposed)), c(.025,.975), na.rm = TRUE)

  point_estimate_rabs <- round(unexposed_risk - exposed_risk)
  qci2 <- quantile((1-boot_unexposed) * 1e4 - (1-boot_exposed) * 1e4, c(.025,.975), na.rm = TRUE)

  tibble(
       group = group_description, 
       unexposed_events = unexposed_events,
       unexposed_risk = round(unexposed_risk),
       exposed_events = exposed_events,
       exposed_risk = round(exposed_risk),
       rr_eff = glue("{round(point_estimate_rr, 1)}% \\
                     (95% CI {round(qci[1], 1)}-{round(qci[2], 1)})"), 
       rabs_eff = glue("{round(point_estimate_rabs, 2)} \\
                     ({round(qci2[1])}, {round(qci2[2])})"))

}

```


```{r}
# hacer tabla 2----------------------------------------------------------------
hacer_tabla_2 <- function(tt_tibble, 
                          tiempo_efecto = 7) {

# arreglar tibble
  tt_tibble_after_7_days <- tt_tibble %>% 
  # filtrar parejas con seguimiento después del día 7
      mutate(survtime = date_end - date_start) %>% 
      group_by(target_trial_number, par) %>% 
      filter(min(survtime) >= tiempo_efecto) %>%
      ungroup() %>% 
  # crear tiempo desde pauta completa
      mutate(
        t_desde_pauta_completa = 90 + temp_var,
        t_desde_pauta_completa_cat = case_when(
          between(t_desde_pauta_completa, 91, 195) ~  "91-195 days",
          between(t_desde_pauta_completa, 196, 220) ~ "196-220 days",
        t_desde_pauta_completa > 221 ~  "&ge; 221 days"),
        t_desde_pauta_completa_cat = factor(t_desde_pauta_completa_cat,
          levels = c("91-195 days", "196-220 days", "&ge; 221 days")))

  tt_tibble_after_7_days_pairs <- tt_tibble_after_7_days %>% 
    filter(exposicion == 1) %>% nrow()

# hacer estratos tabla 2
  table2_overall <- hacer_grupo_tabla_2(
    tt_tibble = tt_tibble_after_7_days, 
    group_description = "Overall",
    number_of_bootstrap = 500)

  table2_grupo_edad_1 <- hacer_grupo_tabla_2(
    tt_tibble = tt_tibble_after_7_days %>% 
      filter(p_edad_cat_label %in% c(
        "[40,45)",	
        "[45,50)",		
        "[50,55)")), 
    group_description = "40-54",
    number_of_bootstrap = 500)

  table2_grupo_edad_2 <- hacer_grupo_tabla_2(
    tt_tibble = tt_tibble_after_7_days %>% 
      filter(min(survtime) >= tiempo_efecto) %>%
      ungroup() %>% 
      filter(p_edad_cat_label %nin% c(
        "[40,45)",	
        "[45,50)",		
        "[50,55)")), 
    group_description = "&ge; 55",
    number_of_bootstrap = 500)

  table2_hombre <- hacer_grupo_tabla_2(
    tt_tibble = tt_tibble_after_7_days %>% 
      filter(gen == 1), 
    group_description = "Male",
    number_of_bootstrap = 500)

  table2_mujer <- hacer_grupo_tabla_2(
    tt_tibble = tt_tibble_after_7_days %>% 
      filter(gen == 2), 
    group_description = "Female",
    number_of_bootstrap = 500)

  table2_vac_pfizer <- hacer_grupo_tabla_2(
    tt_tibble = tt_tibble_after_7_days %>% 
      filter(tvc1_label == "BioNTech/Pfizer"), 
    group_description = "BioNTech/Pfizer",
    number_of_bootstrap = 500)

  table2_vac_janssen <- hacer_grupo_tabla_2(
    tt_tibble = tt_tibble_after_7_days %>% 
      filter(tvc1_label == "J&J / Janssen"), 
    group_description = "J&J / Janssen",
    number_of_bootstrap = 500)

  table2_vac_moderna <- hacer_grupo_tabla_2(
    tt_tibble = tt_tibble_after_7_days %>% 
      filter(tvc1_label == "Moderna/Lonza"), 
    group_description = "Moderna/Lonza",
    number_of_bootstrap = 500)

  table2_vac_az <- hacer_grupo_tabla_2(
    tt_tibble = tt_tibble_after_7_days %>% 
      filter(tvc1_label == "Oxford/AstraZeneca"), 
    group_description = "Oxford/AstraZeneca",
    number_of_bootstrap = 500)
  
  table2_tsv1 <- hacer_grupo_tabla_2(
    tt_tibble = tt_tibble_after_7_days %>% 
      filter(t_desde_pauta_completa_cat == "91-195 days"), 
    group_description = "91-195 days",
    number_of_bootstrap = 500)

  table2_tsv2 <- hacer_grupo_tabla_2(
    tt_tibble = tt_tibble_after_7_days %>% 
      filter(t_desde_pauta_completa_cat == "196-220 days"), 
    group_description = "196-220 days",
    number_of_bootstrap = 500)

  table2_tsv3 <- hacer_grupo_tabla_2(
    tt_tibble = tt_tibble_after_7_days %>% 
      filter(t_desde_pauta_completa_cat == "&ge; 221 days"), 
    group_description = "&ge; 221 days",
    number_of_bootstrap = 500)

# para el booster el filtrado es algo más complicado
  ids_pfizer <- tt_tibble_after_7_days %>% 
  # seleccionar booster pfizer
    filter(exposicion == 1,
         tvc3 == "BioNTech/Pfizer") %>% 
    select(target_trial_number, par)
# seleccionar todas las parejas de booster pfizer
  tt_tibble_after_7_days_pfizer <- tt_tibble_after_7_days %>% 
    semi_join(ids_pfizer)

  ids_moderna <- tt_tibble_after_7_days %>% 
  # seleccionar booster pfizer
    filter(exposicion == 1,
         tvc3 == "Moderna/Lonza") %>% 
    select(target_trial_number, par)
# seleccionar todas las parejas de booster pfizer
  tt_tibble_after_7_days_moderna <- tt_tibble_after_7_days %>% 
    semi_join(ids_moderna)

  table2_tvc3_pfizer <- hacer_grupo_tabla_2(
    tt_tibble = tt_tibble_after_7_days_pfizer, 
    group_description = "BioNTech/Pfizer",
    number_of_bootstrap = 500)

  table2_tvc3_moderna <- hacer_grupo_tabla_2(
    tt_tibble = tt_tibble_after_7_days_moderna, 
    group_description = "Moderna/Lonza",
    number_of_bootstrap = 500)

  # ensamblar tabla 2
  table2_ensambled <- bind_rows(
    table2_overall,
    table2_grupo_edad_1,
    table2_grupo_edad_2,
    table2_hombre,
    table2_mujer,
    table2_vac_pfizer,
    table2_vac_janssen,
    table2_vac_moderna,
    table2_vac_az,
    table2_tsv1,
    table2_tsv2,
    table2_tsv3,
    table2_tvc3_pfizer,
    table2_tvc3_moderna)
  
 # kable table 2  
 kable(table2_ensambled, align = c("c"), escape = FALSE,
      col.names = c(" ",
                    "Events",
                    "Risk per 10 000 individuals",
                    "Events",
                    "Risk per 10 000 individuals",
                    " ",
                    " "),
      table.attr = "style = \"color: black;\"",
      caption = "Estimated effectiveness of an mRNA COVID-19 vaccine booster in individuals* who had completed primary vaccination schedule against COVID-19") %>%
   add_header_above(c(" ",
                      "No booster group" = 2, 
                      "Booster group" = 2, 
                      "1–risk ratio (95% CI)", 
                      "Risk difference per 10 000 individuals (95% CI)"),
                    bold = T, color = "white", background = "dodgerblue",
           extra_css = "vertical-align:middle") %>% 
  kable_styling(bootstrap_options = c("striped", 
                                      "hover", 
                                      "condensed"),
                full_width = FALSE, fixed_thead = T) %>%
  row_spec(0, bold = T, color = "white", background = "dodgerblue",
           extra_css = "vertical-align:middle") %>% 
  pack_rows("Age group (years)", 2, 3) %>%
  pack_rows("Sex", 4, 5) %>%
  pack_rows("Type of previous vaccination", 6, 9) %>%
  pack_rows("Time since vaccination completed", 10, 12) %>%  
  pack_rows("Type of booster", 13, 14) %>%
  footnote(footnote_as_chunk = TRUE, 
           general = glue("Analyses based on {tt_tibble_after_7_days_pairs} matched pairs who remained under follow-up by day 7 after the booster dose"))
}

```

## f_tabla_3

```{r}
# hacer tabla 3----------------------------------------------------------------
hacer_fila_tabla_3 <- function(
  tt_tibble,  
  tiempo_efecto = 7,
  description) {

# arreglar tibble
  tt_tibble_after_7_days <- tt_tibble %>% 
  # filtrar parejas con seguimiento después del día 7
      mutate(survtime = date_end - date_start) %>% 
      group_by(target_trial_number, par) %>% 
      filter(min(survtime) >= tiempo_efecto) %>%
      ungroup()
  
  tt_tibble_after_7_days_pairs <- tt_tibble_after_7_days %>% 
    filter(exposicion == 1) %>% nrow()

# hacer estratos tabla 2
  table3_overall <- hacer_grupo_tabla_2(
    tt_tibble = tt_tibble_after_7_days, 
    group_description = description,
    number_of_bootstrap = 500)
  }
  
  
```

```{r}
hacer_tabla_3 <- function(){
t3_main <- hacer_fila_tabla_3(matching_fcv1$tt_tibble,
                      description = "Main analysis")

t3_sens0 <- hacer_fila_tabla_3(matching_fcv1_sens0$tt_tibble,
                      description = "Restricting to people with &ge; 1
                      negative lab test at enrolment")

t3_sens1 <- hacer_fila_tabla_3(matching_fcv1_sens1$tt_tibble,
                      description = "Restricting to persons with no
                      tests in the 7 days before enrolment")

t3_sens2 <- hacer_fila_tabla_3(matching_fcv1_sens2$tt_tibble,
                      description = "Censoring matched pairs 7 days
                      after the control receives a booster rather than 0 days")

t3_sens3 <- hacer_fila_tabla_3(matching_fcv1_sens3$tt_tibble,
                      description = "Adding 2 days to the PCR test date
                      rather than the PCR collection date")

t3_sens4 <- hacer_fila_tabla_3(matching_fcv1_sens4$tt_tibble,
                      description = "Selecting matched controls without
                      replacement")

t3_sens5 <- hacer_fila_tabla_3(matching_fcv1_sens5$tt_tibble,
                      description = "Exact matching by age (year by year vs 5-year
                      groups)")


# ensamblar tabla 3
  table3_ensambled <- bind_rows(
    t3_main,
    t3_sens0,
    t3_sens1,
    t3_sens2,
    t3_sens3,
    t3_sens4,
    t3_sens5) %>% 
    select(group, rr_eff, rabs_eff)
  
 # kable table 3  
 kable(table3_ensambled, align = c("c"), escape = FALSE,
      col.names = c(" ",
                    "1–risk ratio (95% CI)", 
                    "Risk difference per 10 000 individuals (95% CI)"),
      table.attr = "style = \"color: black;\"",
      caption = "Sensitivity analyses of vaccine effectiveness 7–34 days after an mRNA COVID-19 vaccine booster dose") %>%
  kable_styling(bootstrap_options = c("striped", 
                                      "hover", 
                                      "condensed"),
                full_width = FALSE, fixed_thead = T) %>%
  row_spec(0, bold = T, color = "white", background = "dodgerblue",
           extra_css = "vertical-align:middle")
}
```
