==> 0_funciones.Rmd <==
---
title: "Funciones"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: html_document
---


```{r set-global-options, echo = FALSE}
knitr::opts_chunk$set(eval = TRUE, 
                      echo = TRUE, 
                      cache = FALSE,
                      include = TRUE,
                      collapse = FALSE,
                      dependson = NULL,
                      engine = "R", # Chunks will always have R code, unless noted
                      error = TRUE,
                      fig.path="Figures/",  # Set the figure options
                      fig.align = "center", 
                      fig.width = 12,
                      fig.height = 12)
```


```{r libraries, echo=FALSE, message=FALSE, include = FALSE, warning=FALSE}
library(readxl)
library(readr)
library(data.table)
library(lubridate)
library(tidyverse)
library(knitr)
library(kableExtra)
library(plyr)
library(dplyr)
library(ggpubr)
library(scales)
library(stringr)
library(plotly)
library(Matching)
#library(MatchIt)
library(plotrix)  # para sacar piramides descriptivas de las tablas_analisis
library(ggpmisc) # para series temporales
library(survminer)
library(survival)

```

# Descripción

***

En este .Rmd van a estar todas las funciones que se utilizarán para la creación del target trial, así como otras funciones auxiliares que se van a utilizar para el limpiado de la base y la creación de los resultados. El objetivo último, es acabar convirtiendo este .Rmd en un paquete de R. Como manera de estructurar el trabajo, cada función irá por separado en un chunk.

# Funciones auxiliares

***

En esta sección se encuentran las funciones auxiliares

## fymd

Función que transforma los 10 primeros carácteres de una cadena en una fecha mediante la función de ymd de la librería lubridate.

```{r fymd}
# fecha en texto a ymd
fymd <- function(x) {x=as.Date(ymd(str_sub(x, 1, 10)))}

```

## fnull0

Función que transforma una cadena con texto "NULL" a un real 0. Esta función es un poco chunga, ya que pide cambiar de carácteres a numérico.

```{r fnull0}
# de texto "NULL" a 0
fnull0 <- function(x) {x = as.numeric(ifelse(x == "NULL", 0, x))}

```

Función que transforma una cadena con texto "NULL" a NA.

```{r fnullna}

fnullna <- function(x) {x = ifelse(x == "NULL", NA, x)}

```

# Funciones del Target Trial

***

```{r}

```
