==> 1_b_creacion_pobana_0.Rmd <==
---
title: "Preprocesado: b) Crear base pobana_0"
output: 
  html_document:
    code_folding: show
    highlight: espresso
    css: style.css
    keep_md: yes
    number_sections: yes
    theme: readable
    toc: yes  
    toc_float:
      collapsed: yes
      smooth_scroll: yes
      toc_depth: 3     
editor_options: 
  chunk_output_type: inline
---


```{r set-global-options, echo = FALSE}
knitr::opts_chunk$set(eval = TRUE, 
                      echo = TRUE, 
                      cache = FALSE,
                      include = TRUE,
                      collapse = FALSE,
                       dependson = NULL,
                      engine = "R", # Chunks will always have R code, unless noted
                      error = TRUE,
                      fig.path="Figures/",  # Set the figure options
                      fig.align = "center", 
                      fig.width = 12,
                      fig.height = 12)

```

```{r librerias, echo=FALSE, message=FALSE, include = FALSE, warning=FALSE}
# source("0_a_librerias.R")
# ya se cargan cuando se carguen las funciones

```

```{r}
# función para cargar un .Rmd como si se hiciese source de un .r---------------
source_rmd = function(file, ...) {
  tmp_file = tempfile(fileext = ".R")
  on.exit(unlink(tmp_file), add = TRUE)
  knitr::purl(file, output = tmp_file)
  source(file = tmp_file, ...)
}

```

```{r}
# cargar funciones de 0_funciones----------------------------------------------
if (!exists("se_ha_cargado_f")) {
source_rmd("0_b_funciones.rmd")
}

```

```{r datos}
# cargar datos----------------------------------------------------------------
dat220613_sin_dup <- readRDS(file.path("Datos", "Procesados",
                                     "dat220613_sin_dup.RDS")) # 691327

```

# Creación de pobana_0

***

Para crear *POBANA* se usan dos fuentes de datos de poblacion: 

- *pobna50en25* (ref a 21-01-21 ).

- *pobna88* (pobna88). 

*POBANA* es la Tabla 1 del análisis, de la que se extraen los individuos aplicando el tiempo 0 y a la que solo falta vincular el numero de test negativos previos (que sale de la Tabla 2: test).

```{r pobna50en25}
# cargar base------------------------------------------------------------------
pobna50en25 <- readRDS(file.path("Datos", "Brutos", "pobna50en25.RDS"))

# fnac fnacimiento
# fdef fdefuncion
# res Es_R esidente
# pdvac Pob_Diana_Vacunacion
# gvac20/21 Vacuna_Gripe_2020/21
# fvcv1/2/3 VacCovid.Fecha_1Dosis_Vacuna/2/3"
# fcv1/2/3 Fecha_Ini_Covid_1/2/3
# fih1/2/3 Fecha_Ingreso_1/2/3
# fuci Fecha_Estancia_UCI
# cvsn ha_tenido_covid
# cvhsn Ha_Tenido_Hospitalizacion
# cvhusn Hospitalizado_En_UCI

names(pobna50en25)
#  [1] "id"         "gdep"       "gen"        "edad"       "pnac"      
#  [6] "migran"     "ctsi"       "gma"        "peso"       "domi"      
# [11] "res"        "resl"       "ucn8"       "cpob"       "seccion"   
# [16] "lzbs"       "rentasc"    "ptipo"      "bsgr"       "deme"      
# [21] "diab"       "ainm"       "corp"       "irc"        "epoc"      
# [26] "hlip"       "hta"        "ictu"       "obes"       "cimc"      
# [31] "nhos"       "hf1"        "hfu"        "mhos"       "nvc"       
# [36] "vcf2d"      "fvac19"     "tvac19"     "fvac20"     "tvac20"    
# [41] "mfpcr"      "mct"        "mfinh"      "mfinu"      "tesf1"     
# [46] "tesfu"      "tesn"       "test"       "fdef210115" "vcf1d"   

pob_0 <- pobna50en25 %>% 
  select(id, gen, migran, ctsi, res, ucn8, rentasc, ptipo, deme:cimc)

```


```{r pobna88}
# cargar base------------------------------------------------------------------
pobna88 <- readRDS(file.path("Datos", "Brutos", "pobna88.RDS"))

# unir pob_0 con pobna88-------------------------------------------------------
pob <- left_join(pob_0, select(pobna88, id, fnac, lapna, imc))

```

```{r coincidentes}
# comprobar coincidentes ambas bases-------------------------------------------
pobch <- pob %>% select(id) %>% mutate(pob = TRUE)
datvich <- dat220613_sin_dup %>% select(id) %>% mutate(dvi = TRUE)
pobdat <- full_join(pobch, datvich)
table(pobdat$pob, pobdat$dvi, useNA = "always")
#
#            TRUE   <NA>   pob    hay 19472 aprox casos de pob que no aparecen en dvi
# dvi  TRUE 645605  19472   
#      <NA>  45722      0             45722 de dvi que no aparecen en pob

# coinciden en 645605 personas trabajamos sobre ellas

```

La fuente de población es pob (665077), en vacunas-infeccion hay más población (dat220613_sin_dup -691327-) trabajamos sobre los datos coincidentes (645605)

```{r pob2}
# unión pob con dat220613_sin_dup----------------------------------------------
pob2 <- pob %>% 
  inner_join(dat220613_sin_dup %>% select(id:aseg)) %>% 
## relocate
  relocate(fnac2, .after = gen) %>% 
  relocate(fnac, .after = gen) %>% 
  relocate(resi, .after = res) %>%
  relocate(imc, .after = obes) %>%
  relocate(tper, .after = ptipo) %>%
  relocate(aseg, .after = migran) %>%
  relocate(sexm, .after = gen) %>%
  relocate(czbs, .after = rentasc) %>%
  relocate(vuln:depe, .before = deme) %>%
  relocate(fdef, .after = fnac2) %>%
  relocate(lapna, .after = migran)

```


```{r check-pob2}
#### incoherencia de género y fecha de nacimiento ####
# incoherencias género---------------------------------------------
pob2 %>% janitor::tabyl(gen, sexm)
#table(pob2$gen, pob2$sexm) eliminamos los casos de incohencia de género (148+162)
  #             0      1
  # Hombre    148 318701
  # Mujer  326594    162

# incoherencias fechas de nacimiento-------------------------------------------
pob2 %>% select(id, fnac, fnac2) %>% 
  filter(fnac != fnac2) # 1519

# arreglar inchorencias--------------------------------------------------------
pob3 <- pob2 %>% 
  mutate(inc_g = ((gen == "Hombre" & sexm == 0) | (gen == "Mujer" & sexm == 1)),
         inc_e = abs(round(as.numeric(fnac - fnac2)/365.25)) > 2) %>% 
  mutate(inc = as.numeric(inc_g) + as.numeric(inc_e))

pob3 %>% count(inc_g)
pob3 %>% count(inc_e)
pob3 %>% count(inc)

# table(pob3$inc)
#     0      1 
# 645209    396 
## es verdad que son pocos si la diferencia de edad las fijamos en 2 años

```


```{r pob4}
# se excluyen los incoherentes-------------------------------------------------
pob4 <- pob3 %>% 
  filter(inc_g == FALSE & inc_e == FALSE)  %>%
  # eliminar variables de detección de incoherencias
  select(- c(sexm, fnac2, inc_g:inc)) # 645209

```


```{r pob5}
# quitar residancias-----------------------------------------------------------
pob4 %>% count(res, resi)
#table(pob4$resn)
#     0      1      2 
# 635179   3970   6060 

pob5 <- pob4 %>% 
  mutate(resn = res + resi) %>%
  relocate(resn, .after = ctsi) %>% 
  mutate(resn = resn > 0) %>% 
  select(-c(res, resi))

pob5 %>% count(resn)
# FALSE   TRUE 
# 635179  10030 

```


```{r pob6}
# Agregar info vacuna y marcar dependientes------------------------------------
pob6 <- pob5 %>% left_join(select(dat220613_sin_dup, id, vgp20:fuci)) %>% 
  # dependientes
  left_join(select(pobna50en25, id, gdep)) %>% 
  relocate(gdep, .after = depe) %>% 
  mutate(depex = (gdep == "Dependencia Severa" | gdep == "Gran Dependencia" | 
                    (gdep == "No valorado" & depe == 1))) %>% 
  relocate(depex, .after = depe) %>% 
  select( -c(gdep, depe)) %>% 
  # renombrar depex como depe
  rename(depe = depex)

pob6 %>% count(depe)
#  FALSE   TRUE 
# 634173  11036
 
```


```{r pob7}
# marcar sanitarios o sociosan-------------------------------------------------
pob6 %>% count(ptipo, tper)

# bastante incoherente ese cruze.. de nuevo (como en dependiente) extraemos 
# siguiendo un criterio mixto:
pob7 <- pob6 %>% 
  mutate(prfss = tper == "CUN" | 
                 tper == "Profesional Sanitario" | 
                 tper == "CUN" | 
                 tper == "San Miguel" | 
                 tper == "SPRL" | 
                 ptipo == "residencias" |
                 ptipo == "sanitarios") %>% 
  relocate(prfss, .after = tper) %>% 
  select(-ptipo)

# EL ARTICULO EXCLUYE, AL MENOS LO DICE EN EL TEXTO, A LOS DE EDUCACION, aqui no (estan en ptipo) ####

pob7 %>% count(prfss)
#  FALSE   TRUE   <NA> 
# 231040  22719 391450

```


```{r pob8}
# marcar educación-------------------------------------------------------------
pob8 <- pob7 %>%  left_join(select(dat220613_sin_dup, id, tper)) %>% 
  mutate(educa = tper == "Educacion") %>% 
  select(-tper) %>% 
  relocate(educa, .after = prfss)

```


```{r pob9}
# sin pauta completa (previa a 3 dosis)
pob8 %>% count(is.na(fvc2))
#  FALSE   TRUE 
# 465536 179673 
table(is.na(pob8$fvc2))   #no tiene segunda dosis (pero hay que sacar la jansen que solo era una)
#  FALSE   TRUE 
# 465530 179679 

pob8 %>% count(tvc1)

  #     Bharat Biotech      BioNTech/Pfizer         C N Gamaleya 
  #                  1               383349                   15 
  #         CansinoBio      Centro Chumakov                 CIGB 
  #                  2                    3                    4 
  # Institute of India        J&J / Janssen        Moderna/Lonza 
  #                  5                23418                73107 
  #            Novavax   Oxford/AstraZeneca            Sinopharm 
  #                 12                67329                   82 
  # Sinopharm Abu Dabi              SINOVAC Sinovac Biotech Ltd. 
  #                  1                    1                  142


# vacunacion incompleta: no primera dosis o no segunda dosis si es diferente de Janssen
pob9 <- pob8 %>% 
  mutate(vcix = is.na(fvc1) | (tvc1 != "J&J / Janssen" & is.na(fvc2))) %>%
  relocate(vcix, .after = tvc1)

pob9 %>% count(vcix)
# FALSE   TRUE 
# 488781 156428 

```


```{r pob10}
# czbs-------------------------------------------------------------------------
pob10 <- pob9 %>% 
  mutate(czbs = fnullna(czbs))

pob10 %>% count(is.na(czbs))
#  FALSE   TRUE 
# 629649  15560

```


```{r pob11}
# grupos de edad---------------------------------------------------------------
## edad a 2022/01/01
pob11 <- pob10 %>% 
  mutate(gedad = cut(round(as.numeric(ymd("2022-01-01") - fnac)/365.25), 
         breaks = c(0, 40, 45, 50, 55, 60, 65, 70, 75, 80, 85, 90, 110))) %>%
  relocate(gedad, .after = fnac) %>% 
  select(-aseg)

pob11 %>% count(gedad)

# varios criterios de edad-----------------------------------------------------
# - GEDAD  (años cumplidos a 2022/01/01)
pob11 %>% count(gedad != "(0,40]") 
# Excluye a 270654 + 3 NA = 270657

# - mayores de 40 años (cumplen 40 años en 2021) FNAC 
pob11 %>% count(lubridate::year(fnac) <= 1981)
# Excluye a 266177

# - 40 años o mayores de 40 (cumplen 40 años en 2022) FNAC 
pob11 %>% count(lubridate::year(fnac) <= 1982)
# Excluye a 257298

```


```{r pob12}
# Domicilio--------------------------------------------------------------------
pob12 <- pob11 %>% left_join(select(pobna50en25, id, domi)) %>% 
  mutate(domi = as.numeric(as.factor(domi))) %>% 
  relocate(domi, .after = ucn8) %>% 
  tidylog::filter(!is.na(domi))

# ver distribución de número de personas por domicilio
pob12 %>% count(domi) %>% count(n) # hay domicilios con muchas personas

```

```{r vacunas}
# arreglar base para que las variables se puedan introducir en el matching-----
vacunas_top <- c("Oxford/AstraZeneca",
                 "BioNTech/Pfizer",
                 "J&J / Janssen",
                 "Moderna/Lonza")

pob13 <- pob12 %>% 
  # arreglar vacunas: sólo las 4 principales y sólo NA en tvc2 si tvc1 es J & J
  filter(tvc1 %in% vacunas_top) %>%
  filter(!is.na(tvc2) | tvc1 == "J&J / Janssen") %>% 
  mutate(tvc2 = if_else(is.na(tvc2), "NA", tvc2))

```

```{r pobana_0}
# guardar base de pobana antes de hacer las exclusiones------------------------
pobana_0 <- pob13
saveRDS(pobana_0, file = file.path("Datos", "Procesados", "pobana_0.RDS"))

```

# Descripción de pobana

***

```{r}
e_pres <- c()
e_pres[1] <-  "Código paciente (pac_unif_cod)"
e_pres[2] <-  "Género"
e_pres[3] <-  "Fecha nacimiento (exclusion)"
e_pres[4] <-  "Grupo de edad, quinquenal salvo  (0,40],(90,110]"
e_pres[5] <-  "Fecha defuncion"
e_pres[6] <-  "Migrante (pais de nacimiento no España)"
e_pres[7] <-  "Área (grupo de paises) de nacimiento"
e_pres[8] <-  "Código aportacion farmaceutica (tsi)"
e_pres[9] <-  "Vive en residencia true"
e_pres[10] <- "Unidad de convivencia, truncado (8 ó mas)"
e_pres[11] <- "Indicador numerico de domicilio"
e_pres[12] <- "Renta seccion censal categorizada -promedio 2011-2016-"
e_pres[13] <- "Código de zona básica -por centro de atención-"
e_pres[14] <- "Profesional sociosanitario true"
e_pres[15] <- "Profesional educativo true"
e_pres[16] <- "Vulnerable"
e_pres[17] <- "Dependiente true"
e_pres[18] <- "Demencia"
e_pres[19] <- "Diabetes"
e_pres[20] <- "Enfermedad autoinmune"
e_pres[21] <- "Coronariopatia"
e_pres[22] <- "Insuficiencia renal cronica"
e_pres[23] <- "Enfermedad pulmonar obstructiva cronica-EPOC"
e_pres[24] <- "Hiperlipemia"
e_pres[25] <- "Hipertension arterial (hta)"
e_pres[26] <- "Ictus"
e_pres[27] <- "Obesidad (imc>30)"
e_pres[28] <- "Indice masa corporal (imc)"
e_pres[29] <- "Indice masa corporal, categorias"
e_pres[30] <- "Recibe vacuna gripe campaña 2020"
e_pres[31] <- "Recibe vacuna gripe campaña 2021"
e_pres[32] <- "Tipo vacuna covid 1"
e_pres[33] <- "Vacunacion incompleta true"
e_pres[34] <- "Fecha vacuna covid 1"
e_pres[35] <- "Tipo vacuna covid 2"
e_pres[36] <- "Fecha vacuna covid 2"
e_pres[37] <- "Tipo vacuna covid 3"
e_pres[38] <- "Fecha vacuna covid 3"
e_pres[39] <- "Infeccion covid si/no (ref junio 2022)"
e_pres[40] <- "Fecha infeccion covid 1"
e_pres[41] <- "Fecha infeccion covid 2"
e_pres[42] <- "Fecha primer ingreso hospitalario"
e_pres[43] <- "Ingreso UCI si/no"
e_pres[44] <- "Fecha ingreso UCI"

 

 
descripcion <- tibble(Etiqueta = e_pres,
                     Variable = names(pobana_0))

kable(descripcion, format = "html")%>%  
        kable_styling(bootstrap_options = c("striped", "hover", "condensed"), 
                      full_width = F, fixed_thead = F) %>%
        column_spec(1:2, border_right = T) %>% column_spec(1, bold=T) %>%
        column_spec(2, )

```
