==> 1_c_arreglar_base_test.Rmd <==
---
title: "Preprocesado"
output: 
  html_document:
    code_folding: show
    highlight: espresso
    css: style.css
    keep_md: yes
    number_sections: yes
    theme: readable
    toc: yes  
    toc_float:
      collapsed: yes
      smooth_scroll: yes
      toc_depth: 3     
editor_options: 
  chunk_output_type: inline
---


```{r set-global-options, echo = FALSE}
knitr::opts_chunk$set(eval = TRUE, 
                      echo = TRUE, 
                      cache = FALSE,
                      include = TRUE,
                      collapse = FALSE,
                       dependson = NULL,
                      engine = "R", # Chunks will always have R code, unless noted
                      error = TRUE,
                      fig.path="Figures/",  # Set the figure options
                      fig.align = "center", 
                      fig.width = 12,
                      fig.height = 12)

```

```{r librerias, echo=FALSE, message=FALSE, include = FALSE, warning=FALSE}
# source("0_a_librerias.R")
# ya se cargan cuando se carguen las funciones

```

```{r}
# función para cargar un .Rmd como si se hiciese source de un .r---------------
source_rmd = function(file, ...) {
  tmp_file = tempfile(fileext = ".R")
  on.exit(unlink(tmp_file), add = TRUE)
  knitr::purl(file, output = tmp_file)
  source(file = tmp_file, ...)
}

```

```{r}
# cargar funciones de 0_funciones----------------------------------------------
if (!exists("se_ha_cargado_f")) {
suppressMessages(source_rmd("0_b_funciones.rmd"))
}

```

```{r datos}
# cargar datos----------------------------------------------------------------
pobana_0 <- readRDS(file = file.path("Datos", "Procesados", "pobana_0.RDS"))
# 645209 

test220613 <- fread(file.path("Datos", "Brutos",
                                   "poblacion_Pruebas_dtcas_06_2022.csv"),
                        col.names = c(
                         "tori",
                         "id",
                         "tfreg",
                         "tfval",
                         "ct",
                         "tmot",
                         "tres",
                         "tcen",
                         "tamb",
                         "tprof",
                         "tmot2")) # 1633910

```

# Agregrar test a pobana_0

***

Nos interesa sacar o bien un basal (antes de mayo 2020), o bien hacer algo dinamico, de modo que lo que guardemos sea las pcr realizadas antes de estar infectado o vacunarse, o 28/05/21.

```{r datos}
# Escoger pruebas con pcr-antigeno---------------------------------------------
test220613 %>% count(tori)

test0 <- test220613 %>%
  # quitar algunas variables
  select(-c(tamb, tmot, tcen)) %>% 
  # seleccionar PCR y antígenos
  filter(tori %in% c("PCR","Antigenos"))  # 1627129

```


```{r repetidos}
# un test día por persona------------------------------------------------------ 
test1 <- test0 %>% 
  mutate(id = as.character(id), 
         tpos = tres == "Positivo", 
         profr = tprof == "Profesional") %>%
  select( -c(tres)) %>% 
  mutate(freg = as.Date(tfreg)) %>% 
  mutate(fval = as.Date(tfval)) %>% 
  # seleccionar fecha de registro como fecha del test
  mutate(ftes = freg) %>% 
  select(-tfreg, -tfval) %>% 
  # seleccionar como mucho un test por persona y día
  group_by(id, ftes) %>%
  # primar si hay un test positivo el mismo día que un negativo
  slice(which.max(tpos))  %>% 
  ungroup() %>%
  tidylog::distinct(id, ftes, .keep_all = TRUE) 
  
```

```{r}
# Contar número de test realizado----------------------------------------------
test220613_cleaned <- test1 %>% 
  group_by(id) %>%
  arrange(ftes) %>%
  mutate(n = 1:n()) %>% 
  ungroup()

```


```{r}
# guardar base test------------------------------------------------------------
saveRDS(test220613_cleaned, file = file.path("Datos", "Procesados", 
                                             "test220613_cleaned.RDS"))

```
