==> 1_d_aclaracion_fechas_test_covid_ingresos.Rmd <==
---
title: "test, fcv1 y fih1"
output: 
  html_document:
    code_folding: show
    highlight: espresso
    css: 
    keep_md: yes
    number_sections: yes
    theme: readable
    toc: yes  
    toc_float:
      collapsed: yes
      smooth_scroll: yes
      toc_depth: 3     
editor_options: 
  chunk_output_type: inline
---


```{r set-global-options, echo = FALSE}
knitr::opts_chunk$set(eval = TRUE, 
                      echo = TRUE, 
                      cache = FALSE,
                      include = TRUE,
                      collapse = FALSE,
                       dependson = NULL,
                      engine = "R", # Chunks will always have R code, unless noted
                      error = TRUE,
                      fig.path="Figures/",  # Set the figure options
                      fig.align = "center", 
                      fig.width = 12,
                      fig.height = 12)
```

```{r librerias, echo=FALSE, message=FALSE, include=FALSE, warning=FALSE}
# source("0_a_librerias.R")
# ya se cargan cuando se carguen las funciones

```

```{r}
# función para cargar un .Rmd como si se hiciese source de un .r---------------
source_rmd = function(file, ...) {
  tmp_file = tempfile(fileext = ".R")
  on.exit(unlink(tmp_file), add = TRUE)
  knitr::purl(file, output = tmp_file)
  source(file = tmp_file, ...)
}

```

```{r}
# cargar funciones de 0_funciones----------------------------------------------
if (!exists("se_ha_cargado_f")) {
suppressMessages(source_rmd("0_b_funciones.rmd"))
}

```

# Cargar bases

***

Se cargan las bases originales: 

```{r cargar-datos, echo=FALSE, eval=TRUE, message=FALSE, warning=FALSE}
# cargar datos-----------------------------------------------------------------
poblacion_Pruebas_dtcas_06_2022 <- fread(file = file.path("Datos", "Brutos",
                                   "poblacion_Pruebas_dtcas_06_2022.csv"))
poblacion_Vacunas_Infecc_06_2022 <- fread(file = file.path("Datos", "Brutos", 
                                   "poblacion_Vacunas_Infecc_06_2022.csv"))

```

```{r}
# inicio omicron navarra-------------------------------------------------------
tiempo_0 <- ymd("2022-01-03")

```

# Comprobar fechas


```{r}
poblacion <- poblacion_Vacunas_Infecc_06_2022 %>% 
  clean_names("snake") %>% 
  mutate(across(starts_with("fecha"), fymd))

```


```{r}
# fcv1 > falta ó fcv1 es NA
fechas_ingreso_discrepantes_con_fecha_covid <- poblacion %>% 
  filter(fecha_ini_covid_1 > fecha_alta_1 | is.na(fecha_ini_covid_1) & !is.na(fecha_ingreso_1)) %>% 
  select(pac_number_unif_cod, fecha_ini_covid_1, fecha_ini_covid_2, 
         fecha_ingreso_1, fecha_alta_1) %>% arrange(fecha_ingreso_1) %>%
  relocate(
    pac_number_unif_cod,
           fecha_ingreso_1,
           fecha_alta_1,
    fecha_ini_covid_1,
           fecha_ini_covid_2)

fwrite(fechas_ingreso_discrepantes_con_fecha_covid, file.path("Resultados", 
       "fechas_ingreso_discrepantes_con_fecha_covid.csv"))

```

```{r}
# fcv1 > fingreso
fechas_ingreso_con_fecha_covid_tardia <- poblacion %>% 
  filter(fecha_ini_covid_1 > fecha_ingreso_1,
         fecha_ini_covid_1 <= fecha_alta_1) %>% 
  select(pac_number_unif_cod, fecha_ini_covid_1, fecha_ini_covid_2, 
         fecha_ingreso_1, fecha_alta_1) %>% arrange(fecha_ingreso_1) %>%
  relocate(
    pac_number_unif_cod,
           fecha_ingreso_1,
           fecha_alta_1,
    fecha_ini_covid_1,
           fecha_ini_covid_2) %>% 
  mutate(dias_desde_ingreso_hasta_fecha_covid = fecha_ini_covid_1 - fecha_ingreso_1) %>% 
  arrange(desc(dias_desde_ingreso_hasta_fecha_covid))

fechas_ingreso_con_fecha_covid_tardia %>% 
  tidylog::filter(dias_desde_ingreso_hasta_fecha_covid <= 5)

fwrite(fechas_ingreso_con_fecha_covid_tardia, file.path("Resultados", 
       "fechas_ingreso_con_fecha_covid_tardia.csv"))

```

```{r}
poblacion2 <- poblacion %>% 
  tidylog::mutate(fih1_new = if_else(fecha_ini_covid_1 > fecha_ingreso_1 %m+% days(5) |
                              is.na(fecha_ini_covid_1), NA_Date_, fecha_ingreso_1)) %>% 
  tidylog::mutate(fcv1_new = if_else(!is.na(fih1_new), fih1_new, fecha_ini_covid_1))

```


```{r}
# check------------------------------------------------------------------------
# fcv1 > falta ó fcv1 es NA
poblacion2 %>% 
  filter(fecha_ini_covid_1 > fecha_alta_1 | is.na(fcv1_new) & !is.na(fih1_new)) %>% 
  select(pac_number_unif_cod, fcv1_new, fecha_ini_covid_2, 
         fih1_new, fecha_alta_1) %>% arrange(fih1_new) %>%
  relocate(
    pac_number_unif_cod,
           fih1_new,
           fecha_alta_1,
    fcv1_new,
           fecha_ini_covid_2)

# fcv1 > fingreso
poblacion2 %>% 
  filter(fcv1_new > fih1_new,
         fcv1_new <= fecha_alta_1) %>% 
  select(pac_number_unif_cod, fcv1_new, fecha_ini_covid_2, 
         fih1_new, fecha_alta_1) %>% arrange(fih1_new) %>%
  relocate(
    pac_number_unif_cod,
           fih1_new,
           fecha_alta_1,
    fcv1_new,
           fecha_ini_covid_2) 

```

# Crear pobana_0b

```{r cargar-datos, echo=FALSE, eval=TRUE, message=FALSE, warning=FALSE}
# cargar datos-----------------------------------------------------------------
pobana_0 <- readRDS(file = file.path("Datos", "Procesados", "pobana_0.RDS"))
test220613_cleaned <- readRDS(file = file.path("Datos", "Procesados", 
                                             "test220613_cleaned.RDS"))

```

Se excluyen los ingresos donde han pasado más de 5 días desde la fecha del ingreso hasta fcv1.

Se actualiza la fcv1 con la fih1 correspondiente. Las infecciones nosocomiales se tendrán en cuenta como infección, pero no como ingreso.

```{r}
# fecha covid por test---------------------------------------------------------
pobana_0b <- pobana_0 %>%
  rename(fih1_legacy = fih1,
         fcv1_legacy = fcv1) %>% 
# crear fcv1_test
  left_join(
  test220613_cleaned %>% 
  filter(tpos == TRUE) %>% 
  group_by(id) %>% 
  summarise(fcv1_test = min(freg))) %>%  
  # arreglar base
  mutate(fih1 = if_else(fcv1_legacy > fih1_legacy %m+% days(5) |
                      is.na(fcv1_legacy), NA_Date_, fih1_legacy)) %>% 
  mutate(fcv1 = if_else(!is.na(fih1), fih1, fcv1_test)) %>% 
  mutate(fcv1_legacy2 = if_else(!is.na(fih1), fih1, fcv1_legacy))

```


```{r}
# check pobana_0b
pobana_0b %>% filter(!is.na(fcv1_test))
pobana_0b %>% filter(!is.na(fcv1_legacy))
pobana_0b %>% filter(!is.na(fih1), is.na(fcv1))

```

```{r pobana_0}
# guardar base de pobana antes de hacer las exclusiones------------------------
saveRDS(pobana_0b, file = file.path("Datos", "Procesados", "pobana_0b.RDS"))

```

# Marcar muerte por COVID-19

```{r}
# muerte por todas las causas en los 30 días posteriores a una PDIA------------
pobana_0c <- pobana_0b |> 
  # crear fecha de muerte por covid
  mutate(fdef_covid = if_else(fdef >= fcv1 & fdef <= fcv1 + days(30), 
                              fdef, NA_Date_)) |> 
  # composite fih y fded_covid
  mutate(fih_def = pmin(fdef_covid, fih1, na.rm = TRUE))

# check numbers
pobana_0c |> count(is.na(fdef_covid), is.na(fih1))
pobana_0c |> count(is.na(fih_def), is.na(fdef_covid), is.na(fih1))


```



```{r pobana_0}
# guardar base de pobana antes de hacer las exclusiones------------------------
saveRDS(pobana_0c, file = file.path("Datos", "Procesados", "pobana_0c.RDS"))

```
