==> 2_c_matching_multiples_tt.Rmd <==
---
title: "Múltiples tiempos + test"
output: 
  html_document:
    code_folding: show
    highlight: espresso
    css:
    keep_md: yes
    number_sections: yes
    theme: readable
    toc: yes  
    toc_float:
      collapsed: yes
      smooth_scroll: yes
      toc_depth: 3     
editor_options: 
  chunk_output_type: inline
---


```{r set-global-options, echo = FALSE}
knitr::opts_chunk$set(eval = TRUE, 
                      echo = TRUE, 
                      cache = FALSE,
                      include = TRUE,
                      collapse = FALSE,
                       dependson = NULL,
                      engine = "R", # Chunks will always have R code, unless noted
                      error = TRUE,
                      fig.path="Figures/",  # Set the figure options
                      fig.align = "center", 
                      fig.width = 12,
                      fig.height = 12)
```

```{r}
# función para cargar un .Rmd como si se hiciese source de un .r---------------
source_rmd = function(file, ...) {
  tmp_file = tempfile(fileext = ".R")
  on.exit(unlink(tmp_file), add = TRUE)
  knitr::purl(file, output = tmp_file)
  source(file = tmp_file, ...)
}

```

```{r}
# cargar funciones de 0_funciones----------------------------------------------
if (!exists("se_ha_cargado_f")) {
suppressMessages(source_rmd("0_b_funciones.rmd"))
}

```


```{r cargar-datos, echo=FALSE, eval=TRUE, message=FALSE, warning=FALSE}
# cargar datos-----------------------------------------------------------------
pobana_2 <- readRDS(file = file.path("Resultados", "pobana_2.RDS"))
test220613_cleaned <- readRDS(file = file.path("Datos", "Procesados", 
                                             "test220613_cleaned.RDS"))

```





```{r}
# se crean listas que dependen del proyecto------------------------------------
lista_exclusion_dep <- list(
  criterios = list(
  c1 = "Excluyendo: infección previa antes del tiempo_0...",
  c2 = "Excluyendo: edad < 40 años en tiempo_0...", 
  c3 = "Excluyendo: pauta incompleta 90 días antes del tiempo_0...",
  c4 = "Excluyendo: fallecidos antes del tiempo_0...",
  c5 = "Excluyendo: boosting antes del tiempo_0..."),
  variables = list(
  v1 = "fcv1",
  v2 = "fnac_40",
  v3 = "fpauta_90",
  v4 = "fdef",
  v5 = "fvc3"),
  tipo_var = list(
    "exc",
    "inc",
    "inc",
    "exc",
    "exc")
  )

```

# variables de matching

```{r}
# vectores de matching---------------------------------------------------------
variables_matching = c(
  "gen",    # 1
  "gedad",  # 2
  "czbs",   # 3
  "tvc1",   # 4
  "migran", # 5
  "vgp20",  # 6
  "vgp21",  # 7
  "ucn8",   # 8
  "domi")   # 9

```

```{r}
# comprobar comorbilidades
# todos
pobana_2 %>% tabyl(diab)
# pobana_2 %>% tabyl(ainm)
pobana_2 %>% tabyl(corp)
pobana_2 %>% tabyl(irc)
pobana_2 %>% tabyl(epoc)
pobana_2 %>% tabyl(hlip)
pobana_2 %>% tabyl(hta)
pobana_2 %>% tabyl(ictu)
pobana_2 %>% tabyl(obes)

# con ingreso
pobana_3 <- pobana_2 %>% filter(!is.na(fih1))
  
pobana_3 %>% tabyl(diab)
# pobana_3 %>% tabyl(ainm)
pobana_3 %>% tabyl(corp)
pobana_3 %>% tabyl(irc)
pobana_3 %>% tabyl(epoc)
pobana_3 %>% tabyl(hlip)
pobana_3 %>% tabyl(hta)
pobana_3 %>% tabyl(ictu)
pobana_3 %>% tabyl(obes)


```

# Prueba con una fecha

```{r}
# matching f_pauta en días
prueba_matching_days <- f_matching_tt(
                data = pobana_2,
                tiempo_0 = "2022-01-03",
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fcv1", 
                vector_match = variables_matching[c(1, 3:8)], 
                base_test = test220613_cleaned,
                temporal_var = "fpauta_90",
                temporal_var_unit = "days",
                birth_date = "fnac",
                age_first = 40,
                age_last = 90,
                age_by = 5,
                censoring = "fdef",
                follow_up = 35,
                verbose = FALSE,
                seed = 3)

prueba_matching_days

# matching f_pauta en semanas (menos restrictivo)
prueba_matching_weeks <- f_matching_tt(
                data = pobana_2,
                tiempo_0 = "2022-01-03",
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fcv1", 
                vector_match = variables_matching[c(1, 3:8)], 
                base_test = test220613_cleaned,
                temporal_var = "fpauta_90",
                temporal_var_unit = "weeks",
                birth_date = "fnac",
                age_first = 40,
                age_last = 90,
                age_by = 5,
                censoring = "fdef",
                follow_up = 35,
                verbose = FALSE,
                seed = 3)

prueba_matching_weeks

# comprobar categorías de edad
prueba_matching_weeks %>% 
tabyl(p_edad_cat_label, exposicion) %>% 
  adorn_title()

```


```{r}
# check prueba-----------------------------------------------------------------
# comprobar todos tiempos positivos
prueba_matching_weeks %>% 
  mutate(survtime = date_end - date_start) %>% 
  arrange(survtime) # ok
# comprobar follow-up se hace bien
prueba_matching_weeks %>% 
  mutate(survtime = date_end - date_start) %>% 
  arrange(desc(survtime)) #ok

prueba_matching_weeks %>% tabyl(exposicion, caso)

```

# Prueba con más de una fecha

```{r}
# matching f_pauta en días (menos restrictivo)
prueba_matching_mult_days <- f_matching_tt_mult(
                data = pobana_2,
                f_ini = "2022-01-03",
                f_fin = "2022-01-05",
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fcv1", 
                vector_match = variables_matching[c(1, 3:8)], 
                base_test = test220613_cleaned,
                temporal_var = "fpauta_90",
                temporal_var_unit = "days",
                birth_date = "fnac",
                age_first = 40,
                age_last = 90,
                age_by = 5,
                censoring = "fdef",
                follow_up = 35,
                verbose = FALSE,
                seed = 3)

prueba_matching_mult_days

# matching f_pauta en semanas (menos restrictivo)
prueba_matching_mult_weeks <- f_matching_tt_mult(
                data = pobana_2,
                f_ini = "2022-01-03",
                f_fin = "2022-01-05",
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fcv1", 
                vector_match = variables_matching[c(1, 3:8)], 
                base_test = test220613_cleaned,
                temporal_var = "fpauta_90",
                temporal_var_unit = "weeks",
                birth_date = "fnac",
                age_first = 40,
                age_last = 90,
                age_by = 5,
                censoring = "fdef",
                follow_up = 35,
                verbose = FALSE,
                seed = 3)

```

```{r}
# comprobar todos tiempos positivos
prueba_matching_mult_weeks$tt_tibble %>% 
  mutate(survtime = date_end - date_start) %>% 
  arrange(survtime) # ok
# comprobar follow-up se hace bien
prueba_matching_mult_weeks$tt_tibble %>% 
  mutate(survtime = date_end - date_start) %>% 
  arrange(desc(survtime)) #ok

```

```{r}
# guardar resultados-----------------------------------------------------------
saveRDS(prueba_matching_mult_weeks, file.path("Resultados",
        "prueba_matching_mult_weeks.RDS"))

```

# TT Monge-Hernán

Se replica el análisis realizado en España en el cuál se investiga la efectividad del booster frente a la infección.

## fcv1

Las fechas que ponen en el artículo son del *2022-01-03* al *2022-02-06*.

### Análisis principal

Se utilizan las variables de matching y criterios de exclusión del **análisis principal** como en el artículo.

```{r}
# hacer ajuste como en el artículo
matching_fcv1 <- f_matching_tt_mult(
                data = pobana_2,
                f_ini = "2022-01-03",
                f_fin = "2022-02-06",
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fcv1", 
                vector_match = variables_matching[c(1, 3:8)], 
                base_test = test220613_cleaned,
                temporal_var = "fpauta_90",
                temporal_var_unit = "weeks",
                birth_date = "fnac",
                age_first = 40,
                age_last = 90,
                age_by = 5,
                censoring = "fdef",
                follow_up = 35,
                verbose = FALSE,
                seed = 3)

# guardar resultados-----------------------------------------------------------
saveRDS(matching_fcv1, file.path("Resultados",
        "matching_fcv1.RDS"))

```

### Sensibilidad 0

Se exige que al menos tengan un test diagnóstico realizado en el momento del reclutamiento. No es necesario volver a ejecutar la función *f_matching_tt_mult*. Se hace un filtrado sobre los resultados de *matching_fcv1*.

### Sensibilidad 1

Se excluyen a los individuos con una prueba diagnóstica la semana antes del reclutamiento.

```{r}
# hacer ajuste con el domicilio como matching
matching_fcv1_sens1 <- f_matching_tt_mult(
                data = pobana_2,
                f_ini = "2022-01-03",
                f_fin = "2022-02-06",
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fcv1", 
                vector_match = variables_matching[c(1, 3:8)], 
                base_test = test220613_cleaned,
                dias_a_excluir_test_previos = 7,
                temporal_var = "fpauta_90",
                temporal_var_unit = "weeks",
                birth_date = "fnac",
                age_first = 40,
                age_last = 90,
                age_by = 5,
                censoring = "fdef",
                follow_up = 35,
                verbose = FALSE,
                seed = 3)

# guardar resultados-----------------------------------------------------------
saveRDS(matching_fcv1_sens1, file.path("Resultados",
        "matching_fcv1_sens1.RDS"))

```
### Sensibilidad 2

Se censuran las parejas 7 días después de que el control se vacune (en lugar de el día en el que control se vacuna, como sucede en el análisis principal).

```{r}
# hacer ajuste con el domicilio como matching
matching_fcv1_sens2 <- f_matching_tt_mult(
                data = pobana_2,
                f_ini = "2022-01-03",
                f_fin = "2022-02-06",
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fcv1", 
                vector_match = variables_matching[c(1, 3:8)], 
                base_test = test220613_cleaned,
                dias_a_excluir_test_previos = NULL,
                temporal_var = "fpauta_90",
                temporal_var_unit = "weeks",
                birth_date = "fnac",
                age_first = 40,
                age_last = 90,
                age_by = 5,
                censoring = "fdef",
                delayed_exposure_censoring = 7,
                follow_up = 35,
                verbose = FALSE,
                seed = 3)

# guardar resultados-----------------------------------------------------------
saveRDS(matching_fcv1_sens2, file.path("Resultados",
        "matching_fcv1_sens2.RDS"))

```

### Sensibilidad 3

Sumarle a la fecha *fcv1* dos días.

```{r}
matching_fcv1_sens3 <- f_matching_tt_mult(
                data = pobana_2 %>% 
                  mutate(fcv1 = fcv1 %m+% days(2)),
                f_ini = "2022-01-03",
                f_fin = "2022-02-06",
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fcv1", 
                vector_match = variables_matching[c(1, 3:8)], 
                base_test = test220613_cleaned,
                dias_a_excluir_test_previos = NULL,
                temporal_var = "fpauta_90",
                temporal_var_unit = "weeks",
                birth_date = "fnac",
                age_first = 40,
                age_last = 90,
                age_by = 5,
                censoring = "fdef",
                delayed_exposure_censoring = NULL,
                follow_up = 35,
                verbose = FALSE,
                seed = 3)

# guardar resultados-----------------------------------------------------------
saveRDS(matching_fcv1_sens3, file.path("Resultados",
        "matching_fcv1_sens3.RDS"))

```

### Sensibilidad 4

Matching sin reemplazamiento.

```{r}
matching_fcv1_sens4 <- f_matching_tt_mult(
                data = pobana_2,
                f_ini = "2022-01-03",
                f_fin = "2022-02-06",
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fcv1", 
                vector_match = variables_matching[c(1, 3:8)],
                match_with_replacement = FALSE,
                base_test = test220613_cleaned,
                dias_a_excluir_test_previos = NULL,
                temporal_var = "fpauta_90",
                temporal_var_unit = "weeks",
                birth_date = "fnac",
                age_first = 40,
                age_last = 90,
                age_by = 5,
                censoring = "fdef",
                delayed_exposure_censoring = NULL,
                follow_up = 35,
                verbose = FALSE,
                seed = 3)

# guardar resultados-----------------------------------------------------------
saveRDS(matching_fcv1_sens4, file.path("Resultados",
        "matching_fcv1_sens4.RDS"))

```

### Sensibilidad 5

Matching sin reemplazamiento.

```{r}
matching_fcv1_sens5 <- f_matching_tt_mult(
                data = pobana_2,
                f_ini = "2022-01-03",
                f_fin = "2022-02-06",
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fcv1", 
                vector_match = variables_matching[c(1, 3:8)],
                match_with_replacement = FALSE,
                base_test = test220613_cleaned,
                dias_a_excluir_test_previos = NULL,
                temporal_var = "fpauta_90",
                temporal_var_unit = "weeks",
                birth_date = "fnac",
                age_first = 40,
                age_last = 90,
                age_by = 1,
                censoring = "fdef",
                delayed_exposure_censoring = NULL,
                follow_up = 35,
                verbose = FALSE,
                seed = 3)

# guardar resultados-----------------------------------------------------------
saveRDS(matching_fcv1_sens5, file.path("Resultados",
        "matching_fcv1_sens5.RDS"))

```

### fcv1 Matching 1: Domicilio

Domicilios con 3 o más convivientes.

```{r}
pobana_2 %>% count(ucn8)
pobana_2 %>% add_count(domi, name = "ucn") 

```


```{r}
# hacer ajuste con el domicilio como matching----------------------------------
matching_fcv1_domi_1 <- f_matching_tt_mult(
                data = pobana_2 %>% filter(ucn8 >= 3) ,
                f_ini = "2022-01-03",
                f_fin = "2022-02-06",
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fcv1", 
                vector_match = "domi", 
                base_test = NULL,
                temporal_var = NULL,
                temporal_var_unit = NULL,
                birth_date = NULL,
                age_first = NULL,
                age_last = NULL,
                age_by = NULL,
                censoring = "fdef",
                follow_up = 35,
                verbose = FALSE,
                seed = 3)

# guardar resultados-----------------------------------------------------------
saveRDS(matching_fcv1_domi_1, file.path("Resultados",
        "matching_fcv1_domi_1.RDS"))

```


### fcv1 Matching 2: Domicilio

Domicilio con 2 o más convivientes

```{r}
# hacer ajuste con el domicilio como matching----------------------------------
matching_fcv1_domi_2 <- f_matching_tt_mult(
                data = pobana_2 %>% 
                  filter(ucn8 >= 2) ,
                f_ini = "2022-01-03",
                f_fin = "2022-02-06",
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fcv1", 
                vector_match = "domi", 
                base_test = NULL,
                temporal_var = NULL,
                temporal_var_unit = NULL,
                birth_date = NULL,
                age_first = NULL,
                age_last = NULL,
                age_by = NULL,
                censoring = "fdef",
                follow_up = 35,
                verbose = FALSE,
                seed = 3)

# guardar resultados-----------------------------------------------------------
saveRDS(matching_fcv1_domi_2, file.path("Resultados",
        "matching_fcv1_domi_2.RDS"))

```

### fcv1 Matching 3: Domicilio

Domicilio con 2 o más convivientes y excluímos a los que aparecen como 30 o más convivientes.

```{r}
# hacer ajuste con el domicilio como matching----------------------------------
matching_fcv1_domi_3 <- f_matching_tt_mult(
                data = pobana_2 %>% 
                  add_count(domi, name = "ucn") %>% 
                  filter(ucn8 >= 2) %>%
                  filter(ucn < 30),
                f_ini = "2022-01-03",
                f_fin = "2022-02-06",
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fcv1", 
                vector_match = "domi", 
                base_test = NULL,
                temporal_var = NULL,
                temporal_var_unit = NULL,
                birth_date = NULL,
                age_first = NULL,
                age_last = NULL,
                age_by = NULL,
                censoring = "fdef",
                follow_up = 35,
                verbose = FALSE,
                seed = 3)

# guardar resultados-----------------------------------------------------------
saveRDS(matching_fcv1_domi_3, file.path("Resultados",
        "matching_fcv1_domi_3.RDS"))

```

### fcv1 Matching 4: Domicilio

Domicilio con 3 o más convivientes y excluímos a los que aparecen como 30 o más convivientes.

```{r}
# hacer ajuste con el domicilio como matching----------------------------------
matching_fcv1_domi_4 <- f_matching_tt_mult(
                data = pobana_2 %>% 
                  add_count(domi, name = "ucn") %>% 
                  filter(ucn8 >= 3) %>%
                  filter(ucn < 30),
                f_ini = "2022-01-03",
                f_fin = "2022-02-06",
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fcv1", 
                vector_match = "domi", 
                base_test = NULL,
                temporal_var = NULL,
                temporal_var_unit = NULL,
                birth_date = NULL,
                age_first = NULL,
                age_last = NULL,
                age_by = NULL,
                censoring = "fdef",
                follow_up = 35,
                verbose = FALSE,
                seed = 3)

# guardar resultados-----------------------------------------------------------
saveRDS(matching_fcv1_domi_4, file.path("Resultados",
        "matching_fcv1_domi_4.RDS"))

```

# Cálculos The Lancet

***

## fcv1 más tiempo. Matching: Variables como en el artículo

```{r}
# follow-up = 34,
matching_fcv1_longer_34 <- f_matching_tt_mult(
                data = pobana_2,
                f_ini = "2022-01-03",
                f_fin = "2022-03-30",
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fcv1", 
                vector_match = variables_matching[c(1, 3:8)], 
                base_test = test220613_cleaned,
                temporal_var = "fpauta_90",
                temporal_var_unit = "weeks",
                birth_date = "fnac",
                age_first = 40,
                age_last = 90,
                age_by = 5,
                censoring = "fdef",
                follow_up = 34,
                verbose = FALSE,
                seed = 3,
                fecha_fin_estudio = ymd("2022-03-31"))

# guardar resultados-----------------------------------------------------------
saveRDS(matching_fcv1_longer_34, file.path("Resultados",
        "matching_fcv1_longer_34.RDS"))

```

```{r}
# follow-up = 70,
matching_fcv1_longer_70 <- f_matching_tt_mult(
                data = pobana_2,
                f_ini = "2022-01-03",
                f_fin = "2022-03-30",
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fcv1", 
                vector_match = variables_matching[c(1, 3:8)], 
                base_test = test220613_cleaned,
                temporal_var = "fpauta_90",
                temporal_var_unit = "weeks",
                birth_date = "fnac",
                age_first = 40,
                age_last = 90,
                age_by = 5,
                censoring = "fdef",
                follow_up = 70,
                verbose = FALSE,
                seed = 3,
                fecha_fin_estudio = ymd("2022-03-31"))

# guardar resultados-----------------------------------------------------------
saveRDS(matching_fcv1_longer_70, file.path("Resultados",
        "matching_fcv1_longer_70.RDS"))

```

## fcv1 más tiempo. Matching: Domicilio

```{r}
# hacer ajuste con el domicilio como matching----------------------------------
matching_fcv1_domi_34 <- f_matching_tt_mult(
                data = pobana_2 %>% filter(ucn8 >= 2) ,
                f_ini = "2022-01-03",
                f_fin = "2022-03-30",
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fcv1", 
                vector_match = "domi", 
                base_test = NULL,
                temporal_var = NULL,
                birth_date = NULL,
                calcular_edad = "fnac",
                calcular_test = test220613_cleaned,
                calcular_temp_var = "fpauta_90",
                temporal_var_unit = "weeks",
                age_first = 40,
                age_last = 90,
                age_by = 5,
                censoring = "fdef",
                follow_up = 34,
                verbose = FALSE,
                seed = 3,
                fecha_fin_estudio = ymd("2022-03-31"))

# guardar resultados-----------------------------------------------------------
saveRDS(matching_fcv1_domi_34, file.path("Resultados",
        "matching_fcv1_domi_34.RDS"))

```

```{r}
# hacer ajuste con el domicilio como matching----------------------------------
matching_fcv1_domi_70 <- f_matching_tt_mult(
                data = pobana_2 %>% filter(ucn8 >= 2) ,
                f_ini = "2022-01-03",
                f_fin = "2022-03-30",
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fcv1", 
                vector_match = "domi", 
                base_test = NULL,
                temporal_var = NULL,
                birth_date = NULL,
                calcular_edad = "fnac",
                calcular_test = test220613_cleaned,
                calcular_temp_var = "fpauta_90",
                temporal_var_unit = "weeks",
                age_first = 40,
                age_last = 90,
                age_by = 5,
                censoring = "fdef",
                follow_up = 70,
                verbose = FALSE,
                seed = 3,
                fecha_fin_estudio = ymd("2022-03-31"))

# guardar resultados-----------------------------------------------------------
saveRDS(matching_fcv1_domi_70, file.path("Resultados",
        "matching_fcv1_domi_70.RDS"))

```

## fih1 más tiempo. Matching por edad y sexo

```{r}
# # fih1
# matching_fih1_longer_34 <- f_matching_tt_mult(
#                 data = pobana_2,
#                 f_ini = "2022-01-03",
#                 f_fin = "2022-06-12",
#                 crit_exc = lista_exclusion_dep,
#                 exposure = "fvc3",
#                 out = "fih1", 
#                 vector_match = variables_matching[c(1)], 
#                 # base_test = test220613_cleaned,
#                 # temporal_var = "fpauta_90",
#                 # temporal_var_unit = "weeks",
#                 birth_date = "fnac",
#                 age_first = 40,
#                 age_last = 90,
#                 age_by = 5,
#                 censoring = "fdef",
#                 follow_up = 34,
#                 verbose = FALSE,
#                 seed = 3,
#                 fecha_fin_estudio = ymd("2022-06-13"))
# 
# # guardar resultados-----------------------------------------------------------
# saveRDS(matching_fih1_longer_34, file.path("Resultados",
#         "matching_fih1_longer_34.RDS"))

```


```{r}
# # follow-up = 70, f_fin = 13-04-2022
# matching_fih1_longer_70 <- f_matching_tt_mult(
#                 data = pobana_2,
#                 f_ini = "2022-01-03",
#                 f_fin = "2022-06-12",
#                 crit_exc = lista_exclusion_dep,
#                 exposure = "fvc3",
#                 out = "fih1", 
#                 vector_match = variables_matching[c(1)], 
#                 # base_test = test220613_cleaned,
#                 # temporal_var = "fpauta_90",
#                 # temporal_var_unit = "weeks",
#                 birth_date = "fnac",
#                 age_first = 40,
#                 age_last = 90,
#                 age_by = 5,
#                 censoring = "fdef",
#                 follow_up = 70,
#                 verbose = FALSE,
#                 seed = 3,
#                 fecha_fin_estudio = ymd("2022-06-13"))
# 
# # guardar resultados-----------------------------------------------------------
# saveRDS(matching_fih1_longer_70, file.path("Resultados",
#         "matching_fih1_longer_70.RDS"))

```

## fih1 más tiempo. Matching: edad, sexo y comorbilidades

```{r}
# fih1
matching_fih1_longer_34_bis <- f_matching_tt_mult(
                data = pobana_2,
                f_ini = "2022-01-03",
                f_fin = "2022-06-12",
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fih1", 
                vector_match = c("gen", "diab", "corp", "irc", "epoc", "hlip",
                                 "hta", "ictu", "obes"),
                # base_test = test220613_cleaned,
                # temporal_var = "fpauta_90",
                calcular_temp_var = "fpauta_90",
                temporal_var_unit = "weeks",
                calcular_test = test220613_cleaned,
                birth_date = "fnac",
                age_first = 40,
                age_last = 90,
                age_by = 5,
                censoring = "fdef",
                follow_up = 34,
                verbose = FALSE,
                seed = 3,
                fecha_fin_estudio = ymd("2022-06-13"))

# guardar resultados-----------------------------------------------------------
saveRDS(matching_fih1_longer_34_bis, file.path("Resultados",
        "matching_fih1_longer_34_bis.RDS"))

```


```{r}
# follow-up = 70, f_fin = 13-04-2022
matching_fih1_longer_70_bis <- f_matching_tt_mult(
                data = pobana_2,
                f_ini = "2022-01-03",
                f_fin = "2022-06-12",
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fih1", 
                vector_match = c("gen", "diab", "corp", "irc", "epoc", "hlip",
                                 "hta", "ictu", "obes"),
                # base_test = test220613_cleaned,
                # temporal_var = "fpauta_90",
                calcular_temp_var = "fpauta_90",
                temporal_var_unit = "weeks",
                calcular_test = test220613_cleaned,
                birth_date = "fnac",
                age_first = 40,
                age_last = 90,
                age_by = 5,
                censoring = "fdef",
                follow_up = 70,
                verbose = FALSE,
                seed = 3,
                fecha_fin_estudio = ymd("2022-06-13"))

# guardar resultados-----------------------------------------------------------
saveRDS(matching_fih1_longer_70_bis, file.path("Resultados",
        "matching_fih1_longer_70_bis.RDS"))

```
