==> 2_e_obtaining_short_term_paper.Rmd <==
---
title: "Paper matching"
output: 
  html_document:
    code_folding: show
    highlight: espresso
    css:
    keep_md: yes
    number_sections: yes
    theme: readable
    toc: yes  
    toc_float:
      collapsed: yes
      smooth_scroll: yes
      toc_depth: 3     
editor_options: 
  chunk_output_type: inline
---


```{r set-global-options, echo = FALSE}
knitr::opts_chunk$set(eval = TRUE, 
                      echo = TRUE, 
                      cache = FALSE,
                      include = TRUE,
                      collapse = FALSE,
                       dependson = NULL,
                      engine = "R", # Chunks will always have R code, unless noted
                      error = TRUE,
                      fig.path="Figures/",  # Set the figure options
                      fig.align = "center", 
                      fig.width = 12,
                      fig.height = 12)
```

```{r}
# función para cargar un .Rmd como si se hiciese source de un .r---------------
source_rmd = function(file, ...) {
  tmp_file = tempfile(fileext = ".R")
  on.exit(unlink(tmp_file), add = TRUE)
  knitr::purl(file, output = tmp_file)
  source(file = tmp_file, ...)
}

```

```{r}
# cargar funciones de 0_funciones----------------------------------------------
if (!exists("se_ha_cargado_f")) {
suppressMessages(source_rmd("0_b_funciones.rmd"))
}

```

# Load mid term results

```{r}
# cargar resultados-----------------------------------------------------------
paper_fcv1 <- readRDS(file.path("Resultados",
        "paper_fcv1.RDS"))

paper_fcv1_domi <- readRDS(file.path("Resultados",
        "paper_fcv1_domi.RDS"))

paper_fih_def <- readRDS(file.path("Resultados",
        "paper_fih_def.RDS"))

```

# Calculare short term results

```{r}
paper_fcv1_short_term <- shorten_tibble(
  result = paper_fcv1, 
  shorter_follow_up = 35, 
  outcome_var = fcv1)

paper_fcv1_domi_short_term <- shorten_tibble(
  result = paper_fcv1_domi, 
  shorter_follow_up = 35, 
  outcome_var = fcv1)

paper_fih_def_short_term <- shorten_tibble(
  result = paper_fih_def, 
  shorter_follow_up = 35, 
  outcome_var = fih_def)

```

# Check results

## Matching variables

```{r}
f_analisis(paper_fcv1, tiempo_efecto = 7)
f_analisis(paper_fcv1_short_term, tiempo_efecto = 7)

```

## Home Adress

```{r}
f_analisis(paper_fcv1_domi, tiempo_efecto = 7)
f_analisis(paper_fcv1_domi_short_term, tiempo_efecto = 7)

```

## Hospitalizations and deaths

```{r}
f_analisis(paper_fih_def, tiempo_efecto = 7)
f_analisis(paper_fih_def_short_term, tiempo_efecto = 7)

```
