==> 3_a_supervivencia.Rmd <==
---
title: "Análisis de supervivencia"
output: 
  html_document:
    code_folding: show
    highlight: espresso
    css: style.css
    keep_md: yes
    number_sections: yes
    theme: readable
    toc: yes  
    toc_float:
      collapsed: yes
      smooth_scroll: yes
      toc_depth: 3     
editor_options: 
  chunk_output_type: inline
---


```{r set-global-options, echo = FALSE}
knitr::opts_chunk$set(eval = TRUE, 
                      echo = TRUE, 
                      cache = FALSE,
                      include = TRUE,
                      collapse = FALSE,
                       dependson = NULL,
                      engine = "R", # Chunks will always have R code, unless noted
                      error = TRUE,
                      fig.path="Figures/",  # Set the figure options
                      fig.align = "center", 
                      fig.width = 12,
                      fig.height = 12)
```

```{r echo=FALSE, message=FALSE, include=FALSE, warning=FALSE}
# source("0_a_librerias.R")
# ya se cargan cuando se carguen las funciones

```

```{r}
# función para cargar un .Rmd como si se hiciese source de un .r---------------
source_rmd = function(file, ...) {
  tmp_file = tempfile(fileext = ".R")
  on.exit(unlink(tmp_file), add = TRUE)
  knitr::purl(file, output = tmp_file)
  source(file = tmp_file, ...)
}

```

```{r}
# cargar funciones de 0_funciones----------------------------------------------
if (!exists("se_ha_cargado_f")) {
suppressMessages(source_rmd("0_b_funciones.rmd"))
}

```

# Análisis de supervivencia Omicrón-España

***

Se realizan los análisis como el artículo de Monge-Hernán

```{r}
# cargar datos tt Monge-Hernán-------------------------------------------------
matching_fcv1 <- readRDS(file.path("Resultados",
        "matching_fcv1.RDS"))
matching_fcv1_sens1 <- readRDS(file.path("Resultados",
        "matching_fcv1_sens1.RDS"))
matching_fcv1_sens2 <- readRDS(file.path("Resultados",
        "matching_fcv1_sens2.RDS"))
matching_fcv1_sens3 <- readRDS(file.path("Resultados",
        "matching_fcv1_sens3.RDS"))
matching_fcv1_sens4 <- readRDS(file.path("Resultados",
        "matching_fcv1_sens4.RDS"))
matching_fcv1_sens5 <- readRDS(file.path("Resultados",
        "matching_fcv1_sens5.RDS"))
# prueba domicilio-------------------------------------------------------------
matching_fcv1_domi_1 <- readRDS(file.path("Resultados",
        "matching_fcv1_domi_1.RDS"))
matching_fcv1_domi_2 <- readRDS(file.path("Resultados",
        "matching_fcv1_domi_2.RDS"))
matching_fcv1_domi_3 <- readRDS(file.path("Resultados",
        "matching_fcv1_domi_3.RDS"))
matching_fcv1_domi_4 <- readRDS(file.path("Resultados",
        "matching_fcv1_domi_4.RDS"))
# longer
matching_fcv1_longer <- readRDS(file.path("Resultados",
        "matching_fcv1_longer.RDS"))
matching_fcv1_domi <- readRDS(file.path("Resultados",
        "matching_fcv1_domi.RDS"))
# hospi
matching_fih1_longer <- readRDS(file.path("Resultados",
        "matching_fih1_longer.RDS"))
matching_fih1_longer_2 <- readRDS(file.path("Resultados",
        "matching_fih1_longer_2.RDS"))

```


## Análisis principal

```{r}
# análisis con variables de matching como el artículo--------------------------
f_analisis(lista = matching_fcv1,
           tiempo_efecto = 7,
           lim_1 = 0.8,
           lim_zoom = 0.9)

```

## Sensibilidad 0

No hace falta ejecutar la f_matching de nuevo. Se hace un filtrado en el tibble de la lista y con eso es suficiente.

```{r}
# análisis con variables de matching como el artículo--------------------------
matching_fcv1_sens0 <- matching_fcv1
matching_fcv1_sens0$tt_tibble <- matching_fcv1_sens0$tt_tibble %>%
  filter(n_0 >= 1)

saveRDS(matching_fcv1_sens0, file.path("Resultados", "matching_fcv1_sens0.RDS"))

f_analisis(lista = matching_fcv1_sens0,
           tiempo_efecto = 7,
           lim_1 = 0.8,
           lim_zoom = 0.9)

```

## Sensibilidad 1

```{r}
# análisis con variables de matching como el artículo--------------------------
f_analisis(lista = matching_fcv1_sens1,
           tiempo_efecto = 7,
           tipo = "event",
           lim_1 = 0.8,
           lim_zoom = 0.9
           )

f_analisis2(lista = matching_fcv1_sens1,
           tiempo_efecto = 7,
           # tipo = "event",
           lim_1 = 0.8,
           lim_zoom = 0.9
           )


```

## Sensibilidad 2

```{r}
# análisis con variables de matching como el artículo--------------------------
f_analisis(lista = matching_fcv1_sens2,
           tiempo_efecto = 7,
           lim_1 = 0.8,
           lim_zoom = 0.9)

```

## Sensibilidad 3

```{r}
# análisis con variables de matching como el artículo--------------------------
f_analisis(lista = matching_fcv1_sens3,
           tiempo_efecto = 7,
           lim_1 = 0.8,
           lim_zoom = 0.9)

```

## Sensibilidad 4

```{r}
# análisis con variables de matching como el artículo--------------------------
f_analisis(lista = matching_fcv1_sens4,
           tiempo_efecto = 7,
           lim_1 = 0.8,
           lim_zoom = 0.9)

```

## Sensibilidad 5

```{r}
# análisis con variables de matching como el artículo--------------------------
f_analisis(lista = matching_fcv1_sens5,
           tiempo_efecto = 7,
           lim_1 = 0.8,
           lim_zoom = 0.9)

```

## Prueba con domicilio 1

```{r}
# análisis con domicilio como variable de matching-----------------------------
f_analisis(lista = matching_fcv1_domi_1,
           tiempo_efecto = 7,
           lim_1 = 0.80,
           lim_zoom = 0.9)

```

## Prueba con domicilio 2

```{r}
# análisis con domicilio como variable de matching-----------------------------
f_analisis(lista = matching_fcv1_domi_2,
           tiempo_efecto = 7,
           lim_1 = 0.80,
           lim_zoom = 0.9)

```

## Prueba con domicilio 3

```{r}
# análisis con domicilio como variable de matching-----------------------------
f_analisis(lista = matching_fcv1_domi_3,
           tiempo_efecto = 7,
           lim_1 = 0.80,
           lim_zoom = 0.9)

```

## Prueba con domicilio 4

```{r}
# análisis con domicilio como variable de matching-----------------------------
f_analisis(lista = matching_fcv1_domi_4,
           tiempo_efecto = 7,
           lim_1 = 0.80,
           lim_zoom = 0.9)

```

# fcv1 Longer

```{r}
# análisis---------------------------------------------------------------------
f_analisis(lista = matching_fcv1_longer,
           tiempo_efecto = 7,
           lim_1 = 0.80,
           lim_zoom = 0.9)

```

# fcv1 Longer. Prueba con domicilio

```{r}
# análisis con domicilio como variable de matching-----------------------------
f_analisis(lista = matching_fcv1_domi,
           tiempo_efecto = 7,
           lim_1 = 0.80,
           lim_zoom = 0.9)

```

# fih1 Longer

Emparejado por edad (grupos de 5 años) y sexo.

```{r}
# análisis---------------------------------------------------------------------
f_analisis(lista = matching_fih1_longer,
           tiempo_efecto = 7,
           lim_1 = 0.997,
           lim_zoom = 0.997)

```

# fih1 Longer_2

Emparejado por el mismo año y sexo.

```{r}
# análisis---------------------------------------------------------------------
f_analisis(lista = matching_fih1_longer_2,
           tiempo_efecto = 7,
           lim_1 = 0.997,
           lim_zoom = 0.997)

```

# Hacer figura para la carta

```{r}
f_analisis_lancet_rr()

```




```{r, fig.height = 4, fig.width = 6.5}
fecha_censura <- "2022-03-31"
  
  base_analisis <- matching_fcv1_longer$tt_tibble %>%
    mutate(date_end = pmin(ymd(fecha_censura), date_end)) %>% 
    mutate(survtime = date_end - date_start)

base_analisis %>% count(survtime)
  
```


```{r, fig.height = 4, fig.width = 6.5}
max_survtime <- base_analisis %>% 
    summarise(max_survtime = as.numeric(max(survtime))) %>% pull()
  
 
  print(base_analisis %>% 
             tabyl(exposicion, caso) %>%
             adorn_totals(where = c('row', 'col')) %>%
             adorn_title())

  fit <- survfit(Surv(survtime, caso) ~ exposicion, data = base_analisis) 
  
  summary(fit)
  
  cpp <- coxph(formula = Surv(survtime, caso) ~ exposicion, 
              data = base_analisis)
  
  tt_tibble_eff_short <- base_analisis %>% 
  # filtrar parejas con seguimiento después del día 7
      group_by(target_trial_number, par) %>% 
      filter(min(survtime) >= 7,
             max(survtime) <= 34) %>%
      ungroup()
  
```


```{r, fig.height = 4, fig.width = 6.5}
gg0 <- ggsurvplot(fit, data = base_analisis,
           fun = "event",
           surv.plot.height	= 0.8,
           risk.table.height	= 0.2,
           palette = "lancet",
           xlab = "Time since start of follow-up (days)",
           break.x.by = 7,
           break.y.by = 0.02,
           font.x = 2,
           ylab = "Cumulative Incidence",
           xlim = c(0, 70),
           ylim = c(0, 1 - 0.8),
           risk.table = "nrisk_cumevents",
           fontsize = 3.8,
           order.legend = TRUE,
           legend = "top",
           legend.labs = c("No booster", "Booster"),
           legend.title = " ",
           title = "Prueba",
           font.title = 16,
           conf.int = FALSE) 
  
    gg0$plot <- gg0$plot + 
      geom_vline(xintercept = 7, 
                          linetype = "dashed", 
                          size = 0.7,
                          colour = "dimgrey") +
      geom_vline(xintercept = 35, 
                          linetype = "dashed", 
                          size = 0.7,
                          colour = "dimgrey") +
      theme(axis.title.x = element_text(size = 14))
  
    gg0$table <- gg0$table + 
      theme(axis.line = element_blank(),
            axis.title.x = element_blank(),
            axis.ticks = element_blank(),
            axis.text.x = element_blank(),
            plot.title = element_text(size = 14, face = "bold")) 

print(gg0)

```


```{r, fig.height = 4, fig.width = 6.5}
# Cálculos de la RR
print("Cálculo del RR")

# arreglar tibble
  tt_tibble_eff_short <- base_analisis %>% 
  # filtrar parejas con seguimiento después del día 7
      group_by(target_trial_number, par) %>% 
      filter(min(survtime) >= tiempo_efecto,
             max(survtime) <= 34) %>%
      ungroup()

eff_short = hacer_grupo_tabla_2(tt_tibble = tt_tibble_eff_short, 
         group_description = title,
         number_of_bootstrap = 500)

# arreglar tibble
  tt_tibble_eff_long <- base_analisis %>% 
  # filtrar parejas con seguimiento después del día 7
      group_by(target_trial_number, par) %>% 
      filter(min(survtime) >= 35,
             max(survtime) <= 70) %>%
      ungroup()


eff_long = hacer_grupo_tabla_2(tt_tibble = tt_tibble_eff_long, 
         group_description = title,
         number_of_bootstrap = 500)

# arreglar tibble
  tt_tibble_eff <- base_analisis %>% 
  # filtrar parejas con seguimiento después del día 7
      group_by(target_trial_number, par) %>% 
      filter(min(survtime) >= 7) %>%
      ungroup()


eff = hacer_grupo_tabla_2(tt_tibble = tt_tibble_eff, 
         group_description = title,
         number_of_bootstrap = 500)

return(list(gg = gg0, eff_short = eff_short, eff_long = eff_long, eff = eff))
}

```

## Figura 1a

```{r, fig.height = 4, fig.width = 6.5}
# hacer matching con las covariables-------------------------------------------
gg1 <- f_analisis_lancet(matching_fcv1_longer,
                         tiempo_efecto = 7,
                         lim_1 = 0.85,
                         title = "A) Infections. Matching for age, sex, previous negative test, primary vaccine type and date, health basic zone, 
                         migrant status, flu vaccinated in 2020 and in 2021, and the number of people at home",
                         fecha_censura = "2022-03-31")
gg1$gg$plot
gg1$eff_short
gg1$eff_long

```

## Figura 1b

```{r}
# hacer matching con el domicilio----------------------------------------------
gg2 <- f_analisis_lancet(matching_fcv1_domi,
                         tiempo_efecto = 7,
                         lim_1 = 0.85,
                         title = "B) Infections. Matching for home address",
                         fecha_censura = "2022-03-31")
gg2$eff

```

## Figura 1c

```{r}
f_analisis_lancet_hosp <- function(
                lista,
                tiempo_efecto = 0,
                title = "Put a fancy title",
                lim_1 = 0.88) {
  
  base_analisis <- lista$tt_tibble %>% 
    mutate(survtime = date_end - date_start)
  
  max_survtime <- base_analisis %>% 
    summarise(max_survtime = as.numeric(max(survtime))) %>% pull()
  
  print(glue("Se está analizando el outcome: {lista$out} \\
             mediante la emulación de un ensayo clínico  \\
             con múltiples inicios de tiempo. Se ha utilizado matching por: \\
             {str_c(lista$vector_match, collapse = ', ')}.
             "))
  
  print(glue("Análisis de supervivencia días 0-{max_survtime}"))  
  print(base_analisis %>% 
             tabyl(exposicion, caso) %>%
             adorn_totals(where = c('row', 'col')) %>%
             adorn_title())

  fit <- survfit(Surv(survtime, caso) ~ exposicion, data = base_analisis) 
  
  
# Hacer figura  
  gg0 <- ggsurvplot(fit, data = base_analisis,
           fun = "event",
           surv.plot.height	= 0.8,
           risk.table.height	= 0.2,
           palette = "lancet",
           xlab = "Time since start of follow-up (days)",
           break.x.by = 7,
           break.y.by = 0.0002,
           font.x = 2,
           ylab = "Cumulative Incidence",
           ylim = c(0, 0.0015),
           xlim = c(0, 70),
           risk.table = "nrisk_cumevents",
           fontsize = 3.8,
           order.legend = TRUE,
           legend = "top",
           legend.labs = c("No booster", "Booster"),
           legend.title = " ",
           title = title,
           font.title = 16,
           conf.int = FALSE) 
  
    gg0$plot <- gg0$plot + 
      geom_vline(xintercept = 7, 
                          linetype = "dashed", 
                          size = 0.7,
                          colour = "dimgrey") +
      # geom_vline(xintercept = 35, 
      #                     linetype = "dashed", 
      #                     size = 0.7,
      #                     colour = "dimgrey") +
      scale_y_continuous(labels = scales::comma, limits = c(0, 0.0012)) +
      theme(axis.title.x = element_text(size = 14))
  
    gg0$table <- gg0$table +
      theme(axis.line = element_blank(),
            axis.title.x = element_blank(),
            axis.ticks = element_blank(),
            axis.text.x = element_blank(),
            plot.title = element_text(size = 14, face = "bold"),
            plot.caption = element_text(size = 12, hjust = 0.5)
            ) 
print(gg0)

print("Cálculo del RR")

# arreglar tibble
  tt_tibble_eff_short <- base_analisis %>% 
  # filtrar parejas con seguimiento después del día 7
      group_by(target_trial_number, par) %>% 
      filter(min(survtime) >= tiempo_efecto,
             max(survtime) <= 34) %>%
      ungroup()

eff_short = hacer_grupo_tabla_2(tt_tibble = tt_tibble_eff_short, 
         group_description = title,
         number_of_bootstrap = 500)

# arreglar tibble
  tt_tibble_eff_long <- base_analisis %>% 
  # filtrar parejas con seguimiento después del día 7
      group_by(target_trial_number, par) %>% 
      filter(min(survtime) >= 35,
             max(survtime) <= 70) %>%
      ungroup()


eff_long = hacer_grupo_tabla_2(tt_tibble = tt_tibble_eff_long, 
         group_description = title,
         number_of_bootstrap = 500)

# arreglar tibble
  tt_tibble_eff <- base_analisis %>% 
  # filtrar parejas con seguimiento después del día 7
      group_by(target_trial_number, par) %>% 
      filter(min(survtime) >= 7) %>%
      ungroup()


eff = hacer_grupo_tabla_2(tt_tibble = tt_tibble_eff, 
         group_description = title,
         number_of_bootstrap = 500)

return(list(gg = gg1, eff_short = eff_short, eff_long = eff_long, eff = eff))
}

```


```{r, fig.height = 4, fig.width = 13}
# hacer matching con el domicilio----------------------------------------------
gg3 <- f_analisis_lancet_hosp(matching_fih1_longer_2,
                         tiempo_efecto = 7,
                         title = "C) Hospitalisations. Matching for age and sex",
                         lim_1 = 0.999)
gg3$eff

```

## Figura 1 compuesta

```{r, fig.height = 12, fig.width = 12}
# componer figura--------------------------------------------------------------
gg_up <- cowplot::plot_grid(
                   gg1$gg$plot,  
                   gg1$gg$table, 
                   gg2$gg$plot,
                   gg2$gg$table,
                   gg3$gg$plot,
                   gg3$gg$table,
                   rel_heights = c(3.5, 1, 3.5, 1, 3.5, 1), nrow = 6)

ggsave("Fig1.svg", height = 13, width = 12, units = "in")

```

## PRUEBAS

```{r, fig.height = 4, fig.width = 13}
# hacer matching con el domicilio----------------------------------------------
gg4 <- f_analisis_lancet(matching_fcv1_longer,
                         tiempo_efecto = 7,
                         title = "C) Hospitalisations. Matching for characteristics*",
                         lim_1 = 0.8,
                         fecha_censura = "2022-03-31")
gg4$eff

```

```{r, fig.height = 4, fig.width = 13}
# hacer matching con el domicilio----------------------------------------------
gg5 <- f_analisis_lancet(matching_fcv1_domi,
                         tiempo_efecto = 7,
                         title = "B) Infections. Matching for home address",
                         lim_1 = 0.88,
                         fecha_censura = "2022-03-31")
gg5$eff

```
