==> 4_a_analisis_tncc.Rmd <==
---
title: "Análisis caso control negativo en el tiempo desde prueba diagnóstica"
output: 
  html_document:
    code_folding: show
    highlight: espresso
    css: 
    keep_md: yes
    number_sections: yes
    theme: readable
    toc: yes  
    toc_float:
      collapsed: yes
      smooth_scroll: yes
      toc_depth: 3     
editor_options: 
  chunk_output_type: inline
---


```{r set-global-options, echo = FALSE}
knitr::opts_chunk$set(eval = TRUE, 
                      echo = TRUE, 
                      cache = FALSE,
                      include = TRUE,
                      collapse = FALSE,
                       dependson = NULL,
                      engine = "R", # Chunks will always have R code, unless noted
                      error = TRUE,
                      fig.path="Figures/",  # Set the figure options
                      fig.align = "center", 
                      fig.width = 12,
                      fig.height = 12)
```

```{r, echo=FALSE, message=FALSE, include=FALSE, warning=FALSE}
# source("0_a_librerias.R")
# ya se cargan cuando se carguen las funciones

```

```{r}
# función para cargar un .Rmd como si se hiciese source de un .r---------------
source_rmd = function(file, ...) {
  tmp_file = tempfile(fileext = ".R")
  on.exit(unlink(tmp_file), add = TRUE)
  knitr::purl(file, output = tmp_file)
  source(file = tmp_file, ...)
}

```

```{r}
# cargar funciones de 0_funciones----------------------------------------------
if (!exists("se_ha_cargado_f")) {
suppressMessages(source_rmd("0_b_funciones.rmd"))
}

```


```{r cargar-datos, echo=FALSE, eval=TRUE, message=FALSE, warning=FALSE}
# cargar datos-----------------------------------------------------------------
pobana_2 <- readRDS(file = file.path("Resultados", "pobana_2.RDS"))
test220613_cleaned <- readRDS(file = file.path("Datos", "Procesados", 
                                             "test220613_cleaned.RDS"))

```

```{r}
# check fecha test-------------------------------------------------------------
test220613_cleaned %>% arrange(freg)
# hay 2 test con fecha 2012, pero son negativos. No problem

```


```{r}
# descriptivo test, fcv1 y fih1------------------------------------------------
test220613_cleaned2 <- test220613_cleaned %>% 
  filter(tpos == TRUE) %>% 
  select(id, ftes) %>%
  group_by(id) %>% 
  mutate(n = 1:n()) %>% 
  ungroup()

test220613_cleaned3 <- test220613_cleaned2 %>%   
pivot_wider(id_cols = id,
            names_from = n,
            names_prefix = "f_test_positivo_",
            values_from = ftes)

pobana_con_test <- pobana_2  %>% 
  select(id, fcv1, fcv2, fih1) %>%
  left_join(test220613_cleaned3)

# comprobar que las fechas están bien------------------------------------------
pobana_con_test %>% filter(!is.na(fcv1), is.na(f_test_positivo_1))
pobana_con_test %>% filter(is.na(fcv1), !is.na(f_test_positivo_1))
# OK

```

# Introducción

***

Para comprobar el funcionamiento del diseño **tncc** se realizan tres análisis:

1) Un análisis donde se estudia a la población general sin aplicar criterios de exclusión dependientes del tiempo (los criterios de exclusión independientes del tiempo sí que se han aplicado: no trabajadores sanitarios ni de educación, no residentes, etc.).

2) Se excluyen a los individuos que no tengan la pauta *primaria* de vacunación (2 dosis de Pfizer, AZ y Moderna, o al menos 1 dosis de Janssen) con una antigüedad de 90 o más días en el *tiempo_0*.

3) Se excluyen a los individuos que **no tengan la pauta *primaria*** de vacunación (2 dosis de Pfizer, AZ y Moderna, o al menos 1 dosis de Janssen) con una antigüedad de 90 o más días en el *tiempo_0*, y, además, se excluyena los individuos que **hayan tenido infección por COVID-19 previa** al *tiempo_0*.

```{r}
# vectores de matching---------------------------------------------------------
variables_matching = c(
  "gen",    # 1
  "gedad",  # 2
  "czbs",   # 3
  "tvc1",   # 4
  "migran", # 5
  "vgp20",  # 6
  "vgp21",  # 7
  "ucn8",   # 8
  "domi")   # 9

```

# Análisis 1 

***

Se estudia a la población general filtrada por criterios de exclusión independientes del tiempo (*Pobana_2*: población sin criterios de exclusión dependientes del tiempo). En esta población, existen 6 categorías posibles de tratamiento:

1) Sin pauta de vacunación previa completa (sin booster), y sin registro de COVID-19 previa.
2) Sin pauta de vacunación previa completa (sin booster), y con registro de COVID-19 previa.
3) Con pauta de vacunación previa completa, sin booster, y sin registro de COVID-19 previa.
4) Con pauta de vacunación previa completa, sin booster, y con registro de COVID-19 previa.
5) Con pauta de vacunación previa completa, con booster, y sin registro de COVID-19 previa.
6) Con pauta de vacunación previa completa, con booster, y con registro de COVID-19 previa.

```{r}
data_out_1 <- f_matching_tncc(
  data = pobana_2,
  tiempo_0 = ymd("2022-01-03"),
  caso_control_var = "tpos",
  caso_control_date = "ftes",
  base_test = test220613_cleaned,
  vector_match = c("gen", "migran", "p_edad_cat"),
  birth_date = "fnac",
  age_first = 40,
  age_last = 90,
  age_by = 5)

data_out_1 %>% count(caso_control)

```

```{r}
lista_out_1 <- f_matching_tncc_mult(
  data = pobana_2,
  f_ini = ymd("2022-01-03"),
  f_fin = ymd("2022-06-13"),
  caso_control_var = "tpos",
  caso_control_date = "ftes",
  base_test = test220613_cleaned,
  vector_match = c("gen", "migran", "p_edad_cat"),
  birth_date = "fnac",
  age_first = 40,
  age_last = 90,
  age_by = 5)

lista_out_1$tt_tibble %>% count(caso_control)

```



```{r}
# logística condicional
analisis_1 <- lista_out_1$tt_tibble %>% 
  mutate(tratamiento = case_when(
    is.na(fvc2) & is.na(fcv1_prev) ~ 1,
    is.na(fvc2) & !is.na(fcv1_prev) ~ 2,
    !is.na(fvc2) & is.na(fvc3) & is.na(fcv1_prev) ~ 3,
    !is.na(fvc2) & is.na(fvc3) & !is.na(fcv1_prev) ~ 4,
    !is.na(fvc3) & is.na(fcv1_prev) ~ 5,
    !is.na(fvc3) & !is.na(fcv1_prev) ~ 6
  )) %>% 
  mutate(tratamiento = as_factor(tratamiento))

analisis_1 %>% skimr::skim(fcv1_prev)

analisis_1 %>% count(tratamiento)

res_analisis_1 <- clogit(caso_control ~ tratamiento + strata(par), data = analisis_1)

saveRDS(res_analisis_1, file.path("Resultados", "res_analisis_1.RDS"))

```

```{r}
# resultados del análisis 1
summary(res_analisis_1)

```

# Análisis 2

***

Se estudia a mayores de 40 con pauta previa completa. Por lo tanto, van a existir 4 categorías posibles de tratamiento:

1) Sin booster y sin registro de COVID-19 previa.
2) Sin booster y con registro de COVID-19 previa.
3) Con booster y sin registro de COVID-19 previa.
4) Con booster y con registro de COVID-19 previa.

```{r}
# se crean listas que dependen del proyecto------------------------------------
lista_exclusion_dep_analisis2 <- list(
  criterios = list(
  c1 = "Excluyendo: edad < 40 años en tiempo_0...", 
  c2 = "Excluyendo: pauta incompleta 90 días antes del tiempo_0..."
  ),
  variables = list(
  v1 = "fnac_40",
  v2 = "fpauta_90"
  ),
  tipo_var = list(
    "inc",
    "inc"
    )
  )

```


```{r}
data_out_2 <- f_matching_tncc(
  data = pobana_2,
  tiempo_0 = "2022-01-03",
  caso_control_var = "tpos",
  caso_control_date = "ftes",
  base_test = test220613_cleaned,
  vector_match = c("gen", "migran", "p_edad_cat"),
  crit_exc = lista_exclusion_dep_analisis2,
  birth_date = "fnac",
  age_first = 40,
  age_last = 90,
  age_by = 5)

data_out_2 %>% count(caso_control)

```

```{r}
lista_out_2 <- f_matching_tncc_mult(
  data = pobana_2,
  crit_exc = lista_exclusion_dep_analisis2,
  f_ini = "2022-01-03",
  f_fin = "2022-06-13",
  caso_control_var = "tpos",
  caso_control_date = "ftes",
  base_test = test220613_cleaned,
  vector_match = c("gen", "migran", "p_edad_cat"),
  birth_date = "fnac",
  age_first = 40,
  age_last = 90,
  age_by = 5,
  verbose = FALSE)

lista_out_2$tt_tibble %>% count(caso_control)

```



```{r}
# logística condicional
analisis_2 <- lista_out_2$tt_tibble %>% 
  mutate(tratamiento = case_when(
     is.na(fvc3) & is.na(fcv1_prev) ~  1,
     is.na(fvc3) & !is.na(fcv1_prev) ~ 2,
    !is.na(fvc3) & is.na(fcv1_prev) ~  3,
    !is.na(fvc3) & !is.na(fcv1_prev) ~ 4
  )) %>% 
  mutate(tratamiento = as_factor(tratamiento))

analisis_2 %>% skimr::skim(fcv1_prev)

analisis_2 %>% count(tratamiento)

res_analisis_2 <- clogit(caso_control ~ tratamiento + strata(par), data = analisis_2)

saveRDS(res_analisis_2, file.path("Resultados", "res_analisis_2.RDS"))

```

```{r}
# resultados del análisis 2
summary(res_analisis_2)

```

# Análisis 3

***

Se estudia a los mayores de 40, con pauta previa y sin COVID previa. Por lo tanto, únicamente se tienen 2 categorías de tratamiento:

1) Sin booster.
2) Con booster.

```{r}
# se crean listas que dependen del proyecto------------------------------------
lista_exclusion_dep_analisis3 <- list(
  criterios = list(
  c1 = "Excluyendo: infección previa antes del tiempo_0...",
  c2 = "Excluyendo: edad < 40 años en tiempo_0...", 
  c3 = "Excluyendo: pauta incompleta 90 días antes del tiempo_0..."
  ),
  variables = list(
  v1 = "fcv1",
  v2 = "fnac_40",
  v3 = "fpauta_90"
  ),
  tipo_var = list(
    "exc",
    "inc",
    "inc"
    )
  )

```


```{r}
data_out_3 <- f_matching_tncc(
  data = pobana_2,
  tiempo_0 = "2022-01-03",
  caso_control_var = "tpos",
  caso_control_date = "ftes",
  base_test = test220613_cleaned,
  vector_match = c("gen", "migran", "p_edad_cat"),
  crit_exc = lista_exclusion_dep_analisis3,
  birth_date = "fnac",
  age_first = 40,
  age_last = 90,
  age_by = 5)

data_out_3 %>% count(caso_control)

```

```{r}
lista_out_3 <- f_matching_tncc_mult(
  data = pobana_2,
  crit_exc = lista_exclusion_dep_analisis3,
  f_ini = "2022-01-03",
  f_fin = "2022-06-13",
  caso_control_var = "tpos",
  caso_control_date = "ftes",
  base_test = test220613_cleaned,
  vector_match = c("gen", "migran", "p_edad_cat"),
  birth_date = "fnac",
  age_first = 40,
  age_last = 90,
  age_by = 5,
  verbose = FALSE)

lista_out_3$tt_tibble %>% count(caso_control)

```



```{r}
# logística condicional
analisis_3 <- lista_out_3$tt_tibble %>% 
  mutate(tratamiento = case_when(
     is.na(fvc3)  ~  1,
    !is.na(fvc3)  ~  2
  )) %>% 
  mutate(tratamiento = as_factor(tratamiento))

analisis_3 %>% skimr::skim(fcv1_prev)

analisis_3 %>% count(tratamiento)

res_analisis_3 <- clogit(caso_control ~ tratamiento + strata(par), data = analisis_3)

saveRDS(res_analisis_3, file.path("Resultados", "res_analisis_3.RDS"))

```

```{r}
# resultados del análisis 3
summary(res_analisis_3)

```


# Análisis 4

***

Se estudia a los mayores de 40, con pauta previa y sin COVID previa. Por lo tanto, únicamente se tienen 2 categorías de tratamiento:

1) Sin booster.
2) Con booster.

```{r}
# se crean listas que dependen del proyecto------------------------------------
lista_exclusion_dep_analisis3 <- list(
  criterios = list(
  c1 = "Excluyendo: infección previa antes del tiempo_0...",
  c2 = "Excluyendo: edad < 40 años en tiempo_0...", 
  c3 = "Excluyendo: pauta incompleta 90 días antes del tiempo_0..."
  ),
  variables = list(
  v1 = "fcv1",
  v2 = "fnac_40",
  v3 = "fpauta_90"
  ),
  tipo_var = list(
    "exc",
    "inc",
    "inc"
    )
  )

```


```{r}
data_out_4 <- f_matching_tncc(
  data = pobana_2,
  tiempo_0 = "2022-01-03",
  caso_control_var = "tpos",
  caso_control_date = "ftes",
  base_test = test220613_cleaned,
  vector_match = variables_matching[c(1, 3:8)],
  crit_exc = lista_exclusion_dep_analisis3,
  birth_date = "fnac",
  age_first = 40,
  age_last = 90,
  age_by = 5)

data_out_4 %>% count(caso_control)

```

```{r}
lista_out_4 <- f_matching_tncc_mult(
  data = pobana_2,
  crit_exc = lista_exclusion_dep_analisis3,
  f_ini = "2022-01-10",
  f_fin = "2022-03-10",
  caso_control_var = "tpos",
  caso_control_date = "ftes",
  base_test = test220613_cleaned,
  vector_match = variables_matching[c(1, 3:8)],
  birth_date = "fnac",
  age_first = 40,
  age_last = 90,
  age_by = 5,
  verbose = FALSE)

lista_out_4$tt_tibble %>% count(caso_control)

```



```{r}
# logística condicional
analisis_4 <- lista_out_4$tt_tibble %>% 
  mutate(tratamiento = case_when(
     is.na(fvc3)  ~  1,
    !is.na(fvc3)  ~  2
  )) %>% 
  mutate(tratamiento = as_factor(tratamiento))

analisis_4 %>% skimr::skim(fcv1_prev)

analisis_4 %>% count(tratamiento)

res_analisis_4 <- clogit(caso_control ~ tratamiento + strata(par), data = analisis_4)

saveRDS(res_analisis_4, file.path("Resultados", "res_analisis_4.RDS"))

```

```{r}
# resultados del análisis 4
summary(res_analisis_4)

```
