==> 9_a_Demo.Rmd <==
---
title: "Emulación de target trial"
author: "Francisco Sánchez Sáez"
date: "`r format(Sys.Date())`"
output: 
  bookdown::html_document2:
    keep_md: false 
    theme: cerulean
    highlight: breezedark
    css: style.css
    anchor_sections: false
    toc: yes
    toc_float:
      collapsed: yes
      smooth_scroll: yes
      code_folding: show
    toc_depth: 3
bibliography: References.bib
csl: tl.csl
editor_options: 
  chunk_output_type: inline
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

```{r echo=FALSE, message=FALSE, include=FALSE, warning=FALSE}
# source("0_a_librerias.R")
# ya se cargan cuando se carguen las funciones

```

```{r echo=FALSE, message=FALSE, include=FALSE, warning=FALSE}
# función para cargar un .Rmd como si se hiciese source de un .r---------------
source_rmd = function(file, ...) {
  tmp_file = tempfile(fileext = ".R")
  on.exit(unlink(tmp_file), add = TRUE)
  knitr::purl(file, output = tmp_file)
  source(file = tmp_file, ...)
}

```

```{r echo=FALSE, message=FALSE, include=FALSE, warning=FALSE}
# cargar funciones de 0_funciones----------------------------------------------
if (!exists("se_ha_cargado_f")) {
suppressMessages(source_rmd("0_b_funciones.rmd"))
}

```

# Introducción

***

En este proyecto se va a realizar la metodología del *Target trial*. Dicha metodología consiste en realizar la emulación de un ensayo clínico a través de datos observacionales. En particular se aplica la metodología descrita en [@Danaei2013]. Empleando esta metodología, se van a emular múltiples ensayos con diferentes tiempos índice. Para cada tiempo índice se aplicaran los criterios de inclusión/exclusión del ensayo clínico a emular. Por lo tanto, vamos a tener individuos que pueden participar en más de un ensayo. En concreto, los no-iniciadores durante todo el periodo, y, los iniciadores, en el periodo previo a la iniciación, van a participar en más de una emulación del ensayo clínico siempre y cuando cumplan el resto de criterios de inclusión/exclusión.

El caso de estudio que se va a analizar es el de la efectivad vacunal de la tercera dosis (segunda para los pacientes de Janssen) en la población general de la comunidad foral de Navarra. En el conjunto de la población española, la efectividad estimada entre los días 7 hasta el 34 después de la tercera dosis, con fechas de vacunación comprendidas entre el 3 de enero y el 6 de febrero de 2022, fue del 51.3% (95% CI 50.2–52.4) [@Monge2022]. 

<br>

# Exclusiones estáticas

***

```{r cargar-datos, echo=FALSE, eval=TRUE, message=FALSE, warning=FALSE}
# cargar datos-----------------------------------------------------------------
pobana_0 <- readRDS(file = file.path("Datos", "Procesados", "pobana_0.RDS"))
test220613_cleaned <- readRDS(file = file.path("Datos", "Procesados", 
                                             "test220613_cleaned.RDS"))

```

En este .Rmd se realizan las **exclusiones que no dependen del tiempo 0** y por tanto se realizan antes de determinar la fecha (fechas) de inicio de las emulaciones de los ensayos clínicos. Se alimenta a la función con una lista que contiene el texto del criterio de exclusión, para que se muestre por pantalla, y de las variables que definen dicho criterio. De momento, el tipo de variables que definen las exclusiones son de dos tipos:

- Lógicas. Sí el valor es **TRUE**, es necesario excluir al individuo.

- No lógicas. Se excluyen las variables de las cuales no se tiene información (en este caso se utiliza para eliminar las observaciones con NA en la zona básica de salud).

Cuando se genere nueva casuística, se puede añadir entre el if() y el else() de la función *excluir()*.
      
```{r}
# se crean listas que dependen del proyecto------------------------------------
lista_exclusion <- list(
  criterios = list(
  c1 = "Excluyendo: socio y/o sanitarios...",
  c2 = "Excluyendo: trabj educacion...", # no tengo claro que se hayan de excluir
  c3 = "Excluyendo: viven en residencia...",
  c4 = "Excluyendo: dependientes...",
  c5 = "Excluyendo: zona basica desconocida..."
  # c6 = "Excluyendo: tuvo infección previa...", # esta es dependiente del tiempo
  # c7 = "Excluyendo: no finalizo vacunacion completa 90 dias antes..." # dependiente del tiempo
),
variables = list(
  v1 = "prfss",
  v2 = "educa",
  v3 = "resn",
  v4 = "depe",
  v5 = "czbs"
))

```

```{r}
# función para excluir siguiendo un criterio-----------------------------------
excluir <- function(pob, 
                    crit,
                    var)
{
  
  if (is.logical(pob %>% pull(var)) == TRUE) {
  print(glue("Variable lógica. {crit}"))
  pob <- pob %>% filter(is.na(get(var)) |
                     get(var) != TRUE)
    print(glue("n : {nrow(pob)}"))
  pob
  }
  else {
  print(glue("Variable no lógica. {crit}"))
  pob <- pob %>% filter(!is.na(get(var)))
    print(glue("n : {nrow(pob)}"))
  pob
  }
}

```

```{r}
# flujo------------------------------------------------------------------------
hacer_flujo <- function(poblacion,
                        exc = lista_exclusion
                   ) 
  {
  print(glue("n de partida: {nrow(poblacion)}"))

  for(i in seq_along(exc$criterios)) {
  poblacion <- excluir(pob = poblacion,
                       crit = exc$criterios[[i]],
                       var = exc$variables[[i]])
}
  poblacion
  }

```

```{r}
# aplicar exclusiones independientes del tiempo_0------------------------------
pobana_1 <- hacer_flujo(poblacion = pobana_0, 
            exc = lista_exclusion)

```

# Exclusiones dinámicas

***

En este .Rmd se realizan las **exclusiones que sí dependen del tiempo 0** y por tanto se realizan en conjunción con la determinación del tiempo 0. De momento, el tipo de variables que definen las exclusiones son de dos tipos:

- **Variables de exclusión (NA se mantiene)**. Sí la fecha de la variable es menor a la del tiempo 0 se excluye al individuo. Si no hay fecha, se mantiene al individuo (p.e. *fcv1* o *fdef*).

- **Variables inclusión (NA se excluye)**. Sí la fecha de la variable es mayor a la del tiempo 0 se excluye al individuo. Si no hay fecha, se excluye al individuo (p.e. *fnac_40* o *f_pauta_90*).

Cuando se genere nueva casuística, se puede añadir entre el if() y el else() de la función *excluir_dep()*.

```{r}
# arreglar variables exclusión/inclusión dinámicas-----------------------------
pobana_2 <- pobana_1 %>% 
  # crear fecha 40 cumpleaños
  mutate(fnac_40 = fnac %m+% years(40)) %>% 
  # crear fecha 90 días tras pauta completa
  mutate(fpauta_90 = case_when(
    !is.na(fvc2) ~ fvc2 %m+% days(90),
    tvc1 == "J&J / Janssen" ~ fvc1 %m+% days(90),
    T ~ NA_Date_
    )) %>%
  mutate(tvc1_label = tvc1) %>% 
  # pautas de Janssen o de la misma vacuna
  filter(tvc1 == "J&J / Janssen" | tvc1 == tvc2)

```
   
```{r}
# se crean listas que dependen del proyecto------------------------------------
lista_exclusion_dep <- list(
  criterios = list(
  c1 = "Excluyendo: infección previa antes del tiempo_0...",
  c2 = "Excluyendo: edad < 40 años en tiempo_0...", 
  c3 = "Excluyendo: pauta incompleta 90 días antes del tiempo_0...",
  c4 = "Excluyendo: fallecidos antes del tiempo_0...",
  c5 = "Excluyendo: boosting antes del tiempo_0..."),
  variables = list(
  v1 = "fcv1",
  v2 = "fnac_40",
  v3 = "fpauta_90",
  v4 = "fdef",
  v5 = "fvc3"),
  tipo_var = list(
    "exc",
    "inc",
    "inc",
    "exc",
    "exc")
  )

```



```{r}
# aplicar exclusiones dependientes del tiempo_0------------------------------
pobana_2_2022_01_03 <- hacer_flujo_dep(
                                        data = pobana_2, 
                                        exc = lista_exclusion_dep,
                                        tiempo_0 = "2022-01-03"
                                        )

```

# Target Trial (TT) emulation

***

```{r}
# vectores de matching---------------------------------------------------------
variables_matching = c(
  "gen",    # 1
  "gedad",  # 2
  "czbs",   # 3
  "tvc1",   # 4
  "migran", # 5
  "vgp20",  # 6
  "vgp21",  # 7
  "ucn8",   # 8
  "domi")   # 9

```

## TT con una fecha

***

```{r}
# matching f_pauta en días
prueba_matching_days <- f_matching_tt(
                data = pobana_2,
                tiempo_0 = "2022-01-03",
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fcv1", 
                vector_match = variables_matching[c(1, 3:8)],
                match_with_replacement = TRUE,
                dias_a_excluir_test_previos = NULL,
                base_test = test220613_cleaned,
                temporal_var = "fpauta_90",
                temporal_var_unit = "days",
                birth_date = "fnac",
                age_first = 40,
                age_last = 90,
                age_by = 5,
                censoring = "fdef",
                delayed_exposure_censoring = NULL,
                follow_up = 34,
                verbose = FALSE,
                seed = 3)

```


```{r include=FALSE}
# check prueba-----------------------------------------------------------------
prueba_matching_days %>% 
  tabyl(exposicion, caso) %>% 
  adorn_title()

```

## TT con más de una fecha

```{r}
# matching f_pauta en días (menos restrictivo)
prueba_matching_mult_days <- f_matching_tt_mult(
                data = pobana_2,
                f_ini = "2022-01-03",
                f_fin = "2022-01-05",
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fcv1", 
                vector_match = variables_matching[c(1, 3:8)], 
                base_test = test220613_cleaned,
                temporal_var = "fpauta_90",
                temporal_var_unit = "days",
                birth_date = "fnac",
                age_first = 40,
                age_last = 90,
                age_by = 5,
                censoring = "fdef",
                follow_up = 34,
                verbose = FALSE,
                seed = 3)

```

# Análisis

***

```{r}
# análisis con variables de matching como el artículo--------------------------
f_analisis(lista = prueba_matching_mult_days,
           tiempo_efecto = 7,
           lim_1 = 0.8,
           lim_zoom = 0.9)

```

# Hacer tabla 1

***

```{r}
# ejecutar la función para hacer la tabla--------------------------------------
tabla_1_demo <- hacer_tabla_1(tt_tibble = prueba_matching_mult_days$tt_tibble)

tabla_1_demo

```

# Hacer tabla 2

***

```{r include=FALSE, eval=TRUE}
# ejecutar la función para hacer la tabla--------------------------------------
tabla_2_demo <- hacer_tabla_2(tt_tibble = prueba_matching_mult_days$tt_tibble)

```

```{r}
tabla_2_demo

```

# Referencias

***
