==> 9_d_Paper_v1.Rmd <==
---
title: "Long-term vaccines effectiveness of mRNA boosters against COVID-19 infections, severe cases, and deaths on Omicron-dominant period in Navarre, a 640,000 inhabitants region of Spain"
date: "`r format(Sys.Date())`"
bookdown::html_document2:
    keep_md: false 
    theme: cerulean
    highlight: breezedark
    css: style.css
    anchor_sections: false
    toc: no
bibliography: References.bib
csl: tl.csl
editor_options: 
  chunk_output_type: inline
---

```{css, echo=FALSE}

/*-- Add logo (based on https://rstudio4edu.github.io/rstudio4edu-book/rmd-fancy.html): --*/
#TOC::before {content: ""; display: block; height: 60px; margin: 15px 10px 15px; background-image: url("logo.png"); background-size: contain; background-position: center center; background-repeat: no-repeat}
body {
  text-align: justify;
}

```

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)

```

```{r echo=FALSE, message=FALSE, include=FALSE, warning=FALSE}
# source("0_a_librerias.R")----------------------------------------------------
# ya se cargan cuando se carguen las funciones

```

```{r echo=FALSE, message=FALSE, include=FALSE, warning=FALSE}
# función para cargar un .Rmd como si se hiciese source de un .r---------------
source_rmd = function(file, ...) {
  tmp_file = tempfile(fileext = ".R")
  on.exit(unlink(tmp_file), add = TRUE)
  knitr::purl(file, output = tmp_file)
  source(file = tmp_file, ...)
}

```

```{r echo=FALSE, message=FALSE, include=FALSE, warning=FALSE}
# cargar funciones de 0_funciones----------------------------------------------
if (!exists("se_ha_cargado_f")) {
suppressMessages(source_rmd("0_b_funciones.rmd"))
}

```

```{r}
# cargar resultados------------------------------------------------------------
rr_m1    <- readRDS(file.path("Resultados", "rr_m1.RDS"))
rr_m1_st <- readRDS(file.path("Resultados", "rr_m1_st.RDS"))
rr_m2    <- readRDS(file.path("Resultados", "rr_m2.RDS"))
rr_m2_st <- readRDS(file.path("Resultados", "rr_m2_st.RDS"))
rr_m3    <- readRDS(file.path("Resultados", "rr_m3.RDS"))
rr_m3_st <- readRDS(file.path("Resultados", "rr_m3_st.RDS"))

```

***

***Francisco Sanchez-Saez <sup>1,2</sup>, Ibai Tamayo <sup>2,3</sup>, Ana Ariztegui-Echenique <sup>4</sup>, Javier Gorricho <sup>5</sup>, Julian Librero <sup>2,3</sup>.***

*<sup>1</sup> Health Services Research Unit, Foundation for the Promotion of Health and Biomedical Research of the Valencia Community (FISABIO), Valencia, Spain.*

*<sup>2</sup> Network for Research on Chronicity, Primary Care and Health Promotion (RICAPPS).*

*<sup>3</sup> Unidad de Metodología. Navarrabiomed, Hospital Universitario de Navarra (HUN), Universidad Pública de Navarra (UPNA), IdiSNA, Pamplona, Spain.*

*<sup>4</sup> Subdirección de Cuidados, Atención Domiciliaria, Sociosanitaria y Acciones Comunitarias, Servicio Navarro de Salud-Osasunbidea, Spain.*

*<sup>5</sup> Servicios de Evaluación y Difusión de Resultados, Servicio Navarro de Salud-Osasunbidea, Spain.*

***

# Abstract

***

In the estimation of the effectiveness of the booster mRNA vaccines against the omicron variant of SARS-CoV-2, as well as in any infectious diseases, one of the major limitations is the residual confounding among booster and no booster recipient groups in terms of infection exposure. This is expressed by the incidence differences in the exposed and unexposed groups even before the booster likely effect. In order to address this limitations we compared typical matching variables with matching by the home address. In addition, the evidence regarding the effectiveness in the long-term is scarce. 

In this work we followed the target trial emulation methodology. The effectiveness against infections from 2021-12-18 to 2022-03-27 are analysed. Then, effectiveness against severe cases was assessed from 2021-12-18 to 2022-06-13. The difference of selected periods of follow-up is due to the infections in the general population are not fully recorded from 2022-03-31 whereas the severe cases are. We selected individuals aged 18 years or older who had completed their primary vaccination at least 3 months before the start of follow-up, and had not tested positive for SARS-CoV-2 since the start of the pandemic, excluding  educational and health-care workers, and nursing home residents from Navarre (region of Spain of `r 640000 %>%
  prettyNum(big.mark = "&hairsp;")` inhabitants) throughout the BARDENA databases. On each day between 2021-12-17, and 2022-03-26, we matched individuals who had received a booster mRNA vaccine with controls by age (5 years groups), sex, number of previous negative test (0, 1, 2, 3+), type of primary vaccine (homologous or Janssen), time since complete vaccination, health basic zone, migrant status, flu vaccinated in 2020, and in 2021, and the number of people in the households (truncated at eight). We estimated risk of laboratory-confirmed SARS-CoV-2 infection using the Kaplan-Meier method, with 13 weeks of follow-up and we compared groups using risk ratios (RR). Vaccine effectiveness was calculated as one minus RR. In addition, we use the home address for matching in the same period, balancing the exposition in both groups. Finally, we assessed the long-term effect of the booster against severe cases on a period from 2021-12-18 to 2022-06-13 matching by age, sex and comorbidities.

The estimated effectiveness from day 7 to 35 after a mRNA booster was `r rr_m1_st$rr_eff` for the short-term (7-35 days) and `r rr_m1$rr_eff` for the long-term (7-91 days). In both settings we observed differences between the groups in the first 7 days. However, when the matching was based on sharing home, these differences decrease drastically. In this analysis, the short and mid term effectiveness are `r rr_m2_st$rr_eff` and `r rr_m2$rr_eff` respectively. Finally, the long-term effectiveness of booster in severe cases is high: `r rr_m3$rr_eff`, although there are a low number of cases. 

Summarising, i) it is interesting to have the pseudonymized home address of the participants in order to emulate the target trials and remove confounding, and ii) the booster still had moderate effectiveness against infection, even in the long-term time, and higher effectiveness against hospitalisations with less signs of waning effectiveness.

***

<br>

# Introduction

***

The COVID-19 infectious disease originated a pandemic since 2020, causing, at the end of November 2022, more than 638 millions of infections and 6.6 millions of deaths [@WHO2022b]. The challenge of the pandemic pushes the boundaries of science, medicine and research. The fast development of vaccines and the implementation of the vaccination programs against COVID-19 has supposed one of the most effective tools for dealing with the disease. In their development, the vaccines have been passed the different phases of the clinical trials in order to grantee their efficacy and safety. However, this clinical trials are conducted under determined conditions and in different time periods than their application. Thus, real world data (RWD) is needed in order to generate evidence about the vaccine effectiveness (VE).

Moreover, the mutation in the virus and the spread of new strains, such as Omicron, could change the assessed VE. Omicron has been the dominant strain of SARS-CoV-2 worldwide since endings of 2021. Particularly, in Navarre, Omicron was the dominant strain since epidemiological week 52 of 2021 (starting at 2021-12-26). Currently, the booster VE against the omicron variant of SARS-CoV-2 in the short term has been assessed in several works [@Monge2022, PONER MÁS]. Nevertheless, more evidence is needed in the assessment of the VE in the medium or long term. 

In this kind of design, which uses a observational study, some confounding could remain between booster and no booster recipient groups. This is expressed by the differences in the exposed and unexposed group before the booster likely effect (first seven days [@Walsh2020; @Lustig2021]). In order to address this problem we used matching variables getting the groups more comparable. We assessed the VE in a 5 and a 13-weeks periods of the mRNA booster (third dose) during the Omicron-dominant period in infections, in hospital admissions, and in deaths. To this end, we followed the target trial emulation methodology [@Danaei2013].

<br>

# Methods

***

This work uses the target trial framework. In this approach, a target randomized trial is described and emulated with observational data. In particular, we performed four different target trials to answer each of the three research questions. (i) infection as outcome matching with typical confounding variables, (ii) infection as outcome matching by home address, (iii) composite of COVID-19 related hospitalisation and deaths as outcome matching by age, sex and comorbidities.

## Outcomes

COVID-19 infections, severe cases, and deaths are considered. COVID-19 infections are determined by PCR or antigen tests. COVID-19 severe cases are defined as the WHO surveillance case definition for severe acute respiratory illness infection: A hospitalized person with acute respiratory infection, with a history of fever or measured fever of $\geq$ 38°C and cough with onset within the last 10 days [@WHO2022]. COVID-19 deaths are defined as the death which occur after 10 days of infection or in the hospital when the admission is due to the COVID-19.

## Setting

The source population was the whole population of Navarre, a region of Spain of about `r 640000 %>%
  prettyNum(big.mark = "&hairsp;")` inhabitants. We selected people who had programmed the booster vaccination in the omicron-dominant period. Therefore, we selected individuals aged 18 years or older who had completed their primary vaccination at least 3 months before the start of follow-up, and had not tested positive for SARS-CoV-2 since the start of the pandemic. In addition, educational and health-care workers, and nursing home residents are also excluded since their vaccination calendar and their exposure to infection is different to the general population. We obtain all the information throughout the BARDENA databases [@Bardena2022].
  
For assessing the booster VE, we select the Omicron dominant period plus a week before this period, when the likely effect of the vaccine is not produced yet. The end of the collected data was 2022-06-13 which determined the end of study period. However, when the assessed outcome is the COVID-19 infection (i, ii), the selected end of period was limited to 2022-03-27. From 2022-03-28 the Spanish government changed the surveillance methodology and not all the infections were registered. A microbiological diagnosis is only required for people older than 60 years, or hospitalised, or vulnerable of any age. On the other hand, the selected period for assessing the effectiveness against hospital admissions and deaths (iii) was 2021-12-26 to 2022-06-13 as the data about admissions and deaths is reliable even after 2022-03-27.

## Statistical analysis

From 2021-12-19 to 2022-03-19, individuals who had received a booster mRNA vaccine were matched with no booster recipients according to each of the four planned target trial. We compare booster and no booster recipients and we follow them until the outcome is produced, death or the end of the follow-up. In case of the non-treated component of the pair receives the booster, both components are censored. In the first target trial (i), the matching is done by age (5 years groups), sex, number of previous negative test (0, 1, 2, 3+), type of primary vaccine (two homologous doses of Pfizer, AZ, Moderna or one dose Janssen), time since complete vaccination, health basic zone, migrant status, flu vaccinated in 2020, and in 2021, and the number of people in the households (truncated at eight). The second target trial (ii) is identical to the first, but matching for home address. This, allows to reduce the residual confounding between treated and non-treated. However, the sample obtained will be diminished. Finally, in the third (iii) target trial, we compare individuals where the outcomes of interest were hospital admissions or deaths, matching by age, sex and comorbidities. The matching comorbidities are diabetes mellitus, coronary artery disease, chronic kidney failure, chronic obstructive pulmonary disease, hyperdyslipidemia, hypertension, stroke and obesity. For each target trial, we estimated risk of outcome using the Kaplan-Meier method, with 13 weeks of follow-up and we compared groups using risk ratios (RR). Finally, VE was calculated as one minus RR and expressed as percentage. 

This work has been developed following carefully the WHO guidance [references] and the STROBE statement [reference]. All the databases management and statistical analysis has been performed using R 4.2.1.

<br>

# Results

***
  
The source population is composed by `r n_flow$n0 %>%
  prettyNum(big.mark = "&hairsp;")` individuals, but after the exclusion criteria are applied, we have `r n_flow$ttA %>%
  prettyNum(big.mark = "&hairsp;")` individuals in the tt A), `r n_flow$ttA %>%
  prettyNum(big.mark = "&hairsp;")` individuals in the tt B), `r n_flow$ttC %>%
  prettyNum(big.mark = "&hairsp;")` individuals in the tt C), and `r n_flow$ttD %>%
  prettyNum(big.mark = "&hairsp;")` individuals in the tt D).

  
```{r}
# introducir figura del flow---------------------------------------------------

```
  
  
The estimated effectiveness from day 7 to 35 after a booster was 51·3% (95%CI 50·2, 52·4) in Spain [@Monge2022], meanwhile in Navarre region was `r rr_m1_st$rr_eff` for the 5-week period and `r rr_m1$rr_eff` for the 13-week period. In both settings we observed differences between the groups in the first 7 days (Fig 1A). However, when the matching was based on sharing home, these differences decrease drastically. In this analysis, the short and mid term effectiveness are `r rr_m2_st$rr_eff` and `r rr_m2$rr_eff` respectively (Fig 1B). Finally, the mid-term effectiveness of booster in severe cases is high: `r rr_m3$rr_eff`, although there are a low number of cases (Fig 1C). 

<br>

# Discussion

***

When assessing VE from real world data, one of the most important things is to assess which are the key confounder variables which is needed to taken into account in order to reducing the discrepances among treated and untreated groups. With the available observed variables used in the tt i) we can assimilate both groups pretty well although some confounding remains, as it is observed in the first seven days after the vaccine administrations. When we match by the pseudonymized home address the gap among groups is reduced further, obtaining a better balance. When the severe cases or deaths are assessed we can separate their probability as the probability of infection multiplied the probability of severe case or death once the infection is produced. Each of those terms have different factors. While for the infection, as we talked before, there are some variables which determine the exposure, for the severity or death there are another variables which are more important: besides the age and sex there are the comorbidities. In fact, the different comorbidities can influence in the personal decisions and therefore also define the infection exposure. The trade-off here consists of include the important variables which allow to compare groups properly, and at the same time avoid the overmatching.

The results point out that the vaccines are moderately effective in order to prevent the infection but very effective preventing severe cases and deaths.

This study has some limitations. The possible residual confounding between groups is a source of bias. Nevertheless, introducing the home address as the matching variable improves the similarity between groups in the first week.

Summarising, it is interesting to have the pseudonymized home address of the participants in order to emulate the target trials and remove confounding. Furthermore, the booster had moderate effectiveness against infection, even in the mid-term time, and higher effectiveness against hospitalisations. 

<br>

```{r figure, width = 13, height = 15, fig.cap = "Figure 1. Results of the target trial analysis"}
# poner la figura con el análisis principal, el del domicilio y el de los casos severos
knitr::include_graphics("Fig_paper.jpg")

```

<br>

# References

***

<div class="tocify-extend-page" data-unique="tocify-extend-page" style="height: 0;"></div>
