==> 9_e_Paper_Sup_Mat_v1.Rmd <==
---
title: "Supplementary Material. Correspondence: Booster Long-term effectiveness against infection and severe cases on Omicron-dominant period in Navarra"
date: "`r format(Sys.Date())`"
bookdown::html_document2:
    keep_md: false 
    theme: cerulean
    highlight: breezedark
    css: style.css
    anchor_sections: false
    toc: no
bibliography: References.bib
csl: tl.csl
editor_options: 
  chunk_output_type: inline
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)

```

```{r echo=FALSE, message=FALSE, include=FALSE, warning=FALSE}
# source("0_a_librerias.R")
# ya se cargan cuando se carguen las funciones

```

```{r echo=FALSE, message=FALSE, include=FALSE, warning=FALSE}
# función para cargar un .Rmd como si se hiciese source de un .r---------------
source_rmd = function(file, ...) {
  tmp_file = tempfile(fileext = ".R")
  on.exit(unlink(tmp_file), add = TRUE)
  knitr::purl(file, output = tmp_file)
  source(file = tmp_file, ...)
}

```

```{r echo=FALSE, message=FALSE, include=FALSE, warning=FALSE}
# cargar funciones de 0_funciones----------------------------------------------
if (!exists("se_ha_cargado_f")) {
suppressMessages(source_rmd("0_b_funciones.rmd"))
}

```

***Francisco Sanchez-Saez <sup>1,2</sup>, Ibai Tamayo <sup>2,3</sup>, Ana Ariztegui-Echenique <sup>4</sup>, Javier Gorricho <sup>5</sup>, Julian Librero <sup>2,3</sup>.***

*<sup>1</sup> Health Services Research Unit, Foundation for the Promotion of Health and Biomedical Research of the Valencia Community (FISABIO), Valencia, Spain.*

*<sup>2</sup> Network for Research on Chronicity, Primary Care and Health Promotion (RICAPPS).*

*<sup>3</sup> Unidad de Metodología. Navarrabiomed, Hospital Universitario de Navarra (HUN), Universidad Pública de Navarra (UPNA), IdiSNA, Pamplona, Spain.*

*<sup>4</sup> Subdirección de Cuidados, Atención Domiciliaria, Sociosanitaria y Acciones Comunitarias, Servicio Navarro de Salud-Osasunbidea, Spain.*

*<sup>5</sup> Servicios de Evaluación y Difusión de Resultados, Servicio Navarro de Salud-Osasunbidea, Spain.*

***

```{r}
# cargar datos-----------------------------------------------------------------



```

## fih1 con matching por edad, sexo y comorbilidades

```{r}
# fih1
paper_fih1 <- f_matching_tt_mult(
                data = pobana_2,
                f_ini = start_matching_date,
                f_fin = end_matching_severe_date,
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fih1", 
                vector_match = c("gen", "diab", "corp", "irc", "epoc", "hlip",
                                 "hta", "ictu", "obes"),
                # base_test = test220613_cleaned,
                # temporal_var = "fpauta_90",
                calcular_temp_var = "fpauta_90",
                temporal_var_unit = "weeks",
                calcular_test = test220613_cleaned,
                birth_date = "fnac",
                age_first = 20,
                age_last = 90,
                age_by = 5,
                censoring = "fdef",
                follow_up = 91,
                verbose = FALSE,
                seed = 21,
                fecha_fin_estudio = end_of_study_records)

# guardar resultados-----------------------------------------------------------
saveRDS(paper_fih1, file.path("Resultados",
        "paper_fih1.RDS"))

```

## fdef_covid con matching por edad, sexo y comorbilidades

```{r}
# fdef
paper_fdef <- f_matching_tt_mult(
                data = pobana_2,
                f_ini = start_matching_date,
                f_fin = end_matching_severe_date,
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fdef_covid", 
                vector_match = c("gen", "diab", "corp", "irc", "epoc", "hlip",
                                 "hta", "ictu", "obes"),
                # base_test = test220613_cleaned,
                # temporal_var = "fpauta_90",
                calcular_temp_var = "fpauta_90",
                temporal_var_unit = "weeks",
                calcular_test = test220613_cleaned,
                birth_date = "fnac",
                age_first = 20,
                age_last = 90,
                age_by = 5,
                censoring = "fdef",
                follow_up = 91,
                verbose = FALSE,
                seed = 21,
                fecha_fin_estudio = end_of_study_records)

# guardar resultados-----------------------------------------------------------
saveRDS(paper_fdef, file.path("Resultados",
        "paper_fdef.RDS"))

```

## fih1 con matching por edad, sexo y comorbilidades

```{r}
# fih1
paper_fih1_short_term <- f_matching_tt_mult(
                data = pobana_2,
                f_ini = start_matching_date,
                f_fin = end_matching_severe_date,
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fih1", 
                vector_match = matching_by,
                # base_test = test220613_cleaned,
                # temporal_var = "fpauta_90",
                calcular_temp_var = "fpauta_90",
                temporal_var_unit = "weeks",
                calcular_test = test220613_cleaned,
                birth_date = "fnac",
                age_first = 20,
                age_last = 90,
                age_by = 5,
                censoring = "fdef",
                follow_up = 35,
                verbose = FALSE,
                seed = 21,
                fecha_fin_estudio = end_of_study_records)

# guardar resultados-----------------------------------------------------------
saveRDS(paper_fih1_short_term, file.path("Resultados",
        "paper_fih1_short_term.RDS"))

```

## fdef_covid con matching por edad, sexo y comorbilidades

```{r}
# fih1
paper_fdef_short_term <- f_matching_tt_mult(
                data = pobana_2,
                f_ini = start_matching_date,
                f_fin = end_matching_severe_date,
                crit_exc = lista_exclusion_dep,
                exposure = "fvc3",
                out = "fdef_covid", 
                vector_match = c("gen", "diab", "corp", "irc", "epoc", "hlip",
                                 "hta", "ictu", "obes"),
                # base_test = test220613_cleaned,
                # temporal_var = "fpauta_90",
                calcular_temp_var = "fpauta_90",
                temporal_var_unit = "weeks",
                calcular_test = test220613_cleaned,
                birth_date = "fnac",
                age_first = 20,
                age_last = 90,
                age_by = 5,
                censoring = "fdef",
                follow_up = 35,
                verbose = FALSE,
                seed = 21,
                fecha_fin_estudio = end_of_study_records)

# guardar resultados-----------------------------------------------------------
saveRDS(paper_fdef_short_term, file.path("Resultados",
        "paper_fdef_short_term.RDS"))

```


# References

***

<div class="tocify-extend-page" data-unique="tocify-extend-page" style="height: 0;"></div>
