==> 9_9_NEJM_Examples.Rmd <==
---
title: "a"
author: "Francisco Sanchez Saez"
date: "2022-11-13"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Artículo 1

Evidence of the effectiveness of vaccines against severe acute respiratory syndrome coronavirus 2 (SARS-CoV-2) comes largely from studies of the original strain and of variants that emerged before the B.1.1.529 (omicron) variant.1-4 Previous infection confers protection against reinfection,1,5 but both infection-acquired and vaccine-acquired protection against infection wane over time.6-9

Recent studies indicate that vaccinations have continued to be effective against hospitalization and death10-13 but have had reduced effectiveness in protecting against confirmed infection11,14 and symptomatic illness10,15,16 with omicron. In some studies, estimates of hybrid immunity against infection with omicron have been reported.10,14 However, limited information is available with respect to the level of protection conferred by vaccine boosters and the timing of previous infection.

We analyzed data from the California Department of Corrections and Rehabilitation (CDCR), which operates the second largest state prison system in the United States. Prisons and jails are especially risky congregate settings for coronavirus disease 2019 (Covid-19), and many large outbreaks have occurred in these settings during the pandemic.1,17 The CDCR began offering a third (booster) mRNA vaccine dose to residents and staff in late August 2021; by the end of 2021, among persons in the eligible population, 77.9% of residents and 40.3% of staff had been vaccinated with a booster.18 The omicron variant was first identified within the CDCR system among assayed positive samples obtained from correctional staff on December 10, 2021. Substantial outbreaks occurred shortly thereafter among both residents and staff; these outbreaks were consistent with the timing of the worldwide wave of omicron infection.

We evaluated confirmed SARS-CoV-2 infections among nearly 60,000 incarcerated persons and 17,000 prison staff in California during the omicron outbreak period using data collected from December 24, 2021, through April 14, 2022. Our goal was to estimate the levels of protection conferred by mRNA vaccines against infection according to both the number of doses that persons had received and whether they had had a previous documented infection before the start of the observation period. We also estimated the effectiveness of a third mRNA vaccine dose among the persons who had been eligible to receive a booster.

Methods
STUDY DESIGN AND POPULATION
We used a retrospective cohort study design. The period of interest was restricted to outbreak phases that occurred at each of the 35 prisons in the CDCR system between December 24, 2021 (approximately 2 weeks after the omicron variant was initially identified within the prison system), and April 14, 2022. The CDCR considered a prison to be in an outbreak phase if three of more related cases among residents were detected within a 14-day period at the prison. During the outbreak phases, the prisons implemented modified operations, programs, and services for residents and mandatory testing for staff to minimize the risk of ongoing Covid-19 transmission.19,20 The SARS-CoV-2 infections that were detected in the prison system and in the general population in California during this 16-week period were dominated by the omicron variant.21 Our study analyzed two high-risk populations: residents and correctional staff at these prisons.

Residents were eligible for inclusion in the study cohort if they were incarcerated at the start of the outbreak phase at their prison. To focus on correctional workers who had the highest risk of workplace exposure, staff were eligible for inclusion if they worked in custody or health care positions (excluding contract employees), if they had regular direct contact with residents, if they had worked during the first week of the outbreak phase at their prison, and if they had worked at least half the number of days between the start of an outbreak and their data censoring date. In addition, to reduce the potential for misclassification of previous infection and vaccination status, we restricted our study cohort to residents who were incarcerated in a CDCR prison before January 1, 2021, and staff who were employed before January 1, 2021, and had worked consistently during 2021 (details are provided in the Supplementary Appendix, available with the full text of this article at NEJM.org).

Next, residents and staff who met the above criteria were excluded if they had received the adenovirus vector–based vaccine Ad26.COV2.S (Johnson & Johnson–Janssen), if they had received only one mRNA vaccine dose, if they had received an unknown vaccine or a vaccine that was not approved or authorized by the Food and Drug Administration, if they had inconsistencies regarding vaccination dates, if they had off-schedule vaccination,22 if they had incomplete demographic data, or if the prison where they were housed or where they worked was not known. To avoid possible misclassification of reinfections, we excluded persons who had had a recent SARS-CoV-2 infection (i.e., persons who had had an infection that was first detected ≤90 days before the start of the outbreak phase at their prison or persons who continued to test positive ≤30 days before the start of the outbreak phase). The observation time began on the start date of the outbreak and ended on the earliest of the following time points: the day that the sample from a positive test was obtained, the day before new vaccination, or the day before transfer (for residents) or a work-shift change (for staff) to a different prison.

DATA COLLECTION AND KEY MEASURES
CDCR collects and stores deidentified data on all residents and staff on a daily basis. For residents, detailed information from reverse-transcriptase–polymerase-chain-reaction (RT-PCR) and antigen SARS-CoV-2 testing came from a multilayered, voluntary testing program (in which 99.9% of the tests performed were RT-PCR assays) that included risk-based routine testing, surveillance testing, and testing in response to detected outbreaks (see Table S1 in the Supplementary Appendix). During outbreaks, all staff were tested at least once weekly through a mandatory RT-PCR testing program, and staff who worked in prison health care facilities were tested at least twice weekly.23 All staff could undergo testing voluntarily, and close contacts with active cases triggered a compulsory test.24 By January 1, 2022, more than half the prisons had entered an outbreak phase, with the phase of the last prison commencing on January 18, 2022.

All tests and vaccinations that were administered by or reported to the CDCR since March 2020 were used to derive each person’s previous infection status and vaccination status. We defined a previous SARS-CoV-2 infection by the presence of at least one positive test in the CDCR records before the start of the observation period. Previous infections were categorized according to whether they occurred before July 1, 2021, which reflected the period before the emergence of the B.1.617.2 (delta) variant, or whether they occurred on or after July 1, 2021, which corresponded to the period of high prevalence of the delta variant.1,21 Persons infected during both periods were assigned to the latter period.

In addition to detailed information on testing and vaccination (e.g., dates and vaccine brand), the study data included demographic characteristics of the cohort members (sex or gender identity, age, and race or ethnic group), carceral characteristics of the residents (prison, room type, and security level), and carceral characteristics of the staff (prison and type of position). For residents, a documented history of 17 possible coexisting conditions (e.g., hypertension, chronic kidney disease, and asthma) and a composite risk score designed by the CDCR to grade the risk of severe illness from SARS-CoV-2 infection (see Table S2) were provided. Risk scores were top-coded to a score of 2 in accordance with the CDCR definition of residents at moderate to high risk. Residents with scores of 2 or higher were younger than 65 years of age with coexisting conditions or were 65 years of age or older. This risk score was used to guide Covid-19 mitigation policies, including those involving vaccination, testing, and housing. Information on severe outcomes of infection (i.e., hospitalization and death) was available only for residents.

STUDY OVERSIGHT
The study was approved by the institutional review board at Stanford University. Results are reported in accordance with STROBE (Strengthening the Reporting of Observational Studies in Epidemiology) guidelines25; a checklist of STROBE recommendations is provided in the Supplementary Appendix. The authors vouch for the accuracy and completeness of the data.

STATISTICAL ANALYSIS
The goal of our analysis was to estimate the effectiveness of vaccination and previous infection against omicron infection. In our primary analysis, we used multivariate Cox models to examine the incidence of confirmed infections in groups defined according to combinations of vaccination and previous infection histories during the relevant omicron outbreak phase. The models allowed the baseline hazard to vary according to prison. We used inverse-probability weighting to reduce the effects of confounding due to differences in baseline characteristics among cohort members. Covariate-balancing propensity scores26 were calculated according to demographic, clinical, and carceral characteristics. Specifically, residents were weighted according to prison, Covid-19 risk score (0, 1, or ≥2), room type (cell or dormitory), sex (male or female), and age (as a continuous variable). Staff were weighted according to prison, position (custody or health care), age group (18 to 39, 40 to 54, or ≥55 years), and gender identity (man or woman); details are provided in Section 3.1 in the Supplementary Appendix. Stabilized weights were trimmed to reduce instability and bias from extreme scores.27

We estimated vaccine- and infection-acquired protection with a set of indicator variables that described combinations of vaccination and previous infection status. Effectiveness was expressed as a percent reduction in the hazard rate of omicron infection relative to the group that had not had previous documented infection or been vaccinated (measured as 1 minus the hazard ratio). We also repeated the analyses using a modified set of indicators to estimate the additional effectiveness of three doses of vaccine as compared with two doses. We calculated 95% confidence intervals using standard errors clustered according to prison. Analyses were not adjusted for multiple comparisons. We report the results for residents and staff separately.

The approach that was used in our primary analysis, described above, enabled the estimation and comparison of infection risks across various strata that were defined according to combinations of vaccination and previous infection histories. However, we hypothesized that an alternative study design — a rolling matched-cohort design28 — might be better suited to controlling for potential confounding when estimating vaccine effectiveness within each previous-infection stratum. Instead of weighting exposure groups on the basis of characteristics at the start of the omicron outbreak, as in our primary analysis, this alternative approach matched boosted persons with persons who had been eligible for a booster but had been vaccinated with only two doses. Persons were matched on the basis of previous infection status and other covariates relevant to vaccination according to the date that the boosted person was vaccinated with the third dose. Details are provided in Section 3.2 in the Supplementary Appendix. In this secondary analysis, Cox models were used to estimate the association between vaccination status (two doses or three doses) and omicron infection according to previous infection status. These models allowed the baseline hazard to vary according to prison, with adjustment for calendar week.

We performed five sets of sensitivity analyses. In the first set, we assessed the sensitivity of the results from our primary analysis to extreme weights by comparing estimates from a range of analyses that varied the extent to which large weights were trimmed downward. In the second set, we modified the primary analysis to adjust for the type of mRNA vaccine (BNT162b2 [Pfizer–BioNTech], mRNA-1273 [Moderna], or both). In the third set, we examined the sensitivity of estimates to our specification of reinfections by narrowing the exclusion criterion regarding recent infections, whereby the maximum number of days between the date that an infection was first detected and the start of the outbreak phase was shortened from 90 days to 30 days. In the fourth set, given that our primary analysis included both persons who had been eligible for boosters and those who had not been eligible, we excluded tests from persons who had been vaccinated with two doses but were not yet eligible for a third dose. In the fifth set of sensitivity analyses, to assess the sensitivity of our propensity-score model, we expanded the set of demographic and carceral characteristics that had previously been identified as being predictive of vaccine acceptance among incarcerated persons and staff.29,30

All the analyses were performed with the use of R software, version 3.5.2 (R Foundation for Statistical Computing). Additional details regarding model and variable specifications are provided in the Supplementary Appendix.

Results
STUDY POPULATION
Table 1.

Population Counts and Weighted Characteristics of Residents and Staff at Baseline According to Previous Infection and Vaccination Status.
Figure 1.

Study Population.
Figure 2.

Testing and Cases in the Study Cohort.
The weighted characteristics of the study population at baseline are shown in Table 1. Among 59,794 residents who met the inclusion criteria for the study cohort, 16.7% tested positive during the study period, and among 16,572 staff who met the inclusion criteria, 30.3% tested positive during the study period (Figure 1 and Table S4). The persons who were included in the study cohort were broadly representative of the overall population of residents who were incarcerated at any point during the study period and the overall population of staff who worked at any time during the study period (Table S3). The mean (±SD) number of times per week that persons were tested was 0.6±0.6 among residents and 1.5±1.0 among staff (Figure 2 and Fig. S2). A total of 96 hospitalizations and 1 death that were assessed as being related to SARS-CoV-2 infection were documented among residents. Data on hospitalizations and deaths were not available for staff.

Figure 3.

Vaccination and Previous Infection Status of the Study Cohort over Time.
Nearly half the residents and one third of the staff had previous confirmed infections, most of which had occurred before the delta variant became the dominant strain (Figure 3). Among residents, the median interval between the start of their last infection and the start of the outbreak phase in their prison was 393 days (interquartile range, 372 to 435); among staff, the median interval was 367 days (interquartile range, 163 to 400).

Among persons who had received two vaccine doses only, the median time since receipt of the second dose was 246 days (interquartile range, 127 to 294) for residents and 205 days (interquartile range, 90 to 324) for staff. Among those who had received three doses, the median time since receipt of the third dose was 35 days (interquartile range, 27 to 48) for residents and 37 days (interquartile range, 13 to 55) for staff.

EFFECTIVENESS OF VACCINATION AND PREVIOUS INFECTION AGAINST OMICRON INFECTION
Figure 4.

Adjusted Estimates of Effectiveness of Vaccination and Previous Infection against Omicron Infection among Residents and Staff in California State Prisons.
Within the resident population, the estimated effectiveness (measured as 1 minus the hazard ratio) of vaccination against confirmed SARS-CoV-2 infection during the period of omicron predominance among persons who had not had known previous infection was 18.6% (95% confidence interval [CI], 7.7 to 28.1) for those who had received only two vaccine doses and 40.9% (95% CI, 31.9 to 48.7) for those who had received three doses (Figure 4). Among the residents who had not been vaccinated, the estimated infection-conferred effectiveness against infection was 27.5% (95% CI, 14.8 to 38.4) among persons who had been infected before the period of delta predominance and 38.3% (95% CI, 6.5 to 59.3) among those who had been infected during the period of delta predominance. Among the residents who had been infected before the period of delta predominance, the estimated effectiveness of vaccination was 51.2% (95% CI, 41.5 to 59.2) with two vaccine doses and 57.7% (95% CI, 50.8 to 63.6) with three doses; among those infected during the period of delta predominance, the estimated effectiveness was 68.7% (95% CI, 38.5 to 84.1) and 84.6% (95% CI, 70.7 to 91.9), respectively.

Analyses of data from residents and staff produced estimates that indicated that the relative levels of effectiveness from vaccination and previous infection were generally consistent (Figure 4 and Table S5A). However, the estimated levels of effectiveness from vaccination among staff who had not had known previous infection were higher than those among residents without known previous infection, with estimates of 40.1% (95% CI, 34.0 to 45.6) among staff who had received two vaccine doses and 72.1% (95% CI, 67.2 to 76.2) among those who had received three doses.

Among the staff who had not been vaccinated, the estimated effectiveness of infection that had occurred before the period of delta predominance was 16.3% (95% CI, 8.1 to 23.7), and the estimated effectiveness of infection that had occurred during the period of delta predominance was 48.9% (95% CI, 41.6 to 55.3). Among the staff who had been infected before the period of delta predominance, the estimated effectiveness of vaccination was 55.8% (95% CI, 49.6 to 61.2) with two doses and 77.6% (95% CI, 71.7 to 82.2) with three doses; among those infected during the period of delta predominance, the estimated effectiveness was 83.2% (95% CI, 77.7 to 87.4) and 87.9% (95% CI, 76.0 to 93.9), respectively.

EFFECTIVENESS OF A THIRD DOSE AMONG BOOSTER-ELIGIBLE PERSONS
In our primary analysis, protection against infection was estimated relative to an individual person’s status at the start of the outbreak phase, and the analysis included both persons who had been eligible to receive a booster and those who had not been eligible to receive a booster. In this analysis, the estimated effectiveness of a third dose, as compared with two doses, among residents who were eligible to receive a booster was 27.4% (95% CI, 19.9 to 34.2) among those who had not had known previous infection, 13.3% (95% CI, 3.2 to 22.3) among those who had been infected before the period of delta predominance, and 50.7% (95% CI, −42.0 to 82.9) among those who had been infected during the period of delta predominance. Among staff, the estimated effectiveness of a third dose was 53.4% (95% CI, 46.7 to 59.3) among those who had not had known previous infection, 49.2% (95% CI, 35.0 to 60.3) among those who had been infected before the period of delta predominance, and 28.0% (95% CI, −33.0 to 61.0) among those who had been infected during the period of delta predominance.

In our secondary analysis, in which boosted persons were matched according to the date of vaccination with the third dose with controls who had been eligible for a third dose but had not received a third dose, the estimated incremental effectiveness of the third dose among residents was 25.0% (95% CI, 16.6 to 32.5) among those who had not had known previous infection and 46.4% (95% CI, 38.3 to 53.4) among those who had been infected before the period of delta predominance (Table S5B). Among staff, the estimated incremental effectiveness of the third dose was 57.9% (95% CI, 48.4 to 65.7) among those who had not had known previous infection and 57.7% (95% CI, 46.2 to 66.7) among those who had been infected before the period of delta predominance. Incremental effectiveness could not be estimated for persons infected during the period of delta predominance because no infections were documented among those who had received a third dose.

SENSITIVITY ANALYSES
Estimates that were derived with the use of untrimmed weights and weights with higher degrees of trimming were not appreciably different from those of the primary analysis, nor were estimates that were derived from alternative model specifications that were adjusted for the type of mRNA vaccine. Analyses of the expanded sample that included more recent previous infections produced slightly higher estimates of effectiveness than those of the primary analysis among persons whose previous infection occurred during the period of delta predominance. When persons who had not been eligible for a third dose were excluded, the estimates of effectiveness were slightly lower than those of the primary analysis among persons who had been vaccinated with two doses. Estimates that were derived with the use of the propensity-score model in which an expanded set of covariates was included produced estimates of effectiveness that were similar to those of the primary analysis. The results of the sensitivity analyses are provided in Tables S5A, S5C, S5D, and S5E.

Discussion
This study involving two high-risk populations evaluated the effectiveness of two or three doses of BNT162b2 and mRNA-1273 vaccines, and of previous infection before and during the period of delta predominance, in reducing the risk of infection with the SARS-CoV-2 omicron variant. Our findings add to the evidence base for vaccine effectiveness and immunity conferred by previous infection by providing estimates of effectiveness that were specific to different combinations of vaccination and previous infection histories. These estimates were based on two vulnerable populations in which reliable ascertainment of infection during the study period was likely, owing to high levels of testing.

Vaccination with three mRNA doses was associated with a reduced risk of infection with the omicron variant. Our estimates among persons who had not had known previous infection were similar to those of a convenience-sample study of the effectiveness of two or three mRNA-1273 doses involving data from a health care system that serves southern California.11 A cohort study that was conducted with the use of data from a health care system that serves Connecticut did not show significant additional protection from vaccination with a third dose among persons with previous infection.14 Our primary analysis included both persons who had been eligible to receive a third dose and those who had not been eligible to receive a third dose, an approach that potentially biased estimates of the effectiveness of a third dose downward. The higher risk of infection among persons who had received two or fewer vaccine doses could have induced selection bias if the most susceptible persons had been infected during the period before omicron predominance, which would also have biased estimates of effectiveness downward. As expected, the estimates of effectiveness were similar but slightly higher in the secondary analysis than in the primary analysis; the secondary analysis used an alternative design that probably corrected for these biases more effectively and provided estimates among persons who survived long enough to become eligible for a third dose. However, the conservative estimates from the primary analysis, as well as those from the secondary and sensitivity analyses, all indicate substantial, additional protection from a third mRNA dose against confirmed infection, irrespective of previous infection history. Our findings suggest that a third dose is beneficial, even in persons with previous confirmed infection.

The absolute levels of effectiveness against infection in our study are substantially lower than those estimated in earlier studies8,31-33 that were conducted before the omicron outbreaks occurred; furthermore, we observed that the estimated protection conferred by previous infection varied substantially with respect to timing, with considerably lower estimates among persons infected before the emergence of the delta variant. We estimated that vaccination augmented protection among persons who had had previous infection, a finding that is consistent with findings reported in other studies.11,14 Three doses of vaccine in combination with previous infection resulted in the highest estimates of effectiveness, and these levels approached those reported in studies involving vaccination with the primary series that were conducted before the occurrence of the omicron outbreaks.

For several combinations of exposure history, the resident population had lower estimates of effectiveness than the staff population. Although the CDCR has implemented an extensive testing program in both populations, testing among residents was neither routine nor compulsory. Undetected previous infections could have contributed to dilution of the estimates of effectiveness among residents, as could underascertainment of infections during the study period. Staff also have a greater ability to engage in protective behaviors, such as masking and social distancing, both at work and outside of work. Staff who remained uninfected before the emergence of the omicron variant may reflect a subpopulation that engages in a higher degree of protective behavior; this situation could also explain some of the differences between population estimates.

In this study, vaccine effectiveness was examined in the context of information about infection- and vaccine-acquired immunity against the omicron variant and information about the timing of previous infections. The study has notable strengths. First, we used detailed daily information regarding vaccination status and key Covid-19 outcomes for each member of the high-risk populations studied. These data allowed us to adjust for key potential confounders, as well as for demographic and exposure-related characteristics. Second, the large sample sizes of the two distinct populations enabled relatively high precision in our estimates, and the fact that estimates of the relative levels of effectiveness across the two populations were similar is notable despite the differences between the populations with respect to living situations, testing programs, and demographic characteristics.

Our study also has several limitations. Although a variety of covariates — including those related to vaccine acceptance and the risk of previous infection — were used to balance baseline characteristics, residual confounding could remain. Vaccine uptake and the occurrence of previous infections varied between residents and staff. Differences in infection risks between the two populations may in part reflect complex interactions of vaccine and previous infection levels and timing. Furthermore, the CDCR conducted limited viral sequencing and molecular testing, so we cannot disentangle the effects of variants from temporal waning, nor can we confirm that all cases that were observed during the study period were omicron infections.

Distinct testing programs and exposures for residents and staff during the study period may also have introduced confounding, although several results provide reassurances. First, although staff testing was inconsistent across vaccination and previous infection statuses, nearly all staff were tested at least once weekly (range of means, 0.9 to 2.0 tests per week), which provided relatively complete case detection. Second, although testing for residents was not routine, random, or compulsory, testing was relatively consistent across vaccination and previous infection statuses (range of means, 0.5 to 0.7 tests per week). Third, it is important to note that the consistency between the relative levels of effectiveness in the stratified analyses in these two populations provides some confidence that bias related to variation in testing practices may not be of major concern.

Several additional limitations warrant mention. Our estimates of effectiveness focused on confirmed infections and not on other important outcomes such as symptomatic infections or severe disease. The incidence of hospitalization and death in our sample was too low to support rigorous analysis of those outcomes according to the combinations of vaccine and previous infection histories analyzed, and symptom reporting was unreliable during the study period.34 In the secondary analysis, we were also unable to estimate the effectiveness of a third dose among persons who had had infection during the period of delta predominance. Finally, the generalizability of our results to jails, other prisons, other high-risk populations (e.g., residents of nursing homes and health care workers), and lower-risk populations is unknown.

The findings from this study suggest that, although mRNA vaccines and previous infection provide less protection against infection with the omicron variant than they did against earlier variants, boosters continue to provide substantial additional effectiveness, including among previously infected persons. Continued emphasis on vaccination and other ongoing mitigation practices is essential in preventing transmission, especially in highly vulnerable populations that have already borne a disproportionately large burden of disease and disruption during the Covid-19 pandemic.


# Articulo 2

The direct and indirect effects of the coronavirus disease 2019 (Covid-19) pandemic on children, their families, and surrounding communities have been substantial. By the end of February 2022, children and adolescents in the United States had a higher prevalence of infection with severe acute respiratory syndrome coronavirus 2 (SARS-CoV-2) than any other age group; children with Covid-19 are at risk for severe acute complications, death, and long-term sequelae (known as long Covid or post-Covid conditions).1-4 Furthermore, by the end of September 2022, more than 265,000 children and adolescents in the United States had had a parent or caregiver die from Covid-19,5,6 and the pandemic had caused substantial interruptions in school settings — including staffing shortages, closures, and missed school days — and had deepened educational inequities.7,8 These effects have been disproportionately borne by groups already made vulnerable by historical and contemporary systems of oppression, including structural racism and settler colonialism.9-11 Black, Latinx, and Indigenous children and adolescents are more likely to have had severe Covid-19, to have had a parent or caregiver die from Covid-19, and to be affected by worsening mental health and by educational disruptions than their White counterparts.6,8,12,13

During the Covid-19 pandemic, schools have become an important setting for implementing policies that minimize inequitable health, educational, social, and economic effects on children and their families. However, even before the pandemic, schools were not uniformly health-promoting environments. Chronic underinvestment in combination with structural racism codified in state-sanctioned historical and contemporary policies and practices (e.g., redlining, exclusionary zoning, disinvestment, and gentrification) eroded tax bases in some school districts and shaped the quality of public school infrastructure and associated environmental hazards.10,14-19 These processes left school districts differentially equipped to respond to the Covid-19 pandemic and concentrated high-risk conditions, such as crowded classrooms and poor indoor air quality due to outdated or absent ventilation or filtration systems, in low-income and Black, Latinx, and Indigenous communities.14,18,19

Alongside other measures, universal masking with high-quality masks or respirators has been an important piece of a layered risk-mitigation strategy to reduce the transmission of SARS-CoV-2 in community and school settings below levels that have been observed with individual (optional) masking.20-31 Massachusetts was among the 18 states plus Washington, DC, that had a universal masking policy in public schools during the 2021–2022 school year.32 The Massachusetts Department of Elementary and Secondary Education (DESE) rescinded the statewide masking policy on February 28, 2022, in accordance with updated guidance from the Centers for Disease Control and Prevention (CDC), and many Massachusetts school districts lifted masking requirements during the subsequent weeks. In the greater Boston area, only two school districts — the Boston and neighboring Chelsea districts — sustained masking requirements through June 2022.

The staggered lifting of masking requirements provided an opportunity to examine the potential effect of universal masking policies in schools. Specifically, this study aimed to assess trends in the observed weekly incidence of Covid-19 according to the length of time that school districts sustained masking requirements; to compare the incidence of Covid-19 among students and staff in districts that lifted masking requirements with the incidence in districts that sustained masking requirements in a given reporting week in order to estimate the effect of lifting masking requirements; and to compare school-district characteristics in districts that chose to sustain masking requirements longer with the characteristics in districts that chose to lift masking requirements earlier.

Methods
STUDY POPULATION
This study considered the 79 public, noncharter school districts in the greater Boston area, defined according to the U.S. Census Bureau as the New England city and town area of Boston–Cambridge–Newton (Fig. S1 in the Supplementary Appendix, available with the full text of this article at NEJM.org). Of these 79 school districts, 7 with unreliable or missing Covid-19 data were excluded (Table S1). The final sample included 72 school districts, which comprised 294,084 students and 46,530 staff during the study period. The study period was defined as the 40 calendar weeks of the 2021–2022 school year, which ended on June 15, 2022 (the end of the last full reporting week in all districts).

INTERVENTION AND PRIMARY OUTCOME
The primary exposure variable was whether a school district lifted or sustained its masking requirement in each reporting week. For all school districts, masking requirements were in place from the start of the study period through February 28, 2022, when the statewide masking policy was rescinded. A school district was considered to have lifted its masking requirement in a given reporting week if the requirement had been lifted before the first day of the reporting week (reporting weeks start on Thursday). The primary outcome was the incidence of Covid-19 among students and staff, considered together and separately.

DATA SOURCES
For each school district, data regarding weekly Covid-19 cases, student enrollment, and staffing during the 2021–2022 school year were publicly available from the Massachusetts DESE.33,34 Throughout the study period, DESE required standardized weekly reporting of all positive tests for Covid-19 among students and staff, regardless of symptoms, testing type or program (e.g., testing of symptomatic persons or pooled polymerase-chain-reaction testing), and testing location (community setting or school setting). Details regarding DESE reporting requirements are provided in the Supplementary Appendix. DESE strongly encouraged, and provided full funding for, school districts to opt in to standardized Covid-19 testing programs; 2311 Massachusetts schools (approximately 95%) participated in at least one such program. From 1 month before the statewide masking policy was rescinded through the end of the school year, statewide testing recommendations did not differ according to masking or vaccination status (Table S2).35

The dates during which masking requirements were in place for each school district were obtained from school-district websites or local news sources. For sensitivity analyses, data to be used for covariate adjustments, including data regarding Covid-19 indicators (i.e., measures of Covid-19 burden) according to city and town and Covid-19 vaccination coverage according to age, were publicly available from the Massachusetts Department of Public Health. For descriptive analyses, data regarding the distribution of students and staff according to sociodemographic characteristics and the distribution of students in populations selected and defined by DESE (low-income students, students with disabilities, and English-language learner [ELL] students) during the 2021–2022 school year were obtained from DESE.34 In addition, data regarding building conditions and learning environment were obtained from the Massachusetts School Building Authority 2016–2017 school survey (most recent data).36

CONTRIBUTIONS
The first, second, and last authors wrote the first draft of the manuscript and vouch for the accuracy and completeness of the data and the fidelity of the study to the protocol, available at NEJM.org. All the authors reviewed and edited the draft and made the decision to submit the manuscript for publication. No external funding was received for this study.

STATISTICAL ANALYSIS
Trends in the observed incidence of Covid-19 (weekly Covid-19 cases per 1000 population) before the statewide masking policy was rescinded were compared with trends after the policy was rescinded according to the length of time that school districts sustained masking requirements. A difference-in-differences analysis for staggered policy implementation was used to compare the incidence of Covid-19 in school districts that lifted masking requirements with the incidence in districts that sustained masking requirements in a given reporting week (i.e., school districts that had not yet lifted masking requirements) in order to estimate the effect of lifting masking requirements.37,38

Difference-in-differences methods allow for the estimation of causal effects of policy changes enacted at the group level by comparing the change in the outcome over time in the intervention group with the change in the control group, under an assumption of parallel trends (i.e., in the absence of the intervention, outcomes in the intervention group and the control group would have remained parallel over time).37,39 Unlike some observational methods, difference-in-differences methods are not biased by unmeasured time-invariant confounders or time-varying confounders with consistent trends across the intervention and control groups; the absence of such bias strengthens causal inferences. In this analysis, the weekly and cumulative effects of lifting masking requirements during the 15 weeks after the statewide masking policy was rescinded were estimated with respect to the incidence of Covid-19 cases in school districts that lifted masking requirements (i.e., the average [mean] treatment effect among the treated). Details regarding the difference-in-differences analysis are provided in the Supplementary Appendix.

Several sensitivity analyses were performed: a formal test for parallel trends before masking requirements were lifted; adjustment for time-varying covariates, including Covid-19 indicators at the community level, vaccination coverage, and previous incidences of infection among students and staff; and an assessment of the potential effects of differences in testing definitions or programs across districts (Tables S3 and S4). In the main analysis, data for the weeks in which school districts did not report Covid-19 cases were corrected (these weeks were originally recorded as having zero cases), all school districts in the greater Boston area were included as comparison districts, and weighting according to school population size was performed to capture the effect of masking policies at the population level.

Finally, to provide insight into Covid-19 policy decisions in schools and their potential to exacerbate or mitigate inequities in Covid-19 incidence and educational outcomes, descriptive analyses were performed. The decision to sustain or lift masking requirements was assessed according to various school-district characteristics, including sociodemographic characteristics of the students and staff and physical characteristics of the learning environment.

Results
PRIMARY ANALYSIS
Of the 72 school districts in the greater Boston area that were included in the study, only Boston Public Schools and Chelsea Public Schools sustained masking requirements throughout the study period (Fig. S2A). Of the remaining school districts, 46 districts (64%) lifted masking requirements in the first reporting week after the statewide masking policy was rescinded, 17 (24%) lifted masking in the second reporting week, and 7 (10%) lifted masking in the third reporting week (Fig. S2B). Cumulatively, 46 districts lifted masking requirements and 26 districts sustained masking requirements by the first reporting week after the policy was rescinded, 63 lifted and 9 sustained masking by the second reporting week, and 70 lifted and 2 sustained masking thereafter.

Figure 1.

Incidence of Covid-19 in School Districts in the Greater Boston Area before and after the Statewide Masking Policy Was Rescinded.
Before the statewide masking policy was rescinded, the trends in the incidence of Covid-19 observed in the Boston and Chelsea districts were similar to the trends in school districts that later lifted masking requirements. However, after the statewide masking policy was rescinded, the trends in the incidence of Covid-19 diverged, with a substantially higher incidence observed in school districts that lifted masking requirements than in school districts that sustained masking requirements. These trends were observed among students and staff overall (Figure 1A), as well as among students alone (Figure 1B) and among staff alone (Figure 1C).

Figure 2.

Difference-in-Differences Estimates of Additional Weekly and Cumulative Covid-19 Cases Associated with the Lifting of Masking Requirements.
Figure 2 shows difference-in-differences estimates of additional weekly and cumulative Covid-19 cases associated with the lifting of masking requirements. Before masking requirements were lifted, difference-in-differences estimates were essentially zero, a finding that supports the assumption of parallel trends. After masking requirements were lifted, the lifting of masking requirements was consistently associated with additional Covid-19 cases. The effect was significant during 12 of the 15 weeks after masking requirements were lifted. Weekly estimates ranged from 1.4 additional cases per 1000 students and staff (95% confidence interval [CI], 0.6 to 2.3) in the first reporting week after masking requirements were lifted to 9.7 additional cases per 1000 students and staff (95% CI, 7.1 to 12.3) in the ninth reporting week.

The strength of the association between school masking policies and the incidence of Covid-19 in school districts varied according to the incidence of Covid-19 in surrounding communities; the strongest associations were observed during the weeks when the incidences in surrounding communities were highest (Figs. S3 and S4). The weekly effects that were observed among students and staff overall were similar to those observed among students alone and those observed among staff alone, with slightly greater effects observed among staff than among students. In addition, the weekly effects were consistent with the cumulative effects.

Table 1.

Cumulative Incidence of Covid-19 and Estimated Effect of Lifting Masking Requirements during the 15 Weeks after the Statewide Masking Policy Was Rescinded.
Overall, the lifting of masking requirements was associated with an additional 44.9 Covid-19 cases per 1000 students and staff (95% CI, 32.6 to 57.1) during the 15 weeks after the statewide masking policy was rescinded (Table 1). This estimate corresponded to an additional 11,901 Covid-19 cases (95% CI, 8651 to 15,151), which accounted for 33.4% of the cases (95% CI, 24.3 to 42.5) in school districts that lifted masking requirements and for 29.4% of the cases (95% CI, 21.4 to 37.5) in all school districts during that period. The effect was more pronounced among staff. The lifting of masking requirements was associated with an additional 81.7 Covid-19 cases per 1000 staff (95% CI, 59.3 to 104.1) during the 15-week period, with these cases accounting for 40.4% of the cases (95% CI, 29.4 to 51.5) among staff in school districts that lifted masking requirements. Because persons who had a positive test for Covid-19 were instructed to isolate for at least 5 days, the additional cases translated to a minimum of approximately 17,500 missed school days for students and 6500 missed school days for staff during the 15-week period (Table S5).

SENSITIVITY ANALYSES
The results were shown to be robust in a range of sensitivity analyses, including analyses that assessed potential differences in testing programs and analyses that adjusted for Covid-19 indicators at the community level and vaccination coverage according to age (Fig. S5). Results of sensitivity analyses are provided in the Supplementary Appendix.

DESCRIPTIVE ANALYSES
Figure 3.

Characteristics of the School Districts.
School districts that chose to sustain masking requirements longer had higher percentages of low-income students, students with disabilities, and ELL students (Figure 3A), as well as higher percentages of Black and Latinx students (Figure 3B) and Black and Latinx staff (Figure 3C), than school districts that chose to lift masking requirements earlier. In addition, school districts that chose to sustain masking requirements longer tended to have school buildings that were older and in worse physical condition (e.g., with outdated or absent ventilation or filtration systems) and to have more students per classroom (Figure 3D).

Discussion
Schools are an important yet politically contested space in the Covid-19 response, which makes analyses such as this one particularly relevant to decision makers. We estimated that the lifting of masking requirements in school districts in the greater Boston area during March 2022 contributed an additional 45 Covid-19 cases per 1000 students and staff during the following 15-week period. Overall, this estimate corresponded to nearly 12,000 additional Covid-19 cases among students and staff, which accounted for one third of the cases in school districts that lifted masking requirements during that time and most likely translated to substantial loss of in-person school days.

We observed that the effect of school masking policies was greatest during the weeks when the background incidences of Covid-19 in surrounding cities and towns were highest, a finding that suggests that universal masking policies may be most effective when they are implemented before and throughout periods of high SARS-CoV-2 transmission. Under the CDC guidance at that time, as well as the updated guidance issued in August 2022, universal masking would not have been recommended until the incidences of Covid-19 in schools and surrounding communities were already nearing their peak (May 2022); by this time, a substantial proportion of the effects of masking polices that we observed had already accrued. As such, relying on lagging metrics such as CDC Covid-19 Community Levels and Covid-19 hospitalizations to inform school masking policies is most likely insufficient to prevent Covid-19 cases and loss of in-person school days, and policymakers might instead consider measures of community transmission (e.g., SARS-CoV-2 wastewater concentration or Covid-19 incidence) to inform such policies.

Understanding Covid-19 policy decisions requires attention to power and existing historical and sociopolitical contexts.10,40 Structural racism and racial capitalism operate through multiple pathways, including higher levels of household crowding and employment in essential industries and lower levels of access to testing, vaccines, and treatment; these structural forces differentially concentrate the risk of both SARS-CoV-2 exposure and severe Covid-19 in low-income and Black, Latinx, and Indigenous communities.9-11,18 In our study, school districts that chose to sustain masking requirements longer tended to have school buildings in worse physical condition and more students per classroom, and these districts had higher percentages of students and staff already made vulnerable by historical and contemporary systems of oppression (e.g., racism, capitalism, xenophobia, and ableism). In Boston and Chelsea, more than 80% of the students are Black, Latinx, or people of color, and these cities were among the Massachusetts cities and towns that were hit hardest by Covid-19. Students and families in these school districts have strongly advocated and organized for governmental action to increase Covid-19 protections in schools, emphasizing their role as essential workers, the risk to vulnerable family members, and the unequal consequences of missed work and school.41,42 The decision in some school districts to sustain school masking policies longer may therefore reflect an understanding among parents and elected officials that structural racism is embedded in public policies and that policy decisions have the potential to rectify or reproduce health inequities.10,14,16,40

A growing body of work suggests that knowledge of differential conditions and inequitable effects may decrease support for Covid-19 protections among systematically advantaged groups, whose relative position largely insulates them from Covid-19 harms, while simultaneously increasing support among groups that are directly affected by systems of oppression.43-45 For example, in a randomized trial in which White persons were assigned to receive information about structural causes of persistent Covid-19 inequities across racial or ethnic groups or to not receive such information, those who received the information were less likely to support Covid-19 prevention policies and were less likely to report individual concern about Covid-19 and empathy for the groups that were most affected.45 In several studies and polls, Black and Latinx parents were more likely than White parents to support school masking requirements and less likely to have confidence that schools could operate safely without additional protections.43,44,46 Failure to consider unequal baseline conditions and ongoing inequitable effects of Covid-19 policies risks further exacerbating inequities in Covid-19 incidence and educational outcomes.

Because universal masking policies in schools have been contentious, we anticipate several critiques. One such critique is that the benefits of universal masking in schools are outstripped by potential disruptions to teaching, learning, and social development. These effects warrant further rigorous evaluation; however, to date, there is no clear existing evidence that masking inhibits learning or harms development.47,48 In addition, such effects might be considered alongside the spectrum of benefits of universal masking, including fewer missed school days and staffing shortages, reduced risk of illness for students and their families, and reduced economic hardship for caregivers, who might miss work if their child is sick or if they become ill themselves. For example, in Lexington, MA, a comparison district approximately 10 miles from Boston, mean student and staff absences due to Covid-19 during weeks when masking was optional were 50% higher than absences during previous weeks, when masking was required (see the Supplementary Appendix).

In addition, severe Covid-19 and post-Covid conditions remain substantial risks among school-age children. Like much of the United States, the greater Boston area has low Covid-19 vaccination coverage among children (only 53% of children 5 to 11 years of age had been fully vaccinated in Boston and Chelsea through October 2022, as compared with 67% in comparison districts), with substantial inequities according to race or ethnic group and socioeconomic status. Furthermore, we observed greater benefits of sustaining masking among staff, a finding that emphasizes that universal masking is an important component of comprehensive workplace protections for staff, who may be at a higher risk for severe Covid-19 than students. In addition, staff absences may be especially consequential for students who need additional educational supports and services, including ELL students and students with disabilities.

A second common critique is that there are alternative approaches to reducing transmission and severe disease, such as improved ventilation and increased vaccination coverage. Our findings show that the better ventilation and higher vaccination coverage in school districts that lifted masking requirements than in districts that sustained masking requirements were insufficient to prevent all Covid-19 cases in these schools. Therefore, although we cannot weigh the full spectrum of individual and societal implications of masking policies, our study highlights the important role of interim universal school masking policies in mitigating the effects of Covid-19 while longer-term, more sustainable policies are developed to increase vaccination uptake and improve learning environments.

A key strength of this study is our use of difference-in-differences methods with staggered dates of the lifting of masking requirements. Although there are some factors related to SARS-CoV-2 exposure that differed across school districts, difference-in-differences methods yield robust analyses in the context of sources of confounding that do not change over time (e.g., sociodemographic characteristics or building conditions) or do not coincide with the policy change of interest. In sensitivity analyses, the benefits of masking requirements persisted after we controlled for Covid-19 indicators at the community level, vaccination coverage, and previous incidence of infection. Furthermore, we found that school districts that lifted masking requirements were districts that would have been expected to have lower incidences of Covid-19 (on average, they had buildings in better physical condition and had higher vaccination coverage), which suggests that any residual confounding by Covid-19 risk would have led to underestimation of the harms of lifting masking requirements overall.

A limitation of this study is that we did not have data regarding Covid-19 testing in individual school districts. However, DESE ended the practice of required testing of only unmasked close contacts in January 2022, and data from that “test-and-stay” program show that far too few schools continued with the program for it to explain our results. Under the most extreme assumptions, additional testing of unmasked close contacts could explain less than 7% of the estimated excess cases. Overall, our findings should be interpreted as the effect of universal masking policies and not as the effect of masking per se, since masks were still encouraged in most school settings. Despite this consideration, the effect of lifting masking requirements was substantial.

The winter wave of the B.1.1.529 (omicron) variant during the 2021–2022 school year will not be the final Covid-19 surge to affect students and staff, and ongoing efforts to address inequitable environmental risks and effects of Covid-19 in school settings are urgently needed. Our results support that universal masking with high-quality masks or respirators during periods of high community transmission is an important strategy for minimizing SARS-CoV-2 spread and loss of in-person school days. Our results also suggest that universal masking may be an important tool for mitigating the effects of structural racism in schools, including the differential risk of severe Covid-19, educational disruptions, and health and economic effects of secondary transmission to household members. School districts could use these findings to develop equitable mitigation plans in anticipation of a potential winter Covid-19 wave during the 2022–2023 school year, as well as clear decision thresholds for removing masks as the wave abates.

# Artículo 3

## Introduction

In both European and U.S. guidelines, prophylactic implantation of an implantable cardioverter–defibrillator (ICD) is a class 1 recommendation for patients with heart failure and reduced left ventricular systolic function.1,2 However, the evidence for a benefit is much stronger for patients with ischemic heart disease than it is for patients with heart failure from other causes. Over the past two decades, ICD implantation has been shown to be associated with substantial reductions in the rate of sudden cardiac death and total mortality in patients with ischemic heart disease.3-6 In the case of patients without ischemic heart disease, one trial showed an association between ICD implantation and a reduction in the rate of death attributed to arrhythmias, but no single trial has shown a convincing effect on total mortality.7-9 The Comparison of Medical Therapy, Pacing, and Defibrillation in Heart Failure (COMPANION) study, which randomly assigned 1520 patients with New York Heart Association (NYHA) class III or IV heart failure to receive medical therapy, a cardiac resynchronization therapy (CRT) pacemaker, or a CRT defibrillator, showed significantly lower all-cause mortality in association with a CRT defibrillator than with medical treatment alone, but a CRT defibrillator was not shown to be superior to a CRT pacemaker.10 The only randomized trial involving patients with nonischemic systolic heart failure in which a significant benefit with regard to all-cause mortality has been reported in association with the implantation of an ICD is the Sudden Cardiac Death in Heart Failure Trial (SCD-HeFT), which included 2521 patients, half of whom had nonischemic systolic heart failure.6 However, the positive effect of ICD treatment was confined to patients in NYHA class II, and no patients received concomitant CRT. In addition, medical treatment for heart failure has changed since SCD-HeFT was conducted.

Given the limited evidence of a benefit from the implantation of an ICD in patients with chronic nonischemic heart failure, we conducted a randomized trial in which patients whose condition was stable and who had chronic symptomatic heart failure, reduced ejection fraction, and increased levels of natriuretic peptide with or without a need for CRT were randomly assigned either to receive or not to receive an ICD.

## Methods

TRIAL OVERSIGHT
The Danish Study to Assess the Efficacy of ICDs in Patients with Non-ischemic Systolic Heart Failure on Mortality (DANISH) was an investigator-initiated multicenter, randomized, unblinded, controlled trial that was conducted at all centers in Denmark at which ICDs were implanted. The trial design and baseline characteristics of the patients have been published previously.11 The trial was designed and overseen by a steering committee (see the Supplementary Appendix, available with the full text of this article at NEJM.org) and was supported by unrestricted grants from Medtronic, St. Jude Medical, TrygFonden, and the Danish Heart Foundation; the funders had no influence on the design or conduct of the trial and were not involved in data collection or analysis, in the writing of the manuscript, or in the decision to submit it for publication. The trial protocol, available at NEJM.org, was approved by the regional scientific ethics committee for the Capital Region of Denmark and the Danish Data Protection Agency. The trial was performed in accordance with the principles of the Declaration of Helsinki. The authors assume responsibility for the accuracy and completeness of the data and analyses, as well as for the fidelity of the trial and this report to the protocol.

PATIENTS
Symptomatic patients (NYHA class II or III, or NYHA class IV if CRT was planned) with nonischemic systolic heart failure (left ventricular ejection fraction ≤35%) and an increased level (>200 pg per milliliter) of N-terminal pro–brain natriuretic peptide (NT-proBNP) were eligible for enrollment. The qualifying left ventricular ejection fraction and NT-proBNP level had to be measured after the doses of angiotensin-converting–enzyme (ACE) inhibitors or angiotensin-receptor blockers and beta-blockers had been increased to target levels, which were the guideline-specified levels whenever possible. A nonischemic cause of heart failure was usually determined by coronary angiography, although a normal computed tomographic (CT) angiogram or nuclear myocardial perfusion imaging study was acceptable. Patients could be included even if they had one or two coronary arteries with stenoses, if the extent of coronary artery disease was not considered to be sufficient to account for the reduced left ventricular systolic function. Patients with an existing conventional pacemaker or CRT pacemaker device could be included if the patients were willing to have the device changed or upgraded. Patients who had permanent atrial fibrillation with a resting heart rate higher than 100 beats per minute or renal failure that was being treated with dialysis were excluded. A complete list of the inclusion and exclusion criteria is provided in the Supplementary Appendix. All participants provided written informed consent.

TRIAL PROCEDURES
Participants were randomly assigned in a 1:1 ratio to either the ICD group or the control group (usual clinical care). Randomization was performed with the use of a Web-based system, in permuted blocks, with block sizes ranging from 2 to 6 patients, and was stratified according to center and according to whether patients were scheduled to receive CRT. The decision to implant a CRT device had to be made before randomization. Implantation of an ICD (or a CRT pacemaker or CRT defibrillator) was planned to be performed no later than 4 weeks after randomization; all patients were seen at follow-up visits after 2 months and every 6 months thereafter until the end of the trial. Device programming and monitoring of the ICD followed routine clinical practice (see the Supplementary Appendix).

OUTCOMES
The primary outcome was death from any cause. The secondary outcomes were sudden cardiac death, cardiovascular death, resuscitated cardiac arrest or sustained ventricular tachycardia, and change from baseline in quality of life (quality-of-life outcomes are not discussed in this report). The outcome designated as “sudden cardiac death” probably included some instances of sudden vascular death, because, as we realized during the adjudication of outcomes, a clear distinction often could not be made between sudden cardiac death and sudden vascular death. However, because the term “sudden cardiac death” was used in the protocol and event definitions, we have retained this term throughout. The safety outcome of device infection was prespecified; all other safety outcomes were based on adverse-event reporting. An end-point classification committee, the members of which were unaware of the treatment assignments, used prespecified criteria to adjudicate all prespecified clinical outcomes (see the Supplementary Appendix for criteria).

STATISTICAL ANALYSIS
Table 1.

Characteristics of the Patients at Baseline.
The study was designed to have 80% power to detect a 25% difference in total mortality between the treatment groups. At least 246 primary outcome events were required for the study to be conclusive, and we planned to include 1000 patients. Because the event rate and enrollment rate were lower than expected, the steering committee decided to prolong enrollment until June 30, 2014, or until 1200 patients were included (whichever came first) and to follow the last randomly assigned patient for at least 2 years. The baseline characteristics were compared between the treatment groups with the use of chi-square and Wilcoxon tests. Outcomes were analyzed with the use of time-to-event methods. Kaplan–Meier plots were calculated for total mortality, and cumulative incidence curves were calculated for events with competing risk (sudden cardiac death and cardiovascular death). The analysis of the primary outcome was performed with a log-rank model stratified according to center and status with respect to planned implantation of a CRT device. The proportional-hazard assumption was assessed with Schoenfeld residuals. Post hoc annual event rate ratios were derived from a Poisson regression. Prespecified subgroup analyses of the primary outcome were performed for the variables shown in Table 1. All analyses were performed in the intention-to-treat population. Two-sided P values of 0.05 or less were considered to indicate statistical significance. Analyses were performed with SAS software, version 9.4 (SAS Institute), and R software, version 3.3.1 (R Project for Statistical Computing).

## Results

CHARACTERISTICS OF THE PATIENTS
Figure 1.

Enrollment and Randomization of Patients.
From February 7, 2008, to June 30, 2014, a total of 1116 patients were enrolled at five centers; 556 patients were randomly assigned to the ICD group, and 560 patients were assigned to the control group (Figure 1). Overall, the two groups were balanced with respect to baseline characteristics (Table 1). In both groups, 58% of patients received CRT. The majority of patients received target doses of heart failure medication in accordance with the guidelines that were available at the time of the trial (Table S1 in the Supplementary Appendix), and among the patients with wide left bundle-branch block (QRS duration, ≥150 msec), 93% received CRT (see Table S2 in the Supplementary Appendix for baseline demographic characteristics according to CRT status and treatment group).

DEVICE IMPLANTATION
In the ICD group, 4 patients required two attempts at implantation, and 14 patients did not receive an ICD (1 patient died before implantation, implantation was unsuccessful in 2 patients, and 11 patients withdrew consent before implantation). The median time from randomization to implantation was 31 days (interquartile range, 19 to 43). A total of 27 patients in the control group received an ICD (24 because of an arrhythmic event and 3 at the request of their physician). During the trial, 85 patients (15.3%) in the ICD group had the battery in their ICD replaced, and 30 patients had the device removed or permanently deactivated because of an infection or by request (Figure 1).

FOLLOW-UP AND OUTCOMES
Follow-up data for all outcomes were available through June 30, 2016. The median follow-up period was 67.6 months (interquartile range, 49 to 85), and no patients were lost to follow-up for the primary outcome.

Figure 2.

Time-to-Event Curves for Death from Any Cause, Cardiovascular Death, and Sudden Cardiac Death.
Table 2.

Outcomes and Adverse Events.
The primary outcome, death from any cause, occurred in 120 patients (21.6%) in the ICD group (4.4 events per 100 person-years) and in 131 patients (23.4%) in the control group (5.0 events per 100 person-years) (Figure 2A and Table 2). The hazard ratio for death from any cause in the ICD group, as compared with the control group, was 0.87 (95% confidence interval [CI] 0.68 to 1.12; P=0.28). The test for the proportional-hazard assumption with Schoenfeld residuals gave a P value of 0.054. Annual event rate ratios from post hoc analyses are provided in Table S3 in the Supplementary Appendix. Cardiovascular death occurred in 77 patients (13.8%) in the ICD group and in 95 patients (17.0%) in the control group (hazard ratio, 0.77; 95% CI, 0.57 to 1.05; P=0.10) (Table 2 and Figure 2B). Sudden cardiac death occurred in 24 patients (4.3%) in the ICD group and in 46 patients (8.2%) in the control group (hazard ratio, 0.50; 95% CI, 0.31 to 0.82; P=0.005) (Table 2 and Figure 2C). The clinical outcome of resuscitated cardiac arrest or sustained ventricular tachycardia occurred with similar frequency in the two groups (Table 2); termination of ventricular tachycardia by antitachycardia pacing occurred in 97 patients (17.4%) in the ICD group, and appropriate shock for ventricular fibrillation or rapid ventricular tachycardia was given for 64 patients (11.5%) in the ICD group.

Figure 3.

Rate of Death from Any Cause (Primary Outcome) in Prespecified Subgroups.
The results of subgroup analyses are shown in Figure 3. The results were similar across all subgroups with the exception of age, for which there was a significant treatment-by-subgroup interaction (P=0.009 for the interaction) (Figure 3, and Fig. S1 in the Supplementary Appendix). In the subgroup of patients who were younger than 68 years of age, the rate of death from any cause was significantly lower in the ICD group than in the control group (hazard ratio, 0.64; 95% CI, 0.45 to 0.90; P=0.01). The effect of ICD implantation was independent of CRT status (P=0.73 for the interaction) (Fig. S2 in the Supplementary Appendix).

SAFETY
Device infections occurred in 27 patients (4.9%) in the ICD group and in 20 patients (3.6%) in the control group (P=0.29) (Table 2). Among patients who were not receiving CRT, the risk of device infection was higher in the ICD group than it was in the control group (12 of 234 patients [5.1%] vs. 2 of 237 patients [0.8%]; hazard ratio, 6.35; 95% CI, 1.38 to 58.87; P=0.006). Inappropriate shocks occurred in 33 patients (5.9%) in the ICD group (Table 2); 28 of the shocks were due to atrial fibrillation, 4 were due to oversensing, and 1 was due to supraventricular arrhythmia other than atrial fibrillaton. One patient had the ICD deactivated because of several shocks. Other complications during implantation were uncommon and occurred in both groups, because many of the control patients received a CRT device (Table 2).

## Discussion

We found that implantation of an ICD in patients who had heart failure that was not caused by ischemic heart disease did not provide an overall survival benefit, although the risk of sudden cardiac death was halved with an ICD. There was an important interaction with age that suggested that younger patients may have a survival benefit in association with ICD implantation. The results were independent of whether a patient received a CRT device.

All time-to-event curves appeared to diverge during the initial 5 years of the trial and then to converge (Figure 2). Post hoc analysis of annual event rate ratios (Supplementary Appendix) supports this interpretation. This observation lends support to the rationale for long-term studies, since the most common causes of death may change over time in patients with heart failure. In previous trials that showed a benefit from ICD implantation in patients with ischemic heart disease, the number needed to treat in order to save one life was as low as 6 over an 8-year period.3,12,13 There may be several explanations for the difference between these results and our findings. Patients with nonischemic heart failure may be less prone to death from arrhythmia than patients with ischemic heart failure, but better medical treatment and CRT also may have reduced the risk of death from arrhythmia for all patients with heart failure. Our event rate was lower than that observed in older studies, which reflects the fact that our study population consisted predominantly of outpatients who were in stable condition and who were treated medically in accordance with the guidelines, with almost every patient receiving beta-blockers and inhibitors of the renin–angiotensin system and 60% of the patients receiving mineralocorticoid-receptor antagonists.

We did not find a difference in the relative effect of an ICD between patients who received CRT and patients who did not. However, data on causes of death from observational studies involving patients receiving CRT suggest that an ICD may not be beneficial in the majority of these patients.14 In the COMPANION trial, the overall difference in mortality between the CRT defibrillator group and the CRT pacemaker group was not tested statistically.10 In later analyses of the data by independent authors, the crude P value based on overall mortality in the CRT pacemaker group versus the CRT defibrillator group was calculated as 0.12.15,16 At present, it is therefore unclear whether patients who are eligible for CRT should routinely receive an ICD, and a study in which patients are randomly assigned to receive a CRT pacemaker or a CRT defibrillator would be of potential clinical importance.16

In the current study, 31% of deaths were attributed to noncardiovascular causes. This is not surprising in an elderly population, but it highlights the importance of selecting patients for ICD implantation carefully. The rate of death among the patients we included in our trial was relatively low, and patients at higher risk may be more likely to benefit from ICD implantation. However, the subgroup data that suggest a lower likelihood of benefit in older patients might be used as an argument for not implanting ICDs in frail patients. Also, the benefit of ICD implantation with respect to sudden cardiac death that was seen in our trial is convincing, and patients who are not expected to die from other causes may be good candidates for ICD implantation. Long-term data from the Multicenter Automatic Defibrillator Implantation Trial (MADIT) II indicate that clinical risk scores may make it possible to identify the patients with ischemic heart disease who will benefit most from ICD implantation.12 Such a method for identifying patients with nonischemic heart failure who are at high risk for death from arrhythmia would be very useful.

The side effects associated with device implantation in our trial were not trivial. Device-related infections were not infrequent, but we did have a high proportion of patients in both groups who were receiving CRT (which requires implantation of an additional lead in the coronary sinus). In addition, 10% of our patients already had a pacemaker and were randomly assigned either to receive or not to receive an ICD upgrade; this is a patient group that has not been included in most of the previous trials. In a recent study, major device-related complications occurred within the first 6 months after implantation in approximately 6% of patients who received an ICD and 11% of patients who received a CRT defibrillator.17 Rates of inappropriate shocks or inappropriate antitachycardia pacing have decreased in recent years with the use of less aggressive ICD settings, but even with these modifications, 6% of patients have inappropriate shocks within 2.5 years after implantation of a dual-chamber ICD.13 These figures are similar to the 5.9% of patients who received inappropriate shocks in our trial.

The DANISH trial adds to the available body of data to consider in formulating the indications for ICD implantation in patients with nonischemic systolic heart failure. In the current American Heart Association guidelines, ICD implantation for primary prevention of sudden cardiac death in patients with symptomatic systolic heart failure is a class 1A recommendation, with no differentiation between patients with ischemic and nonischemic causes.1 In the European guidelines, ICD implantation is a class 1B recommendation for patients with nonischemic heart failure, as opposed to a class 1A recommendation for patients with ischemic heart failure.2 The American Heart Association guidelines refer solely to the SCD-HeFT trial for their recommendations regarding ICD implantation in patients with nonischemic causes of heart failure, whereas the European guidelines also refer to the Defibrillators in Non-Ischemic Cardiomyopathy Treatment Evaluation (DEFINITE) trial. However, SCD-HeFT enrolled patients between 1997 and 2001, and only 69% of enrolled patients received beta-blockers at baseline, 20% of patients received a mineralocorticoid-receptor antagonist, and no patients received concomitant CRT.6 In addition, the designation of ischemic or nonischemic causes in the SCD-HeFT trial was based primarily on patient history, with results from coronary angiography used to designate the cause only if such results were available. The DEFINITE trial, which enrolled patients between 1998 and 2002, included a larger proportion of patients who received beta-blockers (85%) but, again, no patients who received CRT.9

In conclusion, in our trial, prophylactic ICD implantation in patients with symptomatic systolic heart failure that was not caused by coronary artery disease was not found to reduce long-term mortality.
